Test.Cardano.Chain.Block.Model.Examples

Description

Module containing the imports needed to test that a given abstract trace passes the concrete validation. This is useful when debugging counterexamples.

Usage from the ghci repl:

import Hedgehog
Hedgehog.check $ property $ passConcreteValidation trace0

Replace trace0 by the trace under analysis.

Documentation

trace0 ∷ Trace CHAIN

A trace example. When debugging conformance tests you can add such a trace, and use it in the repl. NOTE: Do not commit such trace.