Test.Cardano.Chain.Config

Documentation

readMainetCfg ∷ MonadIO m ⇒ m Config

Read the test mainnet configuration file from the test directory.

An error is thrown if it is not possible to elaborate a genesis configuration from the genesis file.

We use RequiresNoMagic, as it indicates mainnet