Test.Cardano.Chain.UTxO.Model

Documentation

tests ∷ TSGroup

elaborateAndUpdate ∷ UTxOEnv → (UTxO, Map TxId TxId) → Tx → Either UTxOValidationError (UTxO, Map TxId TxId)

Elaborate a single transaction, apply it to the UTxO, and update the TxId map with the new concrete TxId

elaborateInitialUTxO ∷ UTxO → (UTxO, Map TxId TxId)

Create the initial concrete UTxO by elaborating the outputs and updating the map from abstract TxIds to concrete TxIds

elaborateTxWitnesses ∷ Map TxId TxId → [Tx] → ([ATxAux ByteString], Map TxId TxId)

elaborateTxWitsBSWithMap ∷ Map TxId TxId → Tx → (ATxAux ByteString, Map TxId TxId)