Test.Cardano.Ledger.Constrained.Conway.Cert

Description

Specs necessary to generate, environment, state, and signal for the CERT rule

Documentation

certEnvSpec ∷ ∀ era. EraSpecPParams era ⇒ WitUniv era → Specification (CertEnv era)

delegateeSpec ∷ Era era ⇒ WitUniv era → Specification (Set (Credential 'DRepRole))

shelleyCertStateSpec ∷ ∀ era. (EraSpecDeleg era, EraCertState era) ⇒ WitUniv era → Set (Credential 'DRepRole) → Map RewardAccount Coin → Specification (ShelleyCertState era)

conwayCertStateSpec ∷ ∀ era. (EraSpecDeleg era, EraCertState era, ConwayEraCertState era) ⇒ WitUniv era → Set (Credential 'DRepRole) → Map RewardAccount Coin → Specification (ConwayCertState era)

conwayTxCertSpec ∷ ∀ era. era ~ ConwayEra ⇒ WitUniv era → CertEnv era → CertState era → Specification (ConwayTxCert era)

genesisDelegCertSpec ∷ ∀ era. (AtMostEra "Babbage" era, Era era) ⇒ DState era → Specification GenesisDelegCert

Genesis delegations only work through the Babbage era. Hence the (AtMostEra Babbage era)

computeSets ∷ DState era → (KeyHash 'Genesis → Set (VRFVerKeyHash 'GenDelegVRF), KeyHash 'Genesis → Set (KeyHash 'GenesisDelegate))

Compute 2 functions from the DState. Each function, given a KeyHash, returns a Set of Hashes, we expect certain things to be in those sets. This mimics what happens in the Cardano.Ledger.Shelley.Rules.Deleg module

shelleyTxCertSpec ∷ ∀ era. (AtMostEra "Babbage" era, EraSpecPParams era, EraAccounts era) ⇒ WitUniv era → CertEnv era → ShelleyCertState era → Specification (ShelleyTxCert era)

class (HasSpec (TxCert era), EraCertState era) ⇒ EraSpecCert era where

Methods

txCertSpec ∷ WitUniv era → CertEnv era → CertState era → Specification (TxCert era)

txCertKey ∷ TxCert era → CertKey

certStateSpec ∷ WitUniv era → Set (Credential 'DRepRole) → Map RewardAccount Coin → Specification (CertState era)

data CertKey

Used to aggregate the key used in registering a Certificate. Different certificates use different kinds of Keys, that allows us to use one type to represent all kinds of keys (Similar to DepositPurpose)

Constructors

StakeKey !(Credential 'Staking)

PoolKey !(KeyHash 'StakePool)

DRepKey !(Credential 'DRepRole)

ColdKey !(Credential 'ColdCommitteeRole)

GenesisKey !(KeyHash 'Genesis)

MirKey !MIRPot

conwayTxCertKey ∷ ConwayTxCert era → CertKey

Compute the aggregate key type of a Certificater

shelleyTxCertKey ∷ ShelleyTxCert era → CertKey

testGenesisCert ∷ ∀ era. (AtMostEra "Babbage" era, EraSpecDeleg era, EraSpecPParams era, GenScript era) ⇒ Gen Property

testShelleyCert ∷ ∀ era. (Era era, AtMostEra "Babbage" era, EraSpecPParams era, EraSpecDeleg era, GenScript era, EraCertState era) ⇒ Gen Property

testConwayCert ∷ Gen Property