Test.Cardano.Ledger.Constrained.Conway.Certs

Description

Specs necessary to generate, environment, state, and signal for the CERTS rule

Documentation

bootstrapDStateSpec ∷ IsConwayUniv fn ⇒ CertsContext Conway → Specification fn (DState Conway)

coinToWord64 ∷ Coin → Word64

type CertsContext era = Map (RewardAccount (EraCrypto era)) Coin

txZero ∷ AlonzoTx Conway

certsEnvSpec ∷ IsConwayUniv fn ⇒ Specification fn (CertsEnv Conway)

projectEnv ∷ CertsEnv Conway → CertEnv Conway

Project a CertEnv out of a CertsEnv (i.e drop the Tx)

listSeqPairSpec ∷ IsConwayUniv fn ⇒ CertsEnv Conway → CertState Conway → Specification fn ([ConwayTxCert Conway], Seq (ConwayTxCert Conway))

Specify a pair of List and Seq, where they have essentially the same elements EXCEPT, the Seq has duplicate keys filtered out.

txCertsSpec ∷ IsConwayUniv fn ⇒ CertsEnv Conway → CertState Conway → Specification fn (Seq (ConwayTxCert Conway))

data CertKey c

Used to aggregate the key used in registering a Certificate. Different certificates use different kinds of Keys, that allows us to use one type to represent all kinds of keys (Similar to DepositPurpose)

Constructors

StakeKey !(Credential 'Staking c)

PoolKey !(KeyHash 'StakePool c)

DRepKey !(Credential 'DRepRole c)

ColdKey !(Credential 'ColdCommitteeRole c)

noSameKeys ∷ [ConwayTxCert era] → [ConwayTxCert era]