Test.Cardano.Ledger.Constrained.Conway.Deleg

Description

Specs necessary to generate, environment, state, and signal for the DELEG rule

Documentation

someZeros ∷ Specification RDPair

Specify that some of the rewards in the RDPair's are zero. without this in the DState, it is hard to generate the ConwayUnRegCert certificate, since it requires a rewards balance of 0. We also specify that both reward and deposit are greater than (Coin 0)

rewDepMapSpec ∷ Era era ⇒ WitUniv era → Map RewardAccount Coin → Specification (Map (Credential 'Staking) RDPair)

Specification for the RewardDepositMap of the Umap in the DState It must be witnessed, and conform to some properties relating to the withdrawals map, which is part of the context, so passed as an arg.

rewDepMapSpec2 ∷ ∀ era. Era era ⇒ WitUniv era → Map RewardAccount Coin → Specification (Map (Credential 'Staking) RDPair)

coinToWord64 ∷ Coin → Word64

wdrlCredentials ∷ Map RewardAccount Coin → Set (Credential 'Staking)

keyHashWdrl ∷ Map RewardAccount Coin → Set (Credential 'Staking)

dStateSpec ∷ ∀ era. EraSpecDeleg era ⇒ WitUniv era → Map RewardAccount Coin → Specification (DState era)

conwayDelegCertSpec ∷ ∀ era. (EraPParams era, ConwayEraCertState era) ⇒ ConwayDelegEnv era → CertState era → Specification ConwayDelegCert

delegEnvSpec ∷ EraSpecPParams era ⇒ Specification (ConwayDelegEnv era)

shelleyDelegCertSpec ∷ ∀ era. EraPParams era ⇒ WitUniv era → ConwayDelegEnv era → DState era → Specification ShelleyDelegCert

class (Era era, EraPParams era) ⇒ EraSpecDeleg era where

Methods

hasGenDelegs ∷ proxy era → Bool