Test.Cardano.Ledger.Constrained.Conway.GovCert

Description

Specs necessary to generate, environment, state, and signal for the GOVCERT rule

Documentation

vStateSpec ∷ Specification fn (VState (ConwayEra StandardCrypto))

govCertSpec ∷ IsConwayUniv fn ⇒ ConwayGovCertEnv (ConwayEra StandardCrypto) → VState (ConwayEra StandardCrypto) → Specification fn (ConwayGovCert StandardCrypto)

notYetResigned ∷ (HasSpec fn x, Ord x, IsConwayUniv fn) ⇒ Map x (CommitteeAuthorization StandardCrypto) → Term fn x → Pred fn

Operations for authenticating a HotKey, or resigning a ColdKey are illegal if that key has already resigned.

govCertEnvSpec ∷ IsConwayUniv fn ⇒ Specification fn (ConwayGovCertEnv (ConwayEra StandardCrypto))