Test.Cardano.Ledger.Constrained.Conway.Instances

Documentation

class (Reflect era, Eq (PParamsHKD Identity era), Show (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), EraPParams era) ⇒ EraSpecPParams era where

Methods

subsetToPP ∷ SimplePParams era → PParams era

ppToSubset ∷ PParams era → SimplePParams era

updateToPPU ∷ SimplePPUpdate → PParamsUpdate era

ppuToUpdate ∷ PParamsUpdate era → SimplePPUpdate

data SimplePPUpdate

Use this as the SimpleRep of (PParamsUpdate era)

Constructors

SimplePPUpdate

Fields

uminFeeA ∷ StrictMaybe Coin

uminFeeB ∷ StrictMaybe Coin

umaxBBSize ∷ StrictMaybe Word32

umaxTxSize ∷ StrictMaybe Word32

umaxBHSize ∷ StrictMaybe Word32

ukeyDeposit ∷ StrictMaybe Coin

upoolDeposit ∷ StrictMaybe Coin

ueMax ∷ StrictMaybe EpochInterval

unOpt ∷ StrictMaybe Word16

ua0 ∷ StrictMaybe NonNegativeInterval

urho ∷ StrictMaybe UnitInterval

utau ∷ StrictMaybe UnitInterval

udecentral ∷ StrictMaybe UnitInterval

uprotocolVersion ∷ StrictMaybe ProtVer

uminUTxOValue ∷ StrictMaybe Coin

uminPoolCost ∷ StrictMaybe Coin

ucoinsPerUTxOWord ∷ StrictMaybe Coin

ucostModels ∷ StrictMaybe CostModels

uprices ∷ StrictMaybe Prices

umaxTxExUnits ∷ StrictMaybe ExUnits

umaxBlockExUnits ∷ StrictMaybe ExUnits

umaxValSize ∷ StrictMaybe Natural

ucollateralPercentage ∷ StrictMaybe Natural

umaxCollateralInputs ∷ StrictMaybe Natural

ucoinsPerUTxOByte ∷ StrictMaybe Coin

upoolVotingThresholds ∷ StrictMaybe PoolVotingThresholds

udrepVotingThresholds ∷ StrictMaybe DRepVotingThresholds

ucommitteeMinSize ∷ StrictMaybe Natural

ucommitteeMaxTermLength ∷ StrictMaybe EpochInterval

ugovActionLifetime ∷ StrictMaybe EpochInterval

ugovActionDeposit ∷ StrictMaybe Coin

udRepDeposit ∷ StrictMaybe Coin

udRepActivity ∷ StrictMaybe EpochInterval

uminFeeRefScriptCostPerByte ∷ StrictMaybe NonNegativeInterval

data SimplePParams era

Use this as the SimpleRep of (PParams era). It is polymorphic enough to encode PParams in EVERY Era. The EraPParams instances remove the fields that do not appear in that Era.

Constructors

SimplePParams

Fields

minFeeA ∷ Coin

minFeeB ∷ Coin

maxBBSize ∷ Word32

maxTxSize ∷ Word32

maxBHSize ∷ Word32

keyDeposit ∷ Coin

poolDeposit ∷ Coin

eMax ∷ EpochInterval

nOpt ∷ Word16

a0 ∷ NonNegativeInterval

rho ∷ UnitInterval

tau ∷ UnitInterval

decentral ∷ UnitInterval

protocolVersion ∷ ProtVer

minUTxOValue ∷ Coin

minPoolCost ∷ Coin

Alonzo

coinsPerUTxOWord ∷ Coin

costModels ∷ CostModels

prices ∷ Prices

maxTxExUnits ∷ ExUnits

maxBlockExUnits ∷ ExUnits

maxValSize ∷ Natural

collateralPercentage ∷ Natural

maxCollateralInputs ∷ Natural

Babbage

coinsPerUTxOByte ∷ Coin

Conway

poolVotingThresholds ∷ PoolVotingThresholds

drepVotingThresholds ∷ DRepVotingThresholds

committeeMinSize ∷ Natural

committeeMaxTermLength ∷ EpochInterval

govActionLifetime ∷ EpochInterval

govActionDeposit ∷ Coin

dRepDeposit ∷ Coin

dRepActivity ∷ EpochInterval

minFeeRefScriptCostPerByte ∷ NonNegativeInterval

makePrices ∷ Integer → Integer → Prices

makeUnitInterval ∷ Integer → Integer → UnitInterval

makeNonNegativeInterval ∷ Integer → Integer → NonNegativeInterval

cSNothing_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn (StrictMaybe a)

cSJust_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn a → Term fn (StrictMaybe a)

succV_ ∷ BaseUniverse fn ⇒ Term fn Version → Term fn Version

minFeeA_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

minFeeB_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

maxBBSize_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Word32

maxTxSize_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Word32

maxBHSize_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Word32

keyDeposit_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

poolDeposit_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

eMax_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn EpochInterval

nOpt_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Word16

a0_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn NonNegativeInterval

rho_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn UnitInterval

tau_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn UnitInterval

decentral_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn UnitInterval

protocolVersion_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn ProtVer

minUTxOValue_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

minPoolCost_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

coinsPerUTxOWord_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

costModels_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn CostModels

prices_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Prices

maxTxExUnits_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn ExUnits

maxBlockExUnits_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn ExUnits

maxValSize_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Natural

collateralPercentage_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Natural

maxCollateralInputs_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Natural

coinsPerUTxOByte_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

poolVotingThresholds_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn PoolVotingThresholds

drepVotingThresholds_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn DRepVotingThresholds

committeeMinSize_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Natural

committeeMaxTermLength_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn EpochInterval

govActionLifetime_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn EpochInterval

govActionDeposit_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

dRepDeposit_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn Coin

dRepActivity_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn EpochInterval

minFeeRefScriptCostPerByte_ ∷ (EraSpecPParams era, BaseUniverse fn) ⇒ Term fn (SimplePParams era) → Term fn NonNegativeInterval

simplePParamsSpec ∷ ∀ fn era. (EraSpecPParams era, BaseUniverse fn) ⇒ Specification fn (SimplePParams era)

A sample of how to constrain (PParams era) for every Era, by writing a specification for SimplePParams. Constrained but not applicable fields (for that era) are elided in the result. Missing fields are left unconstrained and will appear as random values in the result. This can easily be lifted to PParams: see Test.Cardano.Ledger.Constrained.Conway.PParams(pparamsSpec)

data ProposalsSplit

Constructors

ProposalsSplit

Fields

psPPChange ∷ Integer

psHFInitiation ∷ Integer

psUpdateCommittee ∷ Integer

psNewConstitution ∷ Integer

psOthers ∷ Integer

type ProposalTree era = (StrictMaybe GovActionId, [Tree (GAS era)])

data StringFn (fn ∷ [Type] → Type → Type) as b

type IsConwayUniv fn = (BaseUniverse fn, Member (CoinFn fn) fn, Member (CoerceFn fn) fn, Member (StringFn fn) fn, Member (MapFn fn) fn, Member (FunFn fn) fn, Member (TreeFn fn) fn)

type ConwayFn = Fix (OneofL ConwayUnivFns)

sizedValue_ ∷ (HasSpec fn (Sized a), HasSpec fn a) ⇒ Term fn (Sized a) → Term fn a

sizedSize_ ∷ (HasSpec fn (Sized a), HasSpec fn a) ⇒ Term fn (Sized a) → Term fn Int64

txOutVal_ ∷ (HasSpec fn (Value era), Era era, HasSpec fn (Data era), Val (Value era), HasSpec fn (Script era), IsConwayUniv fn, HasSpec fn (BabbageTxOut era), IsNormalType (Script era)) ⇒ Term fn (BabbageTxOut era) → Term fn (Value era)

maryValueCoin_ ∷ IsConwayUniv fn ⇒ Term fn MaryValue → Term fn Coin

cKeyHashObj ∷ (IsConwayUniv fn, Typeable r) ⇒ Term fn (KeyHash r) → Term fn (Credential r)

cScriptHashObj ∷ (IsConwayUniv fn, Typeable r) ⇒ Term fn ScriptHash → Term fn (Credential r)

strLen_ ∷ ∀ fn s. (Member (StringFn fn) fn, StringLike s, HasSpec fn s) ⇒ Term fn s → Term fn Int

pProcDeposit_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (ProposalProcedure ConwayEra) → Term fn Coin

pProcGovAction_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (ProposalProcedure ConwayEra) → Term fn (GovAction ConwayEra)

gasId_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (GovActionState ConwayEra) → Term fn GovActionId

gasCommitteeVotes_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (GovActionState ConwayEra) → Term fn (Map (Credential 'HotCommitteeRole) Vote)

gasDRepVotes_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (GovActionState ConwayEra) → Term fn (Map (Credential 'DRepRole) Vote)

gasProposalProcedure_ ∷ (EraSpecPParams ConwayEra, IsConwayUniv fn) ⇒ Term fn (GovActionState ConwayEra) → Term fn (ProposalProcedure ConwayEra)

psPParamUpdate_ ∷ (EraSpecPParams era, Arbitrary (Proposals era), IsConwayUniv fn) ⇒ Term fn (Proposals era) → Term fn (ProposalTree era)

proposalSplitSum ∷ ProposalsSplit → Integer

genProposalsSplit ∷ Integer → Gen ProposalsSplit

onJust' ∷ (HasSpec fn a, IsNormalType a, IsPred p fn) ⇒ Term fn (StrictMaybe a) → (Term fn a → p) → Pred fn

onSized ∷ (IsConwayUniv fn, HasSpec fn a, IsPred p fn) ⇒ Term fn (Sized a) → (Term fn a → p) → Pred fn

coerce_ ∷ ∀ a b fn. (Member (CoerceFn fn) fn, HasSpec fn a, HasSpec fn b, CoercibleLike a b) ⇒ Term fn a → Term fn b

toDelta_ ∷ (HasSpec fn Coin, HasSpec fn DeltaCoin, Member (CoinFn fn) fn) ⇒ Term fn Coin → Term fn DeltaCoin

type BabbageTxBodyTypes era = '[Set TxIn, Set TxIn, Set TxIn, [Sized (TxOut era)], Maybe (Sized (TxOut era)), Maybe Coin, [TxCert era], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update era), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern BabbageTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AlonzoTxBodyTypes era = '[Set TxIn, Set TxIn, [TxOut era], [TxCert era], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update era), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern AlonzoTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type MaryTxBodyTypes era = '[Set TxIn, [TxOut era], [TxCert era], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update era), Maybe TxAuxDataHash, MultiAsset]

This is an abstraction of the Pattern MaryTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AllegraTxBodyTypes era = '[Set TxIn, [TxOut era], [TxCert era], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update era), Maybe TxAuxDataHash]

This is an abstraction of the Pattern AllegraTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type ShelleyTxBodyTypes era = '[Set TxIn, [TxOut era], [TxCert era], Map RewardAccount Coin, Coin, SlotNo, Maybe (Update era), Maybe TxAuxDataHash]

This is an abstraction of the Pattern ShelleyTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromShelleyBody ∷ ∀ era. EraTxBody era ⇒ ShelleyTxBody era → TxBody era

fromAllegraBody ∷ ∀ era. AllegraEraTxBody era ⇒ AllegraTxBody era → TxBody era

fromMaryBody ∷ ∀ era. MaryEraTxBody era ⇒ MaryTxBody era → TxBody era

fromAlonzoBody ∷ ∀ era. AlonzoEraTxBody era ⇒ AlonzoTxBody era → TxBody era

fromBabbageBody ∷ ∀ era. BabbageEraTxBody era ⇒ BabbageTxBody era → TxBody era

class (Era era, EraPParams era) ⇒ EraSpecDeleg era where

Methods

hasGenDelegs ∷ proxy era → Bool

data CertKey

Used to aggregate the key used in registering a Certificate. Different certificates use different kinds of Keys, that allows us to use one type to represent all kinds of keys (Similar to DepositPurpose)

Constructors

StakeKey !(Credential 'Staking)

PoolKey !(KeyHash 'StakePool)

DRepKey !(Credential 'DRepRole)

ColdKey !(Credential 'ColdCommitteeRole)

GenesisKey !(KeyHash 'Genesis)

MirKey !MIRPot

class (IsConwayUniv fn, HasSpec fn (TxCert era), Era era) ⇒ EraSpecCert era fn where

Methods

txCertSpec ∷ WitUniv era → CertEnv era → CertState era → Specification fn (TxCert era)

txCertKey ∷ TxCert era → CertKey

certStateSpec ∷ WitUniv era → Set (Credential 'DRepRole) → Map RewardAccount Coin → Specification fn (CertState era)

class (HasSpec fn (StashedAVVMAddresses era), EraSpecPParams era, EraSpecDeleg era, HasSpec fn (TxOut era), IsNormalType (TxOut era), EraTxOut era, GenScript era, IsConwayUniv fn) ⇒ EraSpecTxOut era fn where

The class EraSpecTxOut supports Era parametric Specifications that primarily navigate the differences in types parameterized type Family TxOut. Additional support for phased out Type Families like InstantaneousRewards, GenDelegs, FutureGenDelegs, StashedAVVMAddresses, and Ptrs, are also provided

Methods

irewardSpec ∷ WitUniv era → Term fn AccountState → Specification fn InstantaneousRewards

hasPtrs ∷ proxy era → Term fn Bool

txOutValue_ ∷ Term fn (TxOut era) → Term fn (Value era)

Extract a Value from a TxOut

txOutCoin_ ∷ Term fn (TxOut era) → Term fn Coin

Extract a Coin from a TxOut

txOutAddr_ ∷ Term fn (TxOut era) → Term fn Addr

Extract an Addr from a TxOut

txOutSpec ∷ ∀ fn era. EraSpecTxOut era fn ⇒ WitUniv era → Term fn (Map (Credential 'Staking) (KeyHash 'StakePool)) → Term fn (TxOut era) → Pred fn

An Era polymorhic Specification for type family TxOut

delegatedStakeReference ∷ IsConwayUniv fn ⇒ Term fn (Map (Credential 'Staking) (KeyHash 'StakePool)) → Specification fn StakeReference

Generate random Stake references that have a high probability of being delegated.