Test.Cardano.Ledger.Constrained.Conway.Instances.Basic

Description

This module provides HasSpec and HasSimpleRep instances for Basic types. A type is Basic if it is used to define PParams. See Test.Cardano.Ledger.Constrained.Conway.SimplePParams We divide these HasSpec and HasSimpleRep instances into two files because SimplePParams, needs these instances but not the 100's of other ones defined in Test.Cardano.Ledger.Constrained.Conway.Instances.Ledger And too many instances causes GHC 8.10.7 to blow up.

Documentation

cSNothing_ ∷ (HasSpec a, IsNormalType a) ⇒ Term (StrictMaybe a)

cSJust_ ∷ (HasSpec a, IsNormalType a) ⇒ Term a → Term (StrictMaybe a)

succV_ ∷ Term Version → Term Version

makePrices ∷ Integer → Integer → Prices

makeUnitInterval ∷ Integer → Integer → UnitInterval

makeNonNegativeInterval ∷ Integer → Integer → NonNegativeInterval

data SimplePParams era

Use this as the SimpleRep of (PParams era). It is polymorphic enough to encode PParams in EVERY Era. The EraPParams instances remove the fields that do not appear in that Era.

Constructors

SimplePParams

Fields

minFeeA ∷ Coin

minFeeB ∷ Coin

maxBBSize ∷ Word32

maxTxSize ∷ Word32

maxBHSize ∷ Word32

keyDeposit ∷ Coin

poolDeposit ∷ Coin

eMax ∷ EpochInterval

nOpt ∷ Word16

a0 ∷ NonNegativeInterval

rho ∷ UnitInterval

tau ∷ UnitInterval

decentral ∷ UnitInterval

protocolVersion ∷ ProtVer

minUTxOValue ∷ Coin

minPoolCost ∷ Coin

Alonzo

coinsPerUTxOWord ∷ Coin

costModels ∷ CostModels

prices ∷ Prices

maxTxExUnits ∷ ExUnits

maxBlockExUnits ∷ ExUnits

maxValSize ∷ Natural

collateralPercentage ∷ Natural

maxCollateralInputs ∷ Natural

Babbage

coinsPerUTxOByte ∷ Coin

Conway

poolVotingThresholds ∷ PoolVotingThresholds

drepVotingThresholds ∷ DRepVotingThresholds

committeeMinSize ∷ Natural

committeeMaxTermLength ∷ EpochInterval

govActionLifetime ∷ EpochInterval

govActionDeposit ∷ Coin

dRepDeposit ∷ Coin

dRepActivity ∷ EpochInterval

minFeeRefScriptCostPerByte ∷ NonNegativeInterval

data SimplePPUpdate

Use this as the SimpleRep of (PParamsUpdate era)

Constructors

SimplePPUpdate

Fields

uminFeeA ∷ StrictMaybe Coin

uminFeeB ∷ StrictMaybe Coin

umaxBBSize ∷ StrictMaybe Word32

umaxTxSize ∷ StrictMaybe Word32

umaxBHSize ∷ StrictMaybe Word32

ukeyDeposit ∷ StrictMaybe Coin

upoolDeposit ∷ StrictMaybe Coin

ueMax ∷ StrictMaybe EpochInterval

unOpt ∷ StrictMaybe Word16

ua0 ∷ StrictMaybe NonNegativeInterval

urho ∷ StrictMaybe UnitInterval

utau ∷ StrictMaybe UnitInterval

udecentral ∷ StrictMaybe UnitInterval

uprotocolVersion ∷ StrictMaybe ProtVer

uminUTxOValue ∷ StrictMaybe Coin

uminPoolCost ∷ StrictMaybe Coin

ucoinsPerUTxOWord ∷ StrictMaybe Coin

ucostModels ∷ StrictMaybe CostModels

uprices ∷ StrictMaybe Prices

umaxTxExUnits ∷ StrictMaybe ExUnits

umaxBlockExUnits ∷ StrictMaybe ExUnits

umaxValSize ∷ StrictMaybe Natural

ucollateralPercentage ∷ StrictMaybe Natural

umaxCollateralInputs ∷ StrictMaybe Natural

ucoinsPerUTxOByte ∷ StrictMaybe Coin

upoolVotingThresholds ∷ StrictMaybe PoolVotingThresholds

udrepVotingThresholds ∷ StrictMaybe DRepVotingThresholds

ucommitteeMinSize ∷ StrictMaybe Natural

ucommitteeMaxTermLength ∷ StrictMaybe EpochInterval

ugovActionLifetime ∷ StrictMaybe EpochInterval

ugovActionDeposit ∷ StrictMaybe Coin

udRepDeposit ∷ StrictMaybe Coin

udRepActivity ∷ StrictMaybe EpochInterval

uminFeeRefScriptCostPerByte ∷ StrictMaybe NonNegativeInterval

class (Eq (PParamsHKD Identity era), Show (PParamsHKD Identity era), ToExpr (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), EraPParams era, EraTxOut era, EraGov era, EraTx era) ⇒ EraSpecPParams era where

Methods

subsetToPP ∷ SimplePParams era → PParams era

ppToSubset ∷ PParams era → SimplePParams era

updateToPPU ∷ SimplePPUpdate → PParamsUpdate era

ppuToUpdate ∷ PParamsUpdate era → SimplePPUpdate

prettyE ∷ ToExpr x ⇒ x → Doc

Orphan instances

Num Coin

Instance details

Methods

(+) ∷ Coin → Coin → Coin

(-) ∷ Coin → Coin → Coin

(*) ∷ Coin → Coin → Coin

negate ∷ Coin → Coin

abs ∷ Coin → Coin

signum ∷ Coin → Coin

fromInteger ∷ Integer → Coin

HasSpec OrdExUnits

Instance details

Associated Types

type TypeSpec OrdExUnits

type Prerequisites OrdExUnits

Methods

emptySpec ∷ TypeSpec OrdExUnits

combineSpec ∷ TypeSpec OrdExUnits → TypeSpec OrdExUnits → Specification OrdExUnits

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec OrdExUnits → GenT m OrdExUnits

conformsTo ∷ OrdExUnits → TypeSpec OrdExUnits → Bool

shrinkWithTypeSpec ∷ TypeSpec OrdExUnits → OrdExUnits → [OrdExUnits]

toPreds ∷ Term OrdExUnits → TypeSpec OrdExUnits → Pred

cardinalTypeSpec ∷ TypeSpec OrdExUnits → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec OrdExUnits → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec OrdExUnits → BinaryShow

monadConformsTo ∷ OrdExUnits → TypeSpec OrdExUnits → Writer [String] Bool

typeSpecOpt ∷ TypeSpec OrdExUnits → [OrdExUnits] → Specification OrdExUnits

guardTypeSpec ∷ [String] → TypeSpec OrdExUnits → Specification OrdExUnits

prerequisites ∷ Evidence (Prerequisites OrdExUnits)

HasSpec Version

Instance details

Associated Types

type TypeSpec Version

type Prerequisites Version

Methods

emptySpec ∷ TypeSpec Version

combineSpec ∷ TypeSpec Version → TypeSpec Version → Specification Version

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Version → GenT m Version

conformsTo ∷ Version → TypeSpec Version → Bool

shrinkWithTypeSpec ∷ TypeSpec Version → Version → [Version]

toPreds ∷ Term Version → TypeSpec Version → Pred

cardinalTypeSpec ∷ TypeSpec Version → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec Version → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec Version → BinaryShow

monadConformsTo ∷ Version → TypeSpec Version → Writer [String] Bool

typeSpecOpt ∷ TypeSpec Version → [Version] → Specification Version

guardTypeSpec ∷ [String] → TypeSpec Version → Specification Version

prerequisites ∷ Evidence (Prerequisites Version)

HasSpec DRepVotingThresholds

Instance details

Associated Types

type TypeSpec DRepVotingThresholds

type Prerequisites DRepVotingThresholds

Methods

emptySpec ∷ TypeSpec DRepVotingThresholds

combineSpec ∷ TypeSpec DRepVotingThresholds → TypeSpec DRepVotingThresholds → Specification DRepVotingThresholds

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec DRepVotingThresholds → GenT m DRepVotingThresholds

conformsTo ∷ DRepVotingThresholds → TypeSpec DRepVotingThresholds → Bool

shrinkWithTypeSpec ∷ TypeSpec DRepVotingThresholds → DRepVotingThresholds → [DRepVotingThresholds]

toPreds ∷ Term DRepVotingThresholds → TypeSpec DRepVotingThresholds → Pred

cardinalTypeSpec ∷ TypeSpec DRepVotingThresholds → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec DRepVotingThresholds → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec DRepVotingThresholds → BinaryShow

monadConformsTo ∷ DRepVotingThresholds → TypeSpec DRepVotingThresholds → Writer [String] Bool

typeSpecOpt ∷ TypeSpec DRepVotingThresholds → [DRepVotingThresholds] → Specification DRepVotingThresholds

guardTypeSpec ∷ [String] → TypeSpec DRepVotingThresholds → Specification DRepVotingThresholds

prerequisites ∷ Evidence (Prerequisites DRepVotingThresholds)

HasSpec PoolVotingThresholds

Instance details

Associated Types

type TypeSpec PoolVotingThresholds

type Prerequisites PoolVotingThresholds

Methods

emptySpec ∷ TypeSpec PoolVotingThresholds

combineSpec ∷ TypeSpec PoolVotingThresholds → TypeSpec PoolVotingThresholds → Specification PoolVotingThresholds

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec PoolVotingThresholds → GenT m PoolVotingThresholds

conformsTo ∷ PoolVotingThresholds → TypeSpec PoolVotingThresholds → Bool

shrinkWithTypeSpec ∷ TypeSpec PoolVotingThresholds → PoolVotingThresholds → [PoolVotingThresholds]

toPreds ∷ Term PoolVotingThresholds → TypeSpec PoolVotingThresholds → Pred

cardinalTypeSpec ∷ TypeSpec PoolVotingThresholds → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec PoolVotingThresholds → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec PoolVotingThresholds → BinaryShow

monadConformsTo ∷ PoolVotingThresholds → TypeSpec PoolVotingThresholds → Writer [String] Bool

typeSpecOpt ∷ TypeSpec PoolVotingThresholds → [PoolVotingThresholds] → Specification PoolVotingThresholds

guardTypeSpec ∷ [String] → TypeSpec PoolVotingThresholds → Specification PoolVotingThresholds

prerequisites ∷ Evidence (Prerequisites PoolVotingThresholds)

HasSpec NonNegativeInterval

Instance details

Associated Types

type TypeSpec NonNegativeInterval

type Prerequisites NonNegativeInterval

Methods

emptySpec ∷ TypeSpec NonNegativeInterval

combineSpec ∷ TypeSpec NonNegativeInterval → TypeSpec NonNegativeInterval → Specification NonNegativeInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec NonNegativeInterval → GenT m NonNegativeInterval

conformsTo ∷ NonNegativeInterval → TypeSpec NonNegativeInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec NonNegativeInterval → NonNegativeInterval → [NonNegativeInterval]

toPreds ∷ Term NonNegativeInterval → TypeSpec NonNegativeInterval → Pred

cardinalTypeSpec ∷ TypeSpec NonNegativeInterval → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec NonNegativeInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec NonNegativeInterval → BinaryShow

monadConformsTo ∷ NonNegativeInterval → TypeSpec NonNegativeInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec NonNegativeInterval → [NonNegativeInterval] → Specification NonNegativeInterval

guardTypeSpec ∷ [String] → TypeSpec NonNegativeInterval → Specification NonNegativeInterval

prerequisites ∷ Evidence (Prerequisites NonNegativeInterval)

HasSpec ProtVer

Instance details

Associated Types

type TypeSpec ProtVer

type Prerequisites ProtVer

Methods

emptySpec ∷ TypeSpec ProtVer

combineSpec ∷ TypeSpec ProtVer → TypeSpec ProtVer → Specification ProtVer

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec ProtVer → GenT m ProtVer

conformsTo ∷ ProtVer → TypeSpec ProtVer → Bool

shrinkWithTypeSpec ∷ TypeSpec ProtVer → ProtVer → [ProtVer]

toPreds ∷ Term ProtVer → TypeSpec ProtVer → Pred

cardinalTypeSpec ∷ TypeSpec ProtVer → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec ProtVer → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec ProtVer → BinaryShow

monadConformsTo ∷ ProtVer → TypeSpec ProtVer → Writer [String] Bool

typeSpecOpt ∷ TypeSpec ProtVer → [ProtVer] → Specification ProtVer

guardTypeSpec ∷ [String] → TypeSpec ProtVer → Specification ProtVer

prerequisites ∷ Evidence (Prerequisites ProtVer)

HasSpec UnitInterval

Instance details

Associated Types

type TypeSpec UnitInterval

type Prerequisites UnitInterval

Methods

emptySpec ∷ TypeSpec UnitInterval

combineSpec ∷ TypeSpec UnitInterval → TypeSpec UnitInterval → Specification UnitInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec UnitInterval → GenT m UnitInterval

conformsTo ∷ UnitInterval → TypeSpec UnitInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec UnitInterval → UnitInterval → [UnitInterval]

toPreds ∷ Term UnitInterval → TypeSpec UnitInterval → Pred

cardinalTypeSpec ∷ TypeSpec UnitInterval → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec UnitInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec UnitInterval → BinaryShow

monadConformsTo ∷ UnitInterval → TypeSpec UnitInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec UnitInterval → [UnitInterval] → Specification UnitInterval

guardTypeSpec ∷ [String] → TypeSpec UnitInterval → Specification UnitInterval

prerequisites ∷ Evidence (Prerequisites UnitInterval)

HasSpec Coin

Instance details

Associated Types

type TypeSpec Coin

type Prerequisites Coin

Methods

emptySpec ∷ TypeSpec Coin

combineSpec ∷ TypeSpec Coin → TypeSpec Coin → Specification Coin

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Coin → GenT m Coin

conformsTo ∷ Coin → TypeSpec Coin → Bool

shrinkWithTypeSpec ∷ TypeSpec Coin → Coin → [Coin]

toPreds ∷ Term Coin → TypeSpec Coin → Pred

cardinalTypeSpec ∷ TypeSpec Coin → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec Coin → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec Coin → BinaryShow

monadConformsTo ∷ Coin → TypeSpec Coin → Writer [String] Bool

typeSpecOpt ∷ TypeSpec Coin → [Coin] → Specification Coin

guardTypeSpec ∷ [String] → TypeSpec Coin → Specification Coin

prerequisites ∷ Evidence (Prerequisites Coin)

HasSpec CostModels

Instance details

Associated Types

type TypeSpec CostModels

type Prerequisites CostModels

Methods

emptySpec ∷ TypeSpec CostModels

combineSpec ∷ TypeSpec CostModels → TypeSpec CostModels → Specification CostModels

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec CostModels → GenT m CostModels

conformsTo ∷ CostModels → TypeSpec CostModels → Bool

shrinkWithTypeSpec ∷ TypeSpec CostModels → CostModels → [CostModels]

toPreds ∷ Term CostModels → TypeSpec CostModels → Pred

cardinalTypeSpec ∷ TypeSpec CostModels → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec CostModels → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec CostModels → BinaryShow

monadConformsTo ∷ CostModels → TypeSpec CostModels → Writer [String] Bool

typeSpecOpt ∷ TypeSpec CostModels → [CostModels] → Specification CostModels

guardTypeSpec ∷ [String] → TypeSpec CostModels → Specification CostModels

prerequisites ∷ Evidence (Prerequisites CostModels)

HasSpec ExUnits

Instance details

Associated Types

type TypeSpec ExUnits

type Prerequisites ExUnits

Methods

emptySpec ∷ TypeSpec ExUnits

combineSpec ∷ TypeSpec ExUnits → TypeSpec ExUnits → Specification ExUnits

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec ExUnits → GenT m ExUnits

conformsTo ∷ ExUnits → TypeSpec ExUnits → Bool

shrinkWithTypeSpec ∷ TypeSpec ExUnits → ExUnits → [ExUnits]

toPreds ∷ Term ExUnits → TypeSpec ExUnits → Pred

cardinalTypeSpec ∷ TypeSpec ExUnits → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec ExUnits → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec ExUnits → BinaryShow

monadConformsTo ∷ ExUnits → TypeSpec ExUnits → Writer [String] Bool

typeSpecOpt ∷ TypeSpec ExUnits → [ExUnits] → Specification ExUnits

guardTypeSpec ∷ [String] → TypeSpec ExUnits → Specification ExUnits

prerequisites ∷ Evidence (Prerequisites ExUnits)

HasSpec Prices

Instance details

Associated Types

type TypeSpec Prices

type Prerequisites Prices

Methods

emptySpec ∷ TypeSpec Prices

combineSpec ∷ TypeSpec Prices → TypeSpec Prices → Specification Prices

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Prices → GenT m Prices

conformsTo ∷ Prices → TypeSpec Prices → Bool

shrinkWithTypeSpec ∷ TypeSpec Prices → Prices → [Prices]

toPreds ∷ Term Prices → TypeSpec Prices → Pred

cardinalTypeSpec ∷ TypeSpec Prices → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec Prices → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec Prices → BinaryShow

monadConformsTo ∷ Prices → TypeSpec Prices → Writer [String] Bool

typeSpecOpt ∷ TypeSpec Prices → [Prices] → Specification Prices

guardTypeSpec ∷ [String] → TypeSpec Prices → Specification Prices

prerequisites ∷ Evidence (Prerequisites Prices)

HasSpec EpochInterval

Instance details

Associated Types

type TypeSpec EpochInterval

type Prerequisites EpochInterval

Methods

emptySpec ∷ TypeSpec EpochInterval

combineSpec ∷ TypeSpec EpochInterval → TypeSpec EpochInterval → Specification EpochInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec EpochInterval → GenT m EpochInterval

conformsTo ∷ EpochInterval → TypeSpec EpochInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec EpochInterval → EpochInterval → [EpochInterval]

toPreds ∷ Term EpochInterval → TypeSpec EpochInterval → Pred

cardinalTypeSpec ∷ TypeSpec EpochInterval → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec EpochInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec EpochInterval → BinaryShow

monadConformsTo ∷ EpochInterval → TypeSpec EpochInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec EpochInterval → [EpochInterval] → Specification EpochInterval

guardTypeSpec ∷ [String] → TypeSpec EpochInterval → Specification EpochInterval

prerequisites ∷ Evidence (Prerequisites EpochInterval)

HasSimpleRep OrdExUnits

Instance details

Associated Types

type SimpleRep OrdExUnits

type TheSop OrdExUnits ∷ [Type]

Methods

toSimpleRep ∷ OrdExUnits → SimpleRep OrdExUnits

fromSimpleRep ∷ SimpleRep OrdExUnits → OrdExUnits

HasSimpleRep Version

Instance details

Associated Types

type SimpleRep Version

type TheSop Version ∷ [Type]

Methods

toSimpleRep ∷ Version → SimpleRep Version

fromSimpleRep ∷ SimpleRep Version → Version

HasSimpleRep DRepVotingThresholds

Instance details

Associated Types

type SimpleRep DRepVotingThresholds

type TheSop DRepVotingThresholds ∷ [Type]

Methods

toSimpleRep ∷ DRepVotingThresholds → SimpleRep DRepVotingThresholds

fromSimpleRep ∷ SimpleRep DRepVotingThresholds → DRepVotingThresholds

HasSimpleRep PoolVotingThresholds

Instance details

Associated Types

type SimpleRep PoolVotingThresholds

type TheSop PoolVotingThresholds ∷ [Type]

Methods

toSimpleRep ∷ PoolVotingThresholds → SimpleRep PoolVotingThresholds

fromSimpleRep ∷ SimpleRep PoolVotingThresholds → PoolVotingThresholds

HasSimpleRep ProtVer

Instance details

Associated Types

type SimpleRep ProtVer

type TheSop ProtVer ∷ [Type]

Methods

toSimpleRep ∷ ProtVer → SimpleRep ProtVer

fromSimpleRep ∷ SimpleRep ProtVer → ProtVer

HasSimpleRep Coin

Instance details

Associated Types

type SimpleRep Coin

type TheSop Coin ∷ [Type]

Methods

toSimpleRep ∷ Coin → SimpleRep Coin

fromSimpleRep ∷ SimpleRep Coin → Coin

HasSimpleRep CostModels

Instance details

Associated Types

type SimpleRep CostModels

type TheSop CostModels ∷ [Type]

Methods

toSimpleRep ∷ CostModels → SimpleRep CostModels

fromSimpleRep ∷ SimpleRep CostModels → CostModels

HasSimpleRep ExUnits

Instance details

Associated Types

type SimpleRep ExUnits

type TheSop ExUnits ∷ [Type]

Methods

toSimpleRep ∷ ExUnits → SimpleRep ExUnits

fromSimpleRep ∷ SimpleRep ExUnits → ExUnits

HasSimpleRep Prices

Instance details

Associated Types

type SimpleRep Prices

type TheSop Prices ∷ [Type]

Methods

toSimpleRep ∷ Prices → SimpleRep Prices

fromSimpleRep ∷ SimpleRep Prices → Prices

HasSimpleRep EpochInterval

Instance details

Associated Types

type SimpleRep EpochInterval

type TheSop EpochInterval ∷ [Type]

Methods

toSimpleRep ∷ EpochInterval → SimpleRep EpochInterval

fromSimpleRep ∷ SimpleRep EpochInterval → EpochInterval

HasDivision Coin

Instance details

Methods

doDivide ∷ Coin → Coin → Coin

divideSpec ∷ Coin → TypeSpec Coin → Specification Coin

MaybeBounded Coin

Instance details

Methods

lowerBound ∷ Maybe Coin

upperBound ∷ Maybe Coin

NumLike Coin

Instance details

Methods

subtractSpec ∷ Coin → TypeSpec Coin → Specification Coin

negateSpec ∷ TypeSpec Coin → Specification Coin

safeSubtract ∷ Coin → Coin → Maybe Coin

OrdLike Version

Instance details

Methods

leqSpec ∷ Version → Specification Version

ltSpec ∷ Version → Specification Version

geqSpec ∷ Version → Specification Version

gtSpec ∷ Version → Specification Version

OrdLike Coin

Instance details

Methods

leqSpec ∷ Coin → Specification Coin

ltSpec ∷ Coin → Specification Coin

geqSpec ∷ Coin → Specification Coin

gtSpec ∷ Coin → Specification Coin

OrdLike EpochInterval

Instance details

Methods

leqSpec ∷ EpochInterval → Specification EpochInterval

ltSpec ∷ EpochInterval → Specification EpochInterval

geqSpec ∷ EpochInterval → Specification EpochInterval

gtSpec ∷ EpochInterval → Specification EpochInterval

Foldy Coin

Instance details

Methods

genList ∷ ∀ (m ∷ Type → Type). MonadGenError m ⇒ Specification Coin → Specification Coin → GenT m [Coin]

theAddFn ∷ IntW '[Coin, Coin] Coin

theZero ∷ Coin

genSizedList ∷ ∀ (m ∷ Type → Type). MonadGenError m ⇒ Specification Integer → Specification Coin → Specification Coin → GenT m [Coin]

noNegativeValues ∷ Bool

ToExpr DRepVotingThresholds

Instance details

Methods

toExpr ∷ DRepVotingThresholds → Expr

listToExpr ∷ [DRepVotingThresholds] → Expr

ToExpr PoolVotingThresholds

Instance details

Methods

toExpr ∷ PoolVotingThresholds → Expr

listToExpr ∷ [PoolVotingThresholds] → Expr

HasSpec (NonZero Word64)

Instance details

Associated Types

type TypeSpec (NonZero Word64)

type Prerequisites (NonZero Word64)

Methods

emptySpec ∷ TypeSpec (NonZero Word64)

combineSpec ∷ TypeSpec (NonZero Word64) → TypeSpec (NonZero Word64) → Specification (NonZero Word64)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (NonZero Word64) → GenT m (NonZero Word64)

conformsTo ∷ NonZero Word64 → TypeSpec (NonZero Word64) → Bool

shrinkWithTypeSpec ∷ TypeSpec (NonZero Word64) → NonZero Word64 → [NonZero Word64]

toPreds ∷ Term (NonZero Word64) → TypeSpec (NonZero Word64) → Pred

cardinalTypeSpec ∷ TypeSpec (NonZero Word64) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (NonZero Word64) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (NonZero Word64) → BinaryShow

monadConformsTo ∷ NonZero Word64 → TypeSpec (NonZero Word64) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (NonZero Word64) → [NonZero Word64] → Specification (NonZero Word64)

guardTypeSpec ∷ [String] → TypeSpec (NonZero Word64) → Specification (NonZero Word64)

prerequisites ∷ Evidence (Prerequisites (NonZero Word64))

(EraSpecPParams era, EraTxOut era, EraGov era) ⇒ HasSpec (PParams era)
HasSpec instance for PParams

Instance details

Associated Types

type TypeSpec (PParams era)

type Prerequisites (PParams era)

Methods

emptySpec ∷ TypeSpec (PParams era)

combineSpec ∷ TypeSpec (PParams era) → TypeSpec (PParams era) → Specification (PParams era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (PParams era) → GenT m (PParams era)

conformsTo ∷ PParams era → TypeSpec (PParams era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (PParams era) → PParams era → [PParams era]

toPreds ∷ Term (PParams era) → TypeSpec (PParams era) → Pred

cardinalTypeSpec ∷ TypeSpec (PParams era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (PParams era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (PParams era) → BinaryShow

monadConformsTo ∷ PParams era → TypeSpec (PParams era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (PParams era) → [PParams era] → Specification (PParams era)

guardTypeSpec ∷ [String] → TypeSpec (PParams era) → Specification (PParams era)

prerequisites ∷ Evidence (Prerequisites (PParams era))

EraSpecPParams era ⇒ HasSpec (PParamsUpdate era)
HasSpec instance for PParams

Instance details

Associated Types

type TypeSpec (PParamsUpdate era)

type Prerequisites (PParamsUpdate era)

Methods

emptySpec ∷ TypeSpec (PParamsUpdate era)

combineSpec ∷ TypeSpec (PParamsUpdate era) → TypeSpec (PParamsUpdate era) → Specification (PParamsUpdate era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (PParamsUpdate era) → GenT m (PParamsUpdate era)

conformsTo ∷ PParamsUpdate era → TypeSpec (PParamsUpdate era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (PParamsUpdate era) → PParamsUpdate era → [PParamsUpdate era]

toPreds ∷ Term (PParamsUpdate era) → TypeSpec (PParamsUpdate era) → Pred

cardinalTypeSpec ∷ TypeSpec (PParamsUpdate era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (PParamsUpdate era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (PParamsUpdate era) → BinaryShow

monadConformsTo ∷ PParamsUpdate era → TypeSpec (PParamsUpdate era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (PParamsUpdate era) → [PParamsUpdate era] → Specification (PParamsUpdate era)

guardTypeSpec ∷ [String] → TypeSpec (PParamsUpdate era) → Specification (PParamsUpdate era)

prerequisites ∷ Evidence (Prerequisites (PParamsUpdate era))

Typeable r ⇒ HasSpec (KeyHash r)

Instance details

Associated Types

type TypeSpec (KeyHash r)

type Prerequisites (KeyHash r)

Methods

emptySpec ∷ TypeSpec (KeyHash r)

combineSpec ∷ TypeSpec (KeyHash r) → TypeSpec (KeyHash r) → Specification (KeyHash r)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (KeyHash r) → GenT m (KeyHash r)

conformsTo ∷ KeyHash r → TypeSpec (KeyHash r) → Bool

shrinkWithTypeSpec ∷ TypeSpec (KeyHash r) → KeyHash r → [KeyHash r]

toPreds ∷ Term (KeyHash r) → TypeSpec (KeyHash r) → Pred

cardinalTypeSpec ∷ TypeSpec (KeyHash r) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (KeyHash r) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (KeyHash r) → BinaryShow

monadConformsTo ∷ KeyHash r → TypeSpec (KeyHash r) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (KeyHash r) → [KeyHash r] → Specification (KeyHash r)

guardTypeSpec ∷ [String] → TypeSpec (KeyHash r) → Specification (KeyHash r)

prerequisites ∷ Evidence (Prerequisites (KeyHash r))

(EraGov era, EraTxOut era, EraSpecPParams era) ⇒ HasSpec (FuturePParams era)

Instance details

Associated Types

type TypeSpec (FuturePParams era)

type Prerequisites (FuturePParams era)

Methods

emptySpec ∷ TypeSpec (FuturePParams era)

combineSpec ∷ TypeSpec (FuturePParams era) → TypeSpec (FuturePParams era) → Specification (FuturePParams era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (FuturePParams era) → GenT m (FuturePParams era)

conformsTo ∷ FuturePParams era → TypeSpec (FuturePParams era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (FuturePParams era) → FuturePParams era → [FuturePParams era]

toPreds ∷ Term (FuturePParams era) → TypeSpec (FuturePParams era) → Pred

cardinalTypeSpec ∷ TypeSpec (FuturePParams era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (FuturePParams era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (FuturePParams era) → BinaryShow

monadConformsTo ∷ FuturePParams era → TypeSpec (FuturePParams era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (FuturePParams era) → [FuturePParams era] → Specification (FuturePParams era)

guardTypeSpec ∷ [String] → TypeSpec (FuturePParams era) → Specification (FuturePParams era)

prerequisites ∷ Evidence (Prerequisites (FuturePParams era))

EraSpecPParams era ⇒ HasSpec (ProposedPPUpdates era)

Instance details

Associated Types

type TypeSpec (ProposedPPUpdates era)

type Prerequisites (ProposedPPUpdates era)

Methods

emptySpec ∷ TypeSpec (ProposedPPUpdates era)

combineSpec ∷ TypeSpec (ProposedPPUpdates era) → TypeSpec (ProposedPPUpdates era) → Specification (ProposedPPUpdates era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (ProposedPPUpdates era) → GenT m (ProposedPPUpdates era)

conformsTo ∷ ProposedPPUpdates era → TypeSpec (ProposedPPUpdates era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (ProposedPPUpdates era) → ProposedPPUpdates era → [ProposedPPUpdates era]

toPreds ∷ Term (ProposedPPUpdates era) → TypeSpec (ProposedPPUpdates era) → Pred

cardinalTypeSpec ∷ TypeSpec (ProposedPPUpdates era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (ProposedPPUpdates era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (ProposedPPUpdates era) → BinaryShow

monadConformsTo ∷ ProposedPPUpdates era → TypeSpec (ProposedPPUpdates era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (ProposedPPUpdates era) → [ProposedPPUpdates era] → Specification (ProposedPPUpdates era)

guardTypeSpec ∷ [String] → TypeSpec (ProposedPPUpdates era) → Specification (ProposedPPUpdates era)

prerequisites ∷ Evidence (Prerequisites (ProposedPPUpdates era))

(HasSpec a, IsNormalType a) ⇒ HasSpec (StrictMaybe a)

Instance details

Associated Types

type TypeSpec (StrictMaybe a)

type Prerequisites (StrictMaybe a)

Methods

emptySpec ∷ TypeSpec (StrictMaybe a)

combineSpec ∷ TypeSpec (StrictMaybe a) → TypeSpec (StrictMaybe a) → Specification (StrictMaybe a)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (StrictMaybe a) → GenT m (StrictMaybe a)

conformsTo ∷ StrictMaybe a → TypeSpec (StrictMaybe a) → Bool

shrinkWithTypeSpec ∷ TypeSpec (StrictMaybe a) → StrictMaybe a → [StrictMaybe a]

toPreds ∷ Term (StrictMaybe a) → TypeSpec (StrictMaybe a) → Pred

cardinalTypeSpec ∷ TypeSpec (StrictMaybe a) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (StrictMaybe a) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (StrictMaybe a) → BinaryShow

monadConformsTo ∷ StrictMaybe a → TypeSpec (StrictMaybe a) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (StrictMaybe a) → [StrictMaybe a] → Specification (StrictMaybe a)

guardTypeSpec ∷ [String] → TypeSpec (StrictMaybe a) → Specification (StrictMaybe a)

prerequisites ∷ Evidence (Prerequisites (StrictMaybe a))

HasSimpleRep (NonZero Word64)

Instance details

Associated Types

type SimpleRep (NonZero Word64)

type TheSop (NonZero Word64) ∷ [Type]

Methods

toSimpleRep ∷ NonZero Word64 → SimpleRep (NonZero Word64)

fromSimpleRep ∷ SimpleRep (NonZero Word64) → NonZero Word64

EraSpecPParams era ⇒ HasSimpleRep (PParams era)
SimpleRep instance for PParams

Instance details

Associated Types

type SimpleRep (PParams era)

type TheSop (PParams era) ∷ [Type]

Methods

toSimpleRep ∷ PParams era → SimpleRep (PParams era)

fromSimpleRep ∷ SimpleRep (PParams era) → PParams era

EraSpecPParams era ⇒ HasSimpleRep (PParamsUpdate era)
SimpleRep instance for PParamsUpdate

Instance details

Associated Types

type SimpleRep (PParamsUpdate era)

type TheSop (PParamsUpdate era) ∷ [Type]

Methods

toSimpleRep ∷ PParamsUpdate era → SimpleRep (PParamsUpdate era)

fromSimpleRep ∷ SimpleRep (PParamsUpdate era) → PParamsUpdate era

EraSpecPParams era ⇒ HasSimpleRep (FuturePParams era)

Instance details

Associated Types

type SimpleRep (FuturePParams era)

type TheSop (FuturePParams era) ∷ [Type]

Methods

toSimpleRep ∷ FuturePParams era → SimpleRep (FuturePParams era)

fromSimpleRep ∷ SimpleRep (FuturePParams era) → FuturePParams era

EraSpecPParams era ⇒ HasSimpleRep (ProposedPPUpdates era)

Instance details

Associated Types

type SimpleRep (ProposedPPUpdates era)

type TheSop (ProposedPPUpdates era) ∷ [Type]

Methods

toSimpleRep ∷ ProposedPPUpdates era → SimpleRep (ProposedPPUpdates era)

fromSimpleRep ∷ SimpleRep (ProposedPPUpdates era) → ProposedPPUpdates era

Typeable a ⇒ HasSimpleRep (StrictMaybe a)

Instance details

Associated Types

type SimpleRep (StrictMaybe a)

type TheSop (StrictMaybe a) ∷ [Type]

Methods

toSimpleRep ∷ StrictMaybe a → SimpleRep (StrictMaybe a)

fromSimpleRep ∷ SimpleRep (StrictMaybe a) → StrictMaybe a