Test.Cardano.Ledger.Constrained.Conway.Instances.TxBody

Documentation

type ShelleyTxBodyTypes era = '[Set (TxIn (EraCrypto era)), [TxOut era], [TxCert era], Map (RewardAccount (EraCrypto era)) Coin, Coin, SlotNo, Maybe (Update era), Maybe (AuxiliaryDataHash (EraCrypto era))]

This is an abstraction of the Pattern ShelleyTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromShelleyBody ∷ ∀ era. EraTxBody era ⇒ ShelleyTxBody era → TxBody era

type AllegraTxBodyTypes era = '[Set (TxIn (EraCrypto era)), [TxOut era], [TxCert era], Map (RewardAccount (EraCrypto era)) Coin, Coin, ValidityInterval, Maybe (Update era), Maybe (AuxiliaryDataHash (EraCrypto era))]

This is an abstraction of the Pattern AllegraTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromAllegraBody ∷ ∀ era. AllegraEraTxBody era ⇒ AllegraTxBody era → TxBody era

type MaryTxBodyTypes era = '[Set (TxIn (EraCrypto era)), [TxOut era], [TxCert era], Map (RewardAccount (EraCrypto era)) Coin, Coin, ValidityInterval, Maybe (Update era), Maybe (AuxiliaryDataHash (EraCrypto era)), MultiAsset (EraCrypto era)]

This is an abstraction of the Pattern MaryTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromMaryBody ∷ ∀ era. MaryEraTxBody era ⇒ MaryTxBody era → TxBody era

type AlonzoTxBodyTypes era = '[Set (TxIn (EraCrypto era)), Set (TxIn (EraCrypto era)), [TxOut era], [TxCert era], Map (RewardAccount (EraCrypto era)) Coin, Coin, ValidityInterval, Maybe (Update era), Set (KeyHash 'Witness (EraCrypto era)), MultiAsset (EraCrypto era), Maybe (ScriptIntegrityHash (EraCrypto era)), Maybe (AuxiliaryDataHash (EraCrypto era)), Maybe Network]

This is an abstraction of the Pattern AlonzoTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromAlonzoBody ∷ ∀ era. AlonzoEraTxBody era ⇒ AlonzoTxBody era → TxBody era

type BabbageTxBodyTypes era = '[Set (TxIn (EraCrypto era)), Set (TxIn (EraCrypto era)), Set (TxIn (EraCrypto era)), [Sized (TxOut era)], Maybe (Sized (TxOut era)), Maybe Coin, [TxCert era], Map (RewardAccount (EraCrypto era)) Coin, Coin, ValidityInterval, Maybe (Update era), Set (KeyHash 'Witness (EraCrypto era)), MultiAsset (EraCrypto era), Maybe (ScriptIntegrityHash (EraCrypto era)), Maybe (AuxiliaryDataHash (EraCrypto era)), Maybe Network]

This is an abstraction of the Pattern BabbageTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

fromBabbageBody ∷ ∀ era. BabbageEraTxBody era ⇒ BabbageTxBody era → TxBody era

Orphan instances

(EraSpecPParams era, IsConwayUniv fn, HasSpec fn (TxOut era), HasSpec fn (TxCert era)) ⇒ HasSpec fn (AllegraTxBody era)

Instance details

Associated Types

type TypeSpec fn (AllegraTxBody era)

type Prerequisites fn (AllegraTxBody era)

Methods

emptySpec ∷ TypeSpec fn (AllegraTxBody era)

combineSpec ∷ TypeSpec fn (AllegraTxBody era) → TypeSpec fn (AllegraTxBody era) → Specification fn (AllegraTxBody era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (AllegraTxBody era) → GenT m (AllegraTxBody era)

conformsTo ∷ AllegraTxBody era → TypeSpec fn (AllegraTxBody era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (AllegraTxBody era) → AllegraTxBody era → [AllegraTxBody era]

toPreds ∷ Term fn (AllegraTxBody era) → TypeSpec fn (AllegraTxBody era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (AllegraTxBody era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (AllegraTxBody era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (AllegraTxBody era) → BinaryShow

monadConformsTo ∷ AllegraTxBody era → TypeSpec fn (AllegraTxBody era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (AllegraTxBody era) → [AllegraTxBody era] → Specification fn (AllegraTxBody era)

prerequisites ∷ Evidence (Prerequisites fn (AllegraTxBody era))

(EraSpecPParams era, IsConwayUniv fn, HasSpec fn (TxOut era), HasSpec fn (TxCert era)) ⇒ HasSpec fn (AlonzoTxBody era)

Instance details

Associated Types

type TypeSpec fn (AlonzoTxBody era)

type Prerequisites fn (AlonzoTxBody era)

Methods

emptySpec ∷ TypeSpec fn (AlonzoTxBody era)

combineSpec ∷ TypeSpec fn (AlonzoTxBody era) → TypeSpec fn (AlonzoTxBody era) → Specification fn (AlonzoTxBody era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (AlonzoTxBody era) → GenT m (AlonzoTxBody era)

conformsTo ∷ AlonzoTxBody era → TypeSpec fn (AlonzoTxBody era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (AlonzoTxBody era) → AlonzoTxBody era → [AlonzoTxBody era]

toPreds ∷ Term fn (AlonzoTxBody era) → TypeSpec fn (AlonzoTxBody era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (AlonzoTxBody era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (AlonzoTxBody era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (AlonzoTxBody era) → BinaryShow

monadConformsTo ∷ AlonzoTxBody era → TypeSpec fn (AlonzoTxBody era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (AlonzoTxBody era) → [AlonzoTxBody era] → Specification fn (AlonzoTxBody era)

prerequisites ∷ Evidence (Prerequisites fn (AlonzoTxBody era))

(EraSpecPParams era, BabbageEraTxBody era, IsConwayUniv fn, HasSpec fn (TxOut era), HasSpec fn (TxCert era)) ⇒ HasSpec fn (BabbageTxBody era)

Instance details

Associated Types

type TypeSpec fn (BabbageTxBody era)

type Prerequisites fn (BabbageTxBody era)

Methods

emptySpec ∷ TypeSpec fn (BabbageTxBody era)

combineSpec ∷ TypeSpec fn (BabbageTxBody era) → TypeSpec fn (BabbageTxBody era) → Specification fn (BabbageTxBody era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (BabbageTxBody era) → GenT m (BabbageTxBody era)

conformsTo ∷ BabbageTxBody era → TypeSpec fn (BabbageTxBody era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (BabbageTxBody era) → BabbageTxBody era → [BabbageTxBody era]

toPreds ∷ Term fn (BabbageTxBody era) → TypeSpec fn (BabbageTxBody era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (BabbageTxBody era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (BabbageTxBody era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (BabbageTxBody era) → BinaryShow

monadConformsTo ∷ BabbageTxBody era → TypeSpec fn (BabbageTxBody era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (BabbageTxBody era) → [BabbageTxBody era] → Specification fn (BabbageTxBody era)

prerequisites ∷ Evidence (Prerequisites fn (BabbageTxBody era))

(EraSpecPParams era, IsConwayUniv fn, HasSpec fn (TxOut era), HasSpec fn (TxCert era)) ⇒ HasSpec fn (MaryTxBody era)

Instance details

Associated Types

type TypeSpec fn (MaryTxBody era)

type Prerequisites fn (MaryTxBody era)

Methods

emptySpec ∷ TypeSpec fn (MaryTxBody era)

combineSpec ∷ TypeSpec fn (MaryTxBody era) → TypeSpec fn (MaryTxBody era) → Specification fn (MaryTxBody era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (MaryTxBody era) → GenT m (MaryTxBody era)

conformsTo ∷ MaryTxBody era → TypeSpec fn (MaryTxBody era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (MaryTxBody era) → MaryTxBody era → [MaryTxBody era]

toPreds ∷ Term fn (MaryTxBody era) → TypeSpec fn (MaryTxBody era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (MaryTxBody era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (MaryTxBody era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (MaryTxBody era) → BinaryShow

monadConformsTo ∷ MaryTxBody era → TypeSpec fn (MaryTxBody era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (MaryTxBody era) → [MaryTxBody era] → Specification fn (MaryTxBody era)

prerequisites ∷ Evidence (Prerequisites fn (MaryTxBody era))

(EraSpecPParams era, IsConwayUniv fn) ⇒ HasSpec fn (Update era)

Instance details

Associated Types

type TypeSpec fn (Update era)

type Prerequisites fn (Update era)

Methods

emptySpec ∷ TypeSpec fn (Update era)

combineSpec ∷ TypeSpec fn (Update era) → TypeSpec fn (Update era) → Specification fn (Update era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (Update era) → GenT m (Update era)

conformsTo ∷ Update era → TypeSpec fn (Update era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (Update era) → Update era → [Update era]

toPreds ∷ Term fn (Update era) → TypeSpec fn (Update era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (Update era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (Update era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (Update era) → BinaryShow

monadConformsTo ∷ Update era → TypeSpec fn (Update era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (Update era) → [Update era] → Specification fn (Update era)

prerequisites ∷ Evidence (Prerequisites fn (Update era))

(EraSpecPParams era, IsConwayUniv fn, HasSpec fn (TxOut era), HasSpec fn (TxCert era)) ⇒ HasSpec fn (ShelleyTxBody era)

Instance details

Associated Types

type TypeSpec fn (ShelleyTxBody era)

type Prerequisites fn (ShelleyTxBody era)

Methods

emptySpec ∷ TypeSpec fn (ShelleyTxBody era)

combineSpec ∷ TypeSpec fn (ShelleyTxBody era) → TypeSpec fn (ShelleyTxBody era) → Specification fn (ShelleyTxBody era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (ShelleyTxBody era) → GenT m (ShelleyTxBody era)

conformsTo ∷ ShelleyTxBody era → TypeSpec fn (ShelleyTxBody era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (ShelleyTxBody era) → ShelleyTxBody era → [ShelleyTxBody era]

toPreds ∷ Term fn (ShelleyTxBody era) → TypeSpec fn (ShelleyTxBody era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (ShelleyTxBody era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (ShelleyTxBody era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (ShelleyTxBody era) → BinaryShow

monadConformsTo ∷ ShelleyTxBody era → TypeSpec fn (ShelleyTxBody era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (ShelleyTxBody era) → [ShelleyTxBody era] → Specification fn (ShelleyTxBody era)

prerequisites ∷ Evidence (Prerequisites fn (ShelleyTxBody era))

(EraTxOut era, EraTxCert era) ⇒ HasSimpleRep (AllegraTxBody era)

Instance details

Associated Types

type SimpleRep (AllegraTxBody era)

type TheSop (AllegraTxBody era) ∷ [Type]

Methods

toSimpleRep ∷ AllegraTxBody era → SimpleRep (AllegraTxBody era)

fromSimpleRep ∷ SimpleRep (AllegraTxBody era) → AllegraTxBody era

(EraTxOut era, EraTxCert era) ⇒ HasSimpleRep (AlonzoTxBody era)

Instance details

Associated Types

type SimpleRep (AlonzoTxBody era)

type TheSop (AlonzoTxBody era) ∷ [Type]

Methods

toSimpleRep ∷ AlonzoTxBody era → SimpleRep (AlonzoTxBody era)

fromSimpleRep ∷ SimpleRep (AlonzoTxBody era) → AlonzoTxBody era

(EraTxOut era, EraTxCert era, BabbageEraTxBody era) ⇒ HasSimpleRep (BabbageTxBody era)

Instance details

Associated Types

type SimpleRep (BabbageTxBody era)

type TheSop (BabbageTxBody era) ∷ [Type]

Methods

toSimpleRep ∷ BabbageTxBody era → SimpleRep (BabbageTxBody era)

fromSimpleRep ∷ SimpleRep (BabbageTxBody era) → BabbageTxBody era

(EraTxOut era, EraTxCert era) ⇒ HasSimpleRep (MaryTxBody era)

Instance details

Associated Types

type SimpleRep (MaryTxBody era)

type TheSop (MaryTxBody era) ∷ [Type]

Methods

toSimpleRep ∷ MaryTxBody era → SimpleRep (MaryTxBody era)

fromSimpleRep ∷ SimpleRep (MaryTxBody era) → MaryTxBody era

HasSimpleRep (Update era)

Instance details

Associated Types

type SimpleRep (Update era)

type TheSop (Update era) ∷ [Type]

Methods

toSimpleRep ∷ Update era → SimpleRep (Update era)

fromSimpleRep ∷ SimpleRep (Update era) → Update era

(EraTxOut era, EncCBOR (TxCert era)) ⇒ HasSimpleRep (ShelleyTxBody era)

Instance details

Associated Types

type SimpleRep (ShelleyTxBody era)

type TheSop (ShelleyTxBody era) ∷ [Type]

Methods

toSimpleRep ∷ ShelleyTxBody era → SimpleRep (ShelleyTxBody era)

fromSimpleRep ∷ SimpleRep (ShelleyTxBody era) → ShelleyTxBody era