Test.Cardano.Ledger.Constrained.Conway.InstancesBasic

Description

This module provides HasSpec and HasSimpleRep instances for Basic types. A type is Basic if it is used to define PParams. See Test.Cardano.Ledger.Constrained.Conway.SimplePParams We divide these HasSpec and HasSimpleRep instances into two files because SimplePParams, needs these instances but not the 100's of other ones defined in Test.Cardano.Ledger.Constrained.Conway.Instances. And too many instances causes GHC 8.10.7 to blow up.

Documentation

cSNothing_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn (StrictMaybe a)

cSJust_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn a → Term fn (StrictMaybe a)

succV_ ∷ BaseUniverse fn ⇒ Term fn Version → Term fn Version

makePrices ∷ Integer → Integer → Prices

makeUnitInterval ∷ Integer → Integer → UnitInterval

makeNonNegativeInterval ∷ Integer → Integer → NonNegativeInterval

data SimplePParams era

Use this as the SimpleRep of (PParams era). It is polymorphic enough to encode PParams in EVERY Era. The EraPParams instances remove the fields that do not appear in that Era.

Constructors

SimplePParams

Fields

minFeeA ∷ Coin

minFeeB ∷ Coin

maxBBSize ∷ Word32

maxTxSize ∷ Word32

maxBHSize ∷ Word32

keyDeposit ∷ Coin

poolDeposit ∷ Coin

eMax ∷ EpochInterval

nOpt ∷ Natural

a0 ∷ NonNegativeInterval

rho ∷ UnitInterval

tau ∷ UnitInterval

decentral ∷ UnitInterval

protocolVersion ∷ ProtVer

minUTxOValue ∷ Coin

minPoolCost ∷ Coin

Alonzo

coinsPerUTxOWord ∷ Coin

costModels ∷ CostModels

prices ∷ Prices

maxTxExUnits ∷ ExUnits

maxBlockExUnits ∷ ExUnits

maxValSize ∷ Natural

collateralPercentage ∷ Natural

maxCollateralInputs ∷ Natural

Babbage

coinsPerUTxOByte ∷ Coin

Conway

poolVotingThresholds ∷ PoolVotingThresholds

drepVotingThresholds ∷ DRepVotingThresholds

committeeMinSize ∷ Natural

committeeMaxTermLength ∷ EpochInterval

govActionLifetime ∷ EpochInterval

govActionDeposit ∷ Coin

dRepDeposit ∷ Coin

dRepActivity ∷ EpochInterval

minFeeRefScriptCostPerByte ∷ NonNegativeInterval

data SimplePPUpdate

Use this as the SimpleRep of (PParamsUpdate era)

Constructors

SimplePPUpdate

Fields

uminFeeA ∷ StrictMaybe Coin

uminFeeB ∷ StrictMaybe Coin

umaxBBSize ∷ StrictMaybe Word32

umaxTxSize ∷ StrictMaybe Word32

umaxBHSize ∷ StrictMaybe Word32

ukeyDeposit ∷ StrictMaybe Coin

upoolDeposit ∷ StrictMaybe Coin

ueMax ∷ StrictMaybe EpochInterval

unOpt ∷ StrictMaybe Natural

ua0 ∷ StrictMaybe NonNegativeInterval

urho ∷ StrictMaybe UnitInterval

utau ∷ StrictMaybe UnitInterval

udecentral ∷ StrictMaybe UnitInterval

uprotocolVersion ∷ StrictMaybe ProtVer

uminUTxOValue ∷ StrictMaybe Coin

uminPoolCost ∷ StrictMaybe Coin

ucoinsPerUTxOWord ∷ StrictMaybe Coin

ucostModels ∷ StrictMaybe CostModels

uprices ∷ StrictMaybe Prices

umaxTxExUnits ∷ StrictMaybe ExUnits

umaxBlockExUnits ∷ StrictMaybe ExUnits

umaxValSize ∷ StrictMaybe Natural

ucollateralPercentage ∷ StrictMaybe Natural

umaxCollateralInputs ∷ StrictMaybe Natural

ucoinsPerUTxOByte ∷ StrictMaybe Coin

upoolVotingThresholds ∷ StrictMaybe PoolVotingThresholds

udrepVotingThresholds ∷ StrictMaybe DRepVotingThresholds

ucommitteeMinSize ∷ StrictMaybe Natural

ucommitteeMaxTermLength ∷ StrictMaybe EpochInterval

ugovActionLifetime ∷ StrictMaybe EpochInterval

ugovActionDeposit ∷ StrictMaybe Coin

udRepDeposit ∷ StrictMaybe Coin

udRepActivity ∷ StrictMaybe EpochInterval

uminFeeRefScriptCostPerByte ∷ StrictMaybe NonNegativeInterval

class (Reflect era, Eq (PParamsHKD Identity era), Show (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), EraPParams era) ⇒ EraPP era where

Methods

subsetToPP ∷ SimplePParams era → PParams era

ppToSubset ∷ PParams era → SimplePParams era

updateToPPU ∷ SimplePPUpdate → PParamsUpdate era

ppuToUpdate ∷ PParamsUpdate era → SimplePPUpdate

Orphan instances

Num Coin

Instance details

Methods

(+) ∷ Coin → Coin → Coin

(-) ∷ Coin → Coin → Coin

(*) ∷ Coin → Coin → Coin

negate ∷ Coin → Coin

abs ∷ Coin → Coin

signum ∷ Coin → Coin

fromInteger ∷ Integer → Coin

HasSimpleRep OrdExUnits

Instance details

Associated Types

type SimpleRep OrdExUnits

type TheSop OrdExUnits ∷ [Type]

Methods

toSimpleRep ∷ OrdExUnits → SimpleRep OrdExUnits

fromSimpleRep ∷ SimpleRep OrdExUnits → OrdExUnits

HasSimpleRep Version

Instance details

Associated Types

type SimpleRep Version

type TheSop Version ∷ [Type]

Methods

toSimpleRep ∷ Version → SimpleRep Version

fromSimpleRep ∷ SimpleRep Version → Version

HasSimpleRep DRepVotingThresholds

Instance details

Associated Types

type SimpleRep DRepVotingThresholds

type TheSop DRepVotingThresholds ∷ [Type]

Methods

toSimpleRep ∷ DRepVotingThresholds → SimpleRep DRepVotingThresholds

fromSimpleRep ∷ SimpleRep DRepVotingThresholds → DRepVotingThresholds

HasSimpleRep PoolVotingThresholds

Instance details

Associated Types

type SimpleRep PoolVotingThresholds

type TheSop PoolVotingThresholds ∷ [Type]

Methods

toSimpleRep ∷ PoolVotingThresholds → SimpleRep PoolVotingThresholds

fromSimpleRep ∷ SimpleRep PoolVotingThresholds → PoolVotingThresholds

HasSimpleRep ProtVer

Instance details

Associated Types

type SimpleRep ProtVer

type TheSop ProtVer ∷ [Type]

Methods

toSimpleRep ∷ ProtVer → SimpleRep ProtVer

fromSimpleRep ∷ SimpleRep ProtVer → ProtVer

HasSimpleRep Coin

Instance details

Associated Types

type SimpleRep Coin

type TheSop Coin ∷ [Type]

Methods

toSimpleRep ∷ Coin → SimpleRep Coin

fromSimpleRep ∷ SimpleRep Coin → Coin

HasSimpleRep CostModels

Instance details

Associated Types

type SimpleRep CostModels

type TheSop CostModels ∷ [Type]

Methods

toSimpleRep ∷ CostModels → SimpleRep CostModels

fromSimpleRep ∷ SimpleRep CostModels → CostModels

HasSimpleRep ExUnits

Instance details

Associated Types

type SimpleRep ExUnits

type TheSop ExUnits ∷ [Type]

Methods

toSimpleRep ∷ ExUnits → SimpleRep ExUnits

fromSimpleRep ∷ SimpleRep ExUnits → ExUnits

HasSimpleRep Prices

Instance details

Associated Types

type SimpleRep Prices

type TheSop Prices ∷ [Type]

Methods

toSimpleRep ∷ Prices → SimpleRep Prices

fromSimpleRep ∷ SimpleRep Prices → Prices

HasSimpleRep EpochInterval

Instance details

Associated Types

type SimpleRep EpochInterval

type TheSop EpochInterval ∷ [Type]

Methods

toSimpleRep ∷ EpochInterval → SimpleRep EpochInterval

fromSimpleRep ∷ SimpleRep EpochInterval → EpochInterval

BaseUniverse fn ⇒ Foldy fn Coin

Instance details

Methods

genList ∷ ∀ (m ∷ Type → Type). (BaseUniverse fn, MonadGenError m) ⇒ Specification fn Coin → Specification fn Coin → GenT m [Coin]

theAddFn ∷ fn '[Coin, Coin] Coin

theZero ∷ Coin

BaseUniverse fn ⇒ HasSpec fn OrdExUnits

Instance details

Associated Types

type TypeSpec fn OrdExUnits

type Prerequisites fn OrdExUnits

Methods

emptySpec ∷ TypeSpec fn OrdExUnits

combineSpec ∷ TypeSpec fn OrdExUnits → TypeSpec fn OrdExUnits → Specification fn OrdExUnits

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn OrdExUnits → GenT m OrdExUnits

conformsTo ∷ OrdExUnits → TypeSpec fn OrdExUnits → Bool

shrinkWithTypeSpec ∷ TypeSpec fn OrdExUnits → OrdExUnits → [OrdExUnits]

toPreds ∷ Term fn OrdExUnits → TypeSpec fn OrdExUnits → Pred fn

cardinalTypeSpec ∷ TypeSpec fn OrdExUnits → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn OrdExUnits → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn OrdExUnits → BinaryShow

monadConformsTo ∷ OrdExUnits → TypeSpec fn OrdExUnits → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn OrdExUnits → [OrdExUnits] → Specification fn OrdExUnits

prerequisites ∷ Evidence (Prerequisites fn OrdExUnits)

BaseUniverse fn ⇒ HasSpec fn Version

Instance details

Associated Types

type TypeSpec fn Version

type Prerequisites fn Version

Methods

emptySpec ∷ TypeSpec fn Version

combineSpec ∷ TypeSpec fn Version → TypeSpec fn Version → Specification fn Version

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn Version → GenT m Version

conformsTo ∷ Version → TypeSpec fn Version → Bool

shrinkWithTypeSpec ∷ TypeSpec fn Version → Version → [Version]

toPreds ∷ Term fn Version → TypeSpec fn Version → Pred fn

cardinalTypeSpec ∷ TypeSpec fn Version → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn Version → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn Version → BinaryShow

monadConformsTo ∷ Version → TypeSpec fn Version → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn Version → [Version] → Specification fn Version

prerequisites ∷ Evidence (Prerequisites fn Version)

BaseUniverse fn ⇒ HasSpec fn DRepVotingThresholds

Instance details

Associated Types

type TypeSpec fn DRepVotingThresholds

type Prerequisites fn DRepVotingThresholds

Methods

emptySpec ∷ TypeSpec fn DRepVotingThresholds

combineSpec ∷ TypeSpec fn DRepVotingThresholds → TypeSpec fn DRepVotingThresholds → Specification fn DRepVotingThresholds

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn DRepVotingThresholds → GenT m DRepVotingThresholds

conformsTo ∷ DRepVotingThresholds → TypeSpec fn DRepVotingThresholds → Bool

shrinkWithTypeSpec ∷ TypeSpec fn DRepVotingThresholds → DRepVotingThresholds → [DRepVotingThresholds]

toPreds ∷ Term fn DRepVotingThresholds → TypeSpec fn DRepVotingThresholds → Pred fn

cardinalTypeSpec ∷ TypeSpec fn DRepVotingThresholds → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn DRepVotingThresholds → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn DRepVotingThresholds → BinaryShow

monadConformsTo ∷ DRepVotingThresholds → TypeSpec fn DRepVotingThresholds → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn DRepVotingThresholds → [DRepVotingThresholds] → Specification fn DRepVotingThresholds

prerequisites ∷ Evidence (Prerequisites fn DRepVotingThresholds)

BaseUniverse fn ⇒ HasSpec fn PoolVotingThresholds

Instance details

Associated Types

type TypeSpec fn PoolVotingThresholds

type Prerequisites fn PoolVotingThresholds

Methods

emptySpec ∷ TypeSpec fn PoolVotingThresholds

combineSpec ∷ TypeSpec fn PoolVotingThresholds → TypeSpec fn PoolVotingThresholds → Specification fn PoolVotingThresholds

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn PoolVotingThresholds → GenT m PoolVotingThresholds

conformsTo ∷ PoolVotingThresholds → TypeSpec fn PoolVotingThresholds → Bool

shrinkWithTypeSpec ∷ TypeSpec fn PoolVotingThresholds → PoolVotingThresholds → [PoolVotingThresholds]

toPreds ∷ Term fn PoolVotingThresholds → TypeSpec fn PoolVotingThresholds → Pred fn

cardinalTypeSpec ∷ TypeSpec fn PoolVotingThresholds → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn PoolVotingThresholds → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn PoolVotingThresholds → BinaryShow

monadConformsTo ∷ PoolVotingThresholds → TypeSpec fn PoolVotingThresholds → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn PoolVotingThresholds → [PoolVotingThresholds] → Specification fn PoolVotingThresholds

prerequisites ∷ Evidence (Prerequisites fn PoolVotingThresholds)

BaseUniverse fn ⇒ HasSpec fn NonNegativeInterval

Instance details

Associated Types

type TypeSpec fn NonNegativeInterval

type Prerequisites fn NonNegativeInterval

Methods

emptySpec ∷ TypeSpec fn NonNegativeInterval

combineSpec ∷ TypeSpec fn NonNegativeInterval → TypeSpec fn NonNegativeInterval → Specification fn NonNegativeInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn NonNegativeInterval → GenT m NonNegativeInterval

conformsTo ∷ NonNegativeInterval → TypeSpec fn NonNegativeInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec fn NonNegativeInterval → NonNegativeInterval → [NonNegativeInterval]

toPreds ∷ Term fn NonNegativeInterval → TypeSpec fn NonNegativeInterval → Pred fn

cardinalTypeSpec ∷ TypeSpec fn NonNegativeInterval → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn NonNegativeInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn NonNegativeInterval → BinaryShow

monadConformsTo ∷ NonNegativeInterval → TypeSpec fn NonNegativeInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn NonNegativeInterval → [NonNegativeInterval] → Specification fn NonNegativeInterval

prerequisites ∷ Evidence (Prerequisites fn NonNegativeInterval)

BaseUniverse fn ⇒ HasSpec fn ProtVer

Instance details

Associated Types

type TypeSpec fn ProtVer

type Prerequisites fn ProtVer

Methods

emptySpec ∷ TypeSpec fn ProtVer

combineSpec ∷ TypeSpec fn ProtVer → TypeSpec fn ProtVer → Specification fn ProtVer

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn ProtVer → GenT m ProtVer

conformsTo ∷ ProtVer → TypeSpec fn ProtVer → Bool

shrinkWithTypeSpec ∷ TypeSpec fn ProtVer → ProtVer → [ProtVer]

toPreds ∷ Term fn ProtVer → TypeSpec fn ProtVer → Pred fn

cardinalTypeSpec ∷ TypeSpec fn ProtVer → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn ProtVer → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn ProtVer → BinaryShow

monadConformsTo ∷ ProtVer → TypeSpec fn ProtVer → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn ProtVer → [ProtVer] → Specification fn ProtVer

prerequisites ∷ Evidence (Prerequisites fn ProtVer)

BaseUniverse fn ⇒ HasSpec fn UnitInterval

Instance details

Associated Types

type TypeSpec fn UnitInterval

type Prerequisites fn UnitInterval

Methods

emptySpec ∷ TypeSpec fn UnitInterval

combineSpec ∷ TypeSpec fn UnitInterval → TypeSpec fn UnitInterval → Specification fn UnitInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn UnitInterval → GenT m UnitInterval

conformsTo ∷ UnitInterval → TypeSpec fn UnitInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec fn UnitInterval → UnitInterval → [UnitInterval]

toPreds ∷ Term fn UnitInterval → TypeSpec fn UnitInterval → Pred fn

cardinalTypeSpec ∷ TypeSpec fn UnitInterval → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn UnitInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn UnitInterval → BinaryShow

monadConformsTo ∷ UnitInterval → TypeSpec fn UnitInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn UnitInterval → [UnitInterval] → Specification fn UnitInterval

prerequisites ∷ Evidence (Prerequisites fn UnitInterval)

BaseUniverse fn ⇒ HasSpec fn Coin

Instance details

Associated Types

type TypeSpec fn Coin

type Prerequisites fn Coin

Methods

emptySpec ∷ TypeSpec fn Coin

combineSpec ∷ TypeSpec fn Coin → TypeSpec fn Coin → Specification fn Coin

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn Coin → GenT m Coin

conformsTo ∷ Coin → TypeSpec fn Coin → Bool

shrinkWithTypeSpec ∷ TypeSpec fn Coin → Coin → [Coin]

toPreds ∷ Term fn Coin → TypeSpec fn Coin → Pred fn

cardinalTypeSpec ∷ TypeSpec fn Coin → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn Coin → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn Coin → BinaryShow

monadConformsTo ∷ Coin → TypeSpec fn Coin → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn Coin → [Coin] → Specification fn Coin

prerequisites ∷ Evidence (Prerequisites fn Coin)

BaseUniverse fn ⇒ HasSpec fn CostModels

Instance details

Associated Types

type TypeSpec fn CostModels

type Prerequisites fn CostModels

Methods

emptySpec ∷ TypeSpec fn CostModels

combineSpec ∷ TypeSpec fn CostModels → TypeSpec fn CostModels → Specification fn CostModels

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn CostModels → GenT m CostModels

conformsTo ∷ CostModels → TypeSpec fn CostModels → Bool

shrinkWithTypeSpec ∷ TypeSpec fn CostModels → CostModels → [CostModels]

toPreds ∷ Term fn CostModels → TypeSpec fn CostModels → Pred fn

cardinalTypeSpec ∷ TypeSpec fn CostModels → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn CostModels → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn CostModels → BinaryShow

monadConformsTo ∷ CostModels → TypeSpec fn CostModels → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn CostModels → [CostModels] → Specification fn CostModels

prerequisites ∷ Evidence (Prerequisites fn CostModels)

BaseUniverse fn ⇒ HasSpec fn ExUnits

Instance details

Associated Types

type TypeSpec fn ExUnits

type Prerequisites fn ExUnits

Methods

emptySpec ∷ TypeSpec fn ExUnits

combineSpec ∷ TypeSpec fn ExUnits → TypeSpec fn ExUnits → Specification fn ExUnits

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn ExUnits → GenT m ExUnits

conformsTo ∷ ExUnits → TypeSpec fn ExUnits → Bool

shrinkWithTypeSpec ∷ TypeSpec fn ExUnits → ExUnits → [ExUnits]

toPreds ∷ Term fn ExUnits → TypeSpec fn ExUnits → Pred fn

cardinalTypeSpec ∷ TypeSpec fn ExUnits → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn ExUnits → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn ExUnits → BinaryShow

monadConformsTo ∷ ExUnits → TypeSpec fn ExUnits → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn ExUnits → [ExUnits] → Specification fn ExUnits

prerequisites ∷ Evidence (Prerequisites fn ExUnits)

BaseUniverse fn ⇒ HasSpec fn Prices

Instance details

Associated Types

type TypeSpec fn Prices

type Prerequisites fn Prices

Methods

emptySpec ∷ TypeSpec fn Prices

combineSpec ∷ TypeSpec fn Prices → TypeSpec fn Prices → Specification fn Prices

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn Prices → GenT m Prices

conformsTo ∷ Prices → TypeSpec fn Prices → Bool

shrinkWithTypeSpec ∷ TypeSpec fn Prices → Prices → [Prices]

toPreds ∷ Term fn Prices → TypeSpec fn Prices → Pred fn

cardinalTypeSpec ∷ TypeSpec fn Prices → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn Prices → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn Prices → BinaryShow

monadConformsTo ∷ Prices → TypeSpec fn Prices → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn Prices → [Prices] → Specification fn Prices

prerequisites ∷ Evidence (Prerequisites fn Prices)

BaseUniverse fn ⇒ HasSpec fn EpochInterval

Instance details

Associated Types

type TypeSpec fn EpochInterval

type Prerequisites fn EpochInterval

Methods

emptySpec ∷ TypeSpec fn EpochInterval

combineSpec ∷ TypeSpec fn EpochInterval → TypeSpec fn EpochInterval → Specification fn EpochInterval

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn EpochInterval → GenT m EpochInterval

conformsTo ∷ EpochInterval → TypeSpec fn EpochInterval → Bool

shrinkWithTypeSpec ∷ TypeSpec fn EpochInterval → EpochInterval → [EpochInterval]

toPreds ∷ Term fn EpochInterval → TypeSpec fn EpochInterval → Pred fn

cardinalTypeSpec ∷ TypeSpec fn EpochInterval → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn EpochInterval → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn EpochInterval → BinaryShow

monadConformsTo ∷ EpochInterval → TypeSpec fn EpochInterval → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn EpochInterval → [EpochInterval] → Specification fn EpochInterval

prerequisites ∷ Evidence (Prerequisites fn EpochInterval)

BaseUniverse fn ⇒ NumLike fn Coin

Instance details

Methods

subtractSpec ∷ Coin → TypeSpec fn Coin → Specification fn Coin

negateSpec ∷ TypeSpec fn Coin → Specification fn Coin

safeSubtract ∷ Coin → Coin → Maybe Coin

BaseUniverse fn ⇒ OrdLike fn Version

Instance details

Methods

leqSpec ∷ Version → Specification fn Version

ltSpec ∷ Version → Specification fn Version

geqSpec ∷ Version → Specification fn Version

gtSpec ∷ Version → Specification fn Version

BaseUniverse fn ⇒ OrdLike fn Coin

Instance details

Methods

leqSpec ∷ Coin → Specification fn Coin

ltSpec ∷ Coin → Specification fn Coin

geqSpec ∷ Coin → Specification fn Coin

gtSpec ∷ Coin → Specification fn Coin

BaseUniverse fn ⇒ OrdLike fn EpochInterval

Instance details

Methods

leqSpec ∷ EpochInterval → Specification fn EpochInterval

ltSpec ∷ EpochInterval → Specification fn EpochInterval

geqSpec ∷ EpochInterval → Specification fn EpochInterval

gtSpec ∷ EpochInterval → Specification fn EpochInterval

(BaseUniverse fn, EraPP era, HasSpec fn Coin) ⇒ HasSpec fn (PParams era)
HasSpec instance for PParams

Instance details

Associated Types

type TypeSpec fn (PParams era)

type Prerequisites fn (PParams era)

Methods

emptySpec ∷ TypeSpec fn (PParams era)

combineSpec ∷ TypeSpec fn (PParams era) → TypeSpec fn (PParams era) → Specification fn (PParams era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (PParams era) → GenT m (PParams era)

conformsTo ∷ PParams era → TypeSpec fn (PParams era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (PParams era) → PParams era → [PParams era]

toPreds ∷ Term fn (PParams era) → TypeSpec fn (PParams era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (PParams era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (PParams era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (PParams era) → BinaryShow

monadConformsTo ∷ PParams era → TypeSpec fn (PParams era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (PParams era) → [PParams era] → Specification fn (PParams era)

prerequisites ∷ Evidence (Prerequisites fn (PParams era))

(BaseUniverse fn, EraPP era) ⇒ HasSpec fn (PParamsUpdate era)
HasSpec instance for PParams

Instance details

Associated Types

type TypeSpec fn (PParamsUpdate era)

type Prerequisites fn (PParamsUpdate era)

Methods

emptySpec ∷ TypeSpec fn (PParamsUpdate era)

combineSpec ∷ TypeSpec fn (PParamsUpdate era) → TypeSpec fn (PParamsUpdate era) → Specification fn (PParamsUpdate era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (PParamsUpdate era) → GenT m (PParamsUpdate era)

conformsTo ∷ PParamsUpdate era → TypeSpec fn (PParamsUpdate era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (PParamsUpdate era) → PParamsUpdate era → [PParamsUpdate era]

toPreds ∷ Term fn (PParamsUpdate era) → TypeSpec fn (PParamsUpdate era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (PParamsUpdate era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (PParamsUpdate era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (PParamsUpdate era) → BinaryShow

monadConformsTo ∷ PParamsUpdate era → TypeSpec fn (PParamsUpdate era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (PParamsUpdate era) → [PParamsUpdate era] → Specification fn (PParamsUpdate era)

prerequisites ∷ Evidence (Prerequisites fn (PParamsUpdate era))

(EraPP era, BaseUniverse fn) ⇒ HasSpec fn (FuturePParams era)

Instance details

Associated Types

type TypeSpec fn (FuturePParams era)

type Prerequisites fn (FuturePParams era)

Methods

emptySpec ∷ TypeSpec fn (FuturePParams era)

combineSpec ∷ TypeSpec fn (FuturePParams era) → TypeSpec fn (FuturePParams era) → Specification fn (FuturePParams era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (FuturePParams era) → GenT m (FuturePParams era)

conformsTo ∷ FuturePParams era → TypeSpec fn (FuturePParams era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (FuturePParams era) → FuturePParams era → [FuturePParams era]

toPreds ∷ Term fn (FuturePParams era) → TypeSpec fn (FuturePParams era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (FuturePParams era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (FuturePParams era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (FuturePParams era) → BinaryShow

monadConformsTo ∷ FuturePParams era → TypeSpec fn (FuturePParams era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (FuturePParams era) → [FuturePParams era] → Specification fn (FuturePParams era)

prerequisites ∷ Evidence (Prerequisites fn (FuturePParams era))

(EraPP era, BaseUniverse fn) ⇒ HasSpec fn (ProposedPPUpdates era)

Instance details

Associated Types

type TypeSpec fn (ProposedPPUpdates era)

type Prerequisites fn (ProposedPPUpdates era)

Methods

emptySpec ∷ TypeSpec fn (ProposedPPUpdates era)

combineSpec ∷ TypeSpec fn (ProposedPPUpdates era) → TypeSpec fn (ProposedPPUpdates era) → Specification fn (ProposedPPUpdates era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (ProposedPPUpdates era) → GenT m (ProposedPPUpdates era)

conformsTo ∷ ProposedPPUpdates era → TypeSpec fn (ProposedPPUpdates era) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (ProposedPPUpdates era) → ProposedPPUpdates era → [ProposedPPUpdates era]

toPreds ∷ Term fn (ProposedPPUpdates era) → TypeSpec fn (ProposedPPUpdates era) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (ProposedPPUpdates era) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (ProposedPPUpdates era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (ProposedPPUpdates era) → BinaryShow

monadConformsTo ∷ ProposedPPUpdates era → TypeSpec fn (ProposedPPUpdates era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (ProposedPPUpdates era) → [ProposedPPUpdates era] → Specification fn (ProposedPPUpdates era)

prerequisites ∷ Evidence (Prerequisites fn (ProposedPPUpdates era))

(HasSpec fn a, IsNormalType a) ⇒ HasSpec fn (StrictMaybe a)

Instance details

Associated Types

type TypeSpec fn (StrictMaybe a)

type Prerequisites fn (StrictMaybe a)

Methods

emptySpec ∷ TypeSpec fn (StrictMaybe a)

combineSpec ∷ TypeSpec fn (StrictMaybe a) → TypeSpec fn (StrictMaybe a) → Specification fn (StrictMaybe a)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (StrictMaybe a) → GenT m (StrictMaybe a)

conformsTo ∷ StrictMaybe a → TypeSpec fn (StrictMaybe a) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (StrictMaybe a) → StrictMaybe a → [StrictMaybe a]

toPreds ∷ Term fn (StrictMaybe a) → TypeSpec fn (StrictMaybe a) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (StrictMaybe a) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (StrictMaybe a) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (StrictMaybe a) → BinaryShow

monadConformsTo ∷ StrictMaybe a → TypeSpec fn (StrictMaybe a) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (StrictMaybe a) → [StrictMaybe a] → Specification fn (StrictMaybe a)

prerequisites ∷ Evidence (Prerequisites fn (StrictMaybe a))

(BaseUniverse fn, Typeable r, Crypto c) ⇒ HasSpec fn (KeyHash r c)

Instance details

Associated Types

type TypeSpec fn (KeyHash r c)

type Prerequisites fn (KeyHash r c)

Methods

emptySpec ∷ TypeSpec fn (KeyHash r c)

combineSpec ∷ TypeSpec fn (KeyHash r c) → TypeSpec fn (KeyHash r c) → Specification fn (KeyHash r c)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (KeyHash r c) → GenT m (KeyHash r c)

conformsTo ∷ KeyHash r c → TypeSpec fn (KeyHash r c) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (KeyHash r c) → KeyHash r c → [KeyHash r c]

toPreds ∷ Term fn (KeyHash r c) → TypeSpec fn (KeyHash r c) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (KeyHash r c) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (KeyHash r c) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (KeyHash r c) → BinaryShow

monadConformsTo ∷ KeyHash r c → TypeSpec fn (KeyHash r c) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (KeyHash r c) → [KeyHash r c] → Specification fn (KeyHash r c)

prerequisites ∷ Evidence (Prerequisites fn (KeyHash r c))

EraPP era ⇒ HasSimpleRep (PParams era)
SimpleRep instance for PParams

Instance details

Associated Types

type SimpleRep (PParams era)

type TheSop (PParams era) ∷ [Type]

Methods

toSimpleRep ∷ PParams era → SimpleRep (PParams era)

fromSimpleRep ∷ SimpleRep (PParams era) → PParams era

EraPP era ⇒ HasSimpleRep (PParamsUpdate era)
SimpleRep instance for PParamsUpdate

Instance details

Associated Types

type SimpleRep (PParamsUpdate era)

type TheSop (PParamsUpdate era) ∷ [Type]

Methods

toSimpleRep ∷ PParamsUpdate era → SimpleRep (PParamsUpdate era)

fromSimpleRep ∷ SimpleRep (PParamsUpdate era) → PParamsUpdate era

EraPP era ⇒ HasSimpleRep (FuturePParams era)

Instance details

Associated Types

type SimpleRep (FuturePParams era)

type TheSop (FuturePParams era) ∷ [Type]

Methods

toSimpleRep ∷ FuturePParams era → SimpleRep (FuturePParams era)

fromSimpleRep ∷ SimpleRep (FuturePParams era) → FuturePParams era

EraPP era ⇒ HasSimpleRep (ProposedPPUpdates era)

Instance details

Associated Types

type SimpleRep (ProposedPPUpdates era)

type TheSop (ProposedPPUpdates era) ∷ [Type]

Methods

toSimpleRep ∷ ProposedPPUpdates era → SimpleRep (ProposedPPUpdates era)

fromSimpleRep ∷ SimpleRep (ProposedPPUpdates era) → ProposedPPUpdates era

HasSimpleRep (StrictMaybe a)

Instance details

Associated Types

type SimpleRep (StrictMaybe a)

type TheSop (StrictMaybe a) ∷ [Type]

Methods

toSimpleRep ∷ StrictMaybe a → SimpleRep (StrictMaybe a)

fromSimpleRep ∷ SimpleRep (StrictMaybe a) → StrictMaybe a