Test.Cardano.Ledger.Constrained.Conway.WitnessUniverse

Documentation

genSigningKey ∷ Gen SigningKey

Turn a random bytestring into a SigningKey

genAddrPair ∷ Network → Gen (BootstrapAddress, SigningKey)

Generate a pair, A Byron address, and the key that can sign it.

class (Eq (ProofType hashtype era), EncCBOR (ProofType hashtype era), ToExpr (ProofType hashtype era), NFData (ProofType hashtype era), NFData hashtype, Eq hashtype) ⇒ HasWitness hashtype era where

Associated Types

type ProofType hashtype era

type WitnessType hashtype era

type TypeHashed hashtype era

Methods

hash ∷ TypeHashed hashtype era → hashtype

mkWitness ∷ ProofType hashtype era → WitnessType hashtype era

getTypeHashed ∷ ProofType hashtype era → TypeHashed hashtype era

prettyHash ∷ hashtype → Doc

data WitBlock t era where

A WitBlock is designed to have five purposes 1) To efficiently constrain objects to be witnessed when using constraint generators the (Set hashtype) allows efficient constraints like :: (member_ t) (lit (wbHash witblock)) 2) To efficiently compute witnesses from a hashtype the (Map hashtype (ProofType hashtype era)) can be used like this case Map.lookup hash (wbMap witblock) of Just base -> mkWitness base Nothing -> error "Missing hash. perhaps generator did not constrain the hash to be witnessed?" 3) When (HasWitness hashtype era) holds, the WitBlock can be computed from only [ProofType hashtype era] using getTypeHashed. This makes Gen and CBOR instances, especially easy. We compute only with [ProofType hashtype era] and then reconstruct the rest 4) WitBlock is a Monoid, so we can combine them easily 5) We can easily make (Gen (WitBlock t era)), so we can make them for testing.

Constructors

WitBlock ∷ (Era era, HasWitness t era) ⇒ Set t → Map t (ProofType t era) → WitBlock t era

wbHash ∷ WitBlock t era → Set t

wbMap ∷ WitBlock t era → Map t (ProofType t era)

type BodyHash = SafeHash EraIndependentTxBody

data WitUniv era

Constructors

WitUniv

Fields

wvSize ∷ Int

wvVKey ∷ WitBlock (KeyHash 'Witness) era

wvBoot ∷ WitBlock BootstrapAddress era

wvScript ∷ WitBlock ScriptHash era

wvDats ∷ WitBlock DataHash era

explainWit ∷ String → WitUniv era → Specification t → Specification t

witKeyHashSpec ∷ ∀ era krole. Typeable krole ⇒ WitUniv era → Specification (KeyHash krole)

witScriptHashSpec ∷ ∀ era. WitUniv era → Specification ScriptHash

witBootstrapAddress ∷ ∀ era. WitUniv era → Specification BootstrapAddress

witCredSpec ∷ ∀ era krole. Typeable krole ⇒ WitUniv era → Specification (Credential krole)

witDRepSpec ∷ ∀ era. WitUniv era → Specification DRep

witRewardAccountSpec ∷ ∀ era. WitUniv era → Specification RewardAccount

Used only in Withdrawals, other RewardAccounts, not being withdrawn do not need witnessing

owners_ ∷ Term PoolParams → Term (Set (KeyHash 'Staking))

witPoolParamsSpec ∷ ∀ era. WitUniv era → Specification PoolParams

witPoolStateSpec ∷ ∀ era. WitUniv era → Specification StakePoolState

witGenDelegPairSpec ∷ ∀ era. WitUniv era → Specification GenDelegPair

witShelleyTxCert ∷ ∀ era. Era era ⇒ WitUniv era → Specification (ShelleyTxCert era)

Constrains all the Certificate Authors. Sometimes thay are keyHashes, and sometimes Credentials

witConwayTxCert ∷ ∀ era. Era era ⇒ WitUniv era → Specification (ConwayTxCert era)

Constrains all the Certificate Authors. Sometimes thay are keyHashes, and sometimes Credentials

witnessBootAddr ∷ ∀ era. EraTxWits era ⇒ BodyHash → BootstrapAddress → WitUniv era → TxWits era

witnessKeyHash ∷ ∀ era. EraTxWits era ⇒ BodyHash → KeyHash 'Witness → WitUniv era → TxWits era

witnessScriptHash ∷ ∀ era. EraTxWits era ⇒ ScriptHash → WitUniv era → TxWits era

witnessDataHash ∷ ∀ era. AlonzoEraTxWits era ⇒ DataHash → WitUniv era → TxWits era

blockFromProofList ∷ ∀ t era. (Era era, Ord t, HasWitness t era) ⇒ [ProofType t era] → WitBlock t era

Reconstruct a (WitBlock t era) from only a [ProofType t era]

genWitBlock ∷ ∀ t era. (Era era, Ord t, HasWitness t era) ⇒ Int → Gen (ProofType t era) → Gen (WitBlock t era)

genWitUniv ∷ ∀ era. (GenScript era, HasWitness ScriptHash era) ⇒ Int → Gen (WitUniv era)

genNestedMultiSig ∷ ∀ era. ShelleyEraScript era ⇒ Int → Gen (NativeScript era)

genNestedTimelock ∷ ∀ era. AllegraEraScript era ⇒ Int → Gen (NativeScript era)

class Witnessed era t where

The class of things we know how to witness. This way you don't have to remember long complicated names.

Methods

witness ∷ WitUniv era → Term t → Pred

class (EraTxCert era, HasSpec (TxCert era)) ⇒ EraSpecTxCert era where

Parametric TxCert

Methods

witTxCert ∷ WitUniv era → Specification (TxCert era)

class (EraScript era, ToExpr (NativeScript era), NFData (Script era), ToExpr (Script era)) ⇒ GenScript era where

Parametric Script

Methods

genScript ∷ Gen (Script era)

spec1 ∷ WitUniv ShelleyEra → Specification (Set ScriptHash)

go1 ∷ IO ()

spec2 ∷ WitUniv ShelleyEra → Set ScriptHash → Specification (Set ScriptHash)

go2 ∷ IO ()

conwayWitUniv ∷ Int → WitUniv ConwayEra

babbageWitUniv ∷ Int → WitUniv BabbageEra

alonzoWitUniv ∷ Int → WitUniv AlonzoEra

maryWitUniv ∷ Int → WitUniv MaryEra

allegraWitUniv ∷ Int → WitUniv AllegraEra

shelleyWitUniv ∷ Int → WitUniv ShelleyEra

class EraUniverse era where

Methods

eraWitUniv ∷ Int → WitUniv era

govActionStateWitness ∷ ∀ era. EraSpecPParams era ⇒ WitUniv era → Specification (GovActionState era)

Constrains just the parts that need witnessing in GovActionState

govActionWitness ∷ ∀ era. EraSpecPParams era ⇒ WitUniv era → Specification (GovAction era)

Constrains just the parts that need witnessing in GovAction

proposalProcedureWitness ∷ ∀ era. EraSpecPParams era ⇒ WitUniv era → Specification (ProposalProcedure era)

Constrains just the parts that need witnessing in ProposalProcedure

committeeWitness ∷ EraSpecPParams era ⇒ WitUniv era → Specification (Committee era)

Constrains just the parts that need witnessing in Committee

go9 ∷ IO ()

Orphan instances

ToExpr SigningKey

Instance details

Methods

toExpr ∷ SigningKey → Expr

listToExpr ∷ [SigningKey] → Expr

Era era ⇒ HasSpec (TxDats era)

Instance details

Associated Types

type TypeSpec (TxDats era)

type Prerequisites (TxDats era)

Methods

emptySpec ∷ TypeSpec (TxDats era)

combineSpec ∷ TypeSpec (TxDats era) → TypeSpec (TxDats era) → Specification (TxDats era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxDats era) → GenT m (TxDats era)

conformsTo ∷ TxDats era → TypeSpec (TxDats era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxDats era) → TxDats era → [TxDats era]

toPreds ∷ Term (TxDats era) → TypeSpec (TxDats era) → Pred

cardinalTypeSpec ∷ TypeSpec (TxDats era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxDats era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxDats era) → BinaryShow

monadConformsTo ∷ TxDats era → TypeSpec (TxDats era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxDats era) → [TxDats era] → Specification (TxDats era)

guardTypeSpec ∷ [String] → TypeSpec (TxDats era) → Specification (TxDats era)

prerequisites ∷ Evidence (Prerequisites (TxDats era))

(ShelleyEraScript era, NativeScript era ~ MultiSig era) ⇒ HasSpec (MultiSig era)

Instance details

Associated Types

type TypeSpec (MultiSig era)

type Prerequisites (MultiSig era)

Methods

emptySpec ∷ TypeSpec (MultiSig era)

combineSpec ∷ TypeSpec (MultiSig era) → TypeSpec (MultiSig era) → Specification (MultiSig era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (MultiSig era) → GenT m (MultiSig era)

conformsTo ∷ MultiSig era → TypeSpec (MultiSig era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (MultiSig era) → MultiSig era → [MultiSig era]

toPreds ∷ Term (MultiSig era) → TypeSpec (MultiSig era) → Pred

cardinalTypeSpec ∷ TypeSpec (MultiSig era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (MultiSig era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (MultiSig era) → BinaryShow

monadConformsTo ∷ MultiSig era → TypeSpec (MultiSig era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (MultiSig era) → [MultiSig era] → Specification (MultiSig era)

guardTypeSpec ∷ [String] → TypeSpec (MultiSig era) → Specification (MultiSig era)

prerequisites ∷ Evidence (Prerequisites (MultiSig era))

AlonzoEraScript era ⇒ HasSimpleRep (Redeemers era)

Instance details

Associated Types

type SimpleRep (Redeemers era)

type TheSop (Redeemers era) ∷ [Type]

Methods

toSimpleRep ∷ Redeemers era → SimpleRep (Redeemers era)

fromSimpleRep ∷ SimpleRep (Redeemers era) → Redeemers era

Era era ⇒ HasSimpleRep (TxDats era)

Instance details

Associated Types

type SimpleRep (TxDats era)

type TheSop (TxDats era) ∷ [Type]

Methods

toSimpleRep ∷ TxDats era → SimpleRep (TxDats era)

fromSimpleRep ∷ SimpleRep (TxDats era) → TxDats era