Test.Cardano.Ledger.Constrained.Env

Description

Provides variables (V era t), and mappings of them to objects of type t

Documentation

data V era t where

A proto variable. May or may not contain a Lens (encoded as Access)

Constructors

VRaw ∷ Era era ⇒ String → Rep era t → Int → Access era s t → V era t

Bundled Patterns

pattern V ∷ () ⇒ Era era ⇒ String → Rep era t → Access era s t → V era t
We make hashing of V cheap by memoizing most of the hash computation in in the 3rd argument of VRaw, we provide this pattern to hide this from every other use.

pV ∷ Proof era → String → Rep era t → Access era s t → V era t

Construct a V, dsicharging the (Era era) constraint, using the Proof

data Field era s t where

Fields are like V, except they expose the type of the Lens

Constructors

Field ∷ Era era ⇒ String → Rep era t → Rep era s → Lens' s t → Field era s t

FConst ∷ Rep era t → t → Rep era s → Lens' s t → Field era s t

data AnyF era s where

Hide the type of t in a Field.

Constructors

AnyF ∷ Field era s t → AnyF era s

vToField ∷ Era era ⇒ Rep era s → V era t → Typed (Field era s t)

fieldToV ∷ Era era ⇒ Field era s t → Typed (V era t)

newtype Env era

Constructors

Env (Map String (Payload era))

data Payload era where

Constructors

Payload ∷ Rep era t → t → Access era s t → Payload era

emptyEnv ∷ Env era

findVar ∷ V era t → Env era → Typed t

storeVar ∷ V era t → t → Env era → Env era

findName ∷ Name era → Env era → Maybe (Payload era)

Untyped version of findVar.

storeName ∷ Name era → Payload era → Env era → Env era

Untyped version of storeVar.

restrictEnv ∷ [Name era] → Env era → Env era

Drop any names that are not in the given list from an environment.

data P era where

Constructors

P ∷ V era t → t → P era

bulkStore ∷ [P era] → Env era → Env era

data Name era where

An existentially quantified (V era t), hiding the t The Hashable instance is inherited from V

Constructors

Name ∷ V era t → Name era

data Access era s t where

Constructors

Yes ∷ Rep era s → Lens' s t → Access era s t

No ∷ Access era s t

otherFromEnv ∷ [String] → Env era → [String]

sameName ∷ V era t → V era s → Maybe (t :~: s)

Orphan instances

Eq (Proof e)

Instance details

Methods

(==) ∷ Proof e → Proof e → Bool

(/=) ∷ Proof e → Proof e → Bool

Hashable (Shape a)
Inheriting the Hashable instancefor types that have Shapes, Works for types whose Shaped instances don't mention Esc, Like Evidence, Proof, Rep But we CANNOT use this strategy for (V era t) since its's Shape instance use Esc

Instance details

Methods

hashWithSalt ∷ Int → Shape a → Int

hash ∷ Shape a → Int

Hashable (Proof e)

Instance details

Methods

hashWithSalt ∷ Int → Proof e → Int

hash ∷ Proof e → Int

Eq (Rep e t)

Instance details

Methods

(==) ∷ Rep e t → Rep e t → Bool

(/=) ∷ Rep e t → Rep e t → Bool

Hashable (Rep e t)

Instance details

Methods

hashWithSalt ∷ Int → Rep e t → Int

hash ∷ Rep e t → Int