Test.Cardano.Ledger.Constrained.Examples

Documentation

runCompile ∷ Era era ⇒ [Pred era] → IO ()

data Assembler era where

Constructors

Assemble ∷ PrettyA t ⇒ RootTarget era x t → Assembler era

Skip ∷ Assembler era

stoi ∷ OrderInfo

genMaybeCounterExample ∷ Reflect era ⇒ Proof era → String → Bool → OrderInfo → [Pred era] → Assembler era → Gen (Maybe String)

testn ∷ Reflect era ⇒ Proof era → String → Bool → OrderInfo → [Pred era] → Assembler era → Gen Property

Test that cs :: [Pred] has a solution

failn ∷ Reflect era ⇒ Proof era → String → Bool → OrderInfo → [Pred era] → Assembler era → IO ()

Test that cs :: [Pred] does NOT have a solution. We expect a failure

testSpec ∷ String → Gen Property → IO ()

aCyclicPred ∷ Era era ⇒ [Pred era]

Used to test cyclic dependencies

cyclicPred ∷ Era era ⇒ [Pred era]

test1 ∷ IO ()

test3 ∷ Gen Property

test4 ∷ IO ()

test5 ∷ IO ()

constraints ∷ EraGov era ⇒ Proof era → [Pred era]

test6 ∷ Bool → IO ()

Test that we can find a viable variable ordering

test7 ∷ Bool → IO ()

Test that we can compute a solution

pstateConstraints ∷ Era era ⇒ [Pred era]

test8 ∷ Gen Property

sumPreds ∷ Reflect era ⇒ Proof era → [Pred era]

test9 ∷ Gen Property

test10 ∷ Gen Property

test11 ∷ Gen Property

test12 ∷ Gen Property

componentPreds ∷ EraGov era ⇒ Proof era → [Pred era]

test13 ∷ Gen Property

test14 ∷ IO ()

test15 ∷ Gen Property

preds16 ∷ Era era ⇒ Proof era → [Pred era]

test16 ∷ Gen Property

univPreds ∷ Era era ⇒ Proof era → [Pred era]

pstatePreds ∷ Era era ⇒ Proof era → [Pred era]

dstatePreds ∷ Era era ⇒ Proof era → [Pred era]

accountstatePreds ∷ Proof era → [Pred era]

utxostatePreds ∷ Reflect era ⇒ Proof era → [Pred era]

epochstatePreds ∷ EraGov era ⇒ Proof era → [Pred era]

newepochstatePreds ∷ Era era ⇒ Proof era → [Pred era]

newepochConstraints ∷ Reflect era ⇒ Proof era → [Pred era]

test17 ∷ Gen Property

projPreds1 ∷ Era era ⇒ Proof era → [Pred era]

projPreds2 ∷ Era era ⇒ Proof era → [Pred era]

test18a ∷ Gen Property

test18b ∷ Gen Property

help19 ∷ ∀ era. Era era ⇒ Proof era → Gen (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

test19 ∷ IO ()

preds20 ∷ Era era ⇒ Proof era → [Pred era]

test20 ∷ Gen Property

test21 ∷ Int → IO Bool

testAll ∷ IO ()

allExampleTests ∷ TestTree

demoPreds ∷ Reflect era ⇒ Proof era → [Pred era] → IO ()

Use to generate a solution to preds, test that the solution meets the preds, and puts you in the Repl to inspect the value of each variable solved for. Companion function to testPreds

testPreds ∷ Reflect era ⇒ String → Proof era → [Pred era] → TestTree

Use this to generate a solution to preds, and make a TestTree that tests the solution meets the preds. Companion function to demoPreds

listWherePreds ∷ [Pred (ConwayEra StandardCrypto)]

mainListWhere ∷ IO ()

govPreds ∷ [Pred (ConwayEra StandardCrypto)]

mainGov ∷ IO ()