Test.Cardano.Ledger.Constrained.Lenses

Description

Define Lenses that facilitate accessing the types in the Var Model. Note the types in the Model, are often wrapped in a newtype in the real state, or they are embedded in something like UMap to save space. So we need interesting Lenses to make this possible. Many other (more standard) Lenses are defined in Cardano.Ledger.Shelley.LedgerState

Documentation

iRReservesL ∷ Lens' (InstantaneousRewards c) (Map (Credential 'Staking c) Coin)

iRTreasuryL ∷ Lens' (InstantaneousRewards c) (Map (Credential 'Staking c) Coin)

deltaReservesL ∷ Lens' (InstantaneousRewards c) DeltaCoin

deltaTreasuryL ∷ Lens' (InstantaneousRewards c) DeltaCoin

unGenDelegsL ∷ Lens' (GenDelegs c) (Map (KeyHash 'Genesis c) (GenDelegPair c))

fGenDelegSlotL ∷ Lens' (FutureGenDeleg c) SlotNo

fGenDelegGenKeyHashL ∷ Lens' (FutureGenDeleg c) (KeyHash 'Genesis c)

isCredMapL ∷ Lens' (IncrementalStake c) (Map (Credential 'Staking c) Coin)

isPtrMapL ∷ Lens' (IncrementalStake c) (Map Ptr Coin)

nmLikelihoodsL ∷ Lens' (NonMyopic c) (Map (KeyHash 'StakePool c) [Float])

nmRewardPotL ∷ Lens' (NonMyopic c) Coin

spRewL ∷ Lens' (Split c) (Map (Credential 'Staking c) Coin)

spDepL ∷ Lens' (Split c) (Map (Credential 'Staking c) Coin)

spDelL ∷ Lens' (Split c) (Map (Credential 'Staking c) (KeyHash 'StakePool c))

spRevPtrL ∷ Lens' (Split c) (Map (Credential 'Staking c) (Set Ptr))

spPtrL ∷ Lens' (Split c) (Map Ptr (Credential 'Staking c))

spDRepL ∷ Lens' (Split c) (Map (Credential 'Staking c) (DRep c))

data Split c

Abstract view of the UMap

Constructors

Split

Fields

spRew ∷ Map (Credential 'Staking c) Coin

spDep ∷ Map (Credential 'Staking c) Coin

spDel ∷ Map (Credential 'Staking c) (KeyHash 'StakePool c)

spDRep ∷ Map (Credential 'Staking c) (DRep c)

spRevPtr ∷ Map (Credential 'Staking c) (Set Ptr)

spPtr ∷ Map Ptr (Credential 'Staking c)

splitUMap ∷ UMap c → Split c

The abstraction function, from concrete (UMap) to abstract (Split)

unSplitUMap ∷ Split c → UMap c

The concretization function from abstract (Split) to concrete (UMap)

rewardsUMapL ∷ Lens' (UMap c) (Map (Credential 'Staking c) Coin)

stakeDepositsUMapL ∷ Lens' (UMap c) (Map (Credential 'Staking c) Coin)

ptrsUMapL ∷ Lens' (UMap c) (Map Ptr (Credential 'Staking c))

delegationsUMapL ∷ Lens' (UMap c) (Map (Credential 'Staking c) (KeyHash 'StakePool c))

drepUMapL ∷ Lens' (UMap c) (Map (Credential 'Staking c) (DRep c))

strictMaybeToMaybeL ∷ Lens' (StrictMaybe x) (Maybe x)

idLens ∷ Lens' a a

strictSeqListL ∷ Lens' (StrictSeq a) [a]

mapCompactFormCoinL ∷ Lens' (Map a (CompactForm Coin)) (Map a Coin)

pairL ∷ Lens' x y → Lens' a b → Lens' (x, a) (y, b)

utxoFL ∷ Proof era → Lens' (Map (TxIn (EraCrypto era)) (TxOutF era)) (UTxO era)

Lens to convert from the abstract type UtxO type of the Model, to the concrete UTxO type. The mode uses the type family abstraction TxOutF, and does not wrap the map with the UtxO constructor. Note the getter is liftUTxO from Test.Cardano.Ledger.Constrained.Classes liftUTxO :: Map (TxIn (EraCrypto era)) (TxOutF era) -> UTxO era

fstL ∷ Lens' (a, b) a

sndL ∷ Lens' (a, b) b