Test.Cardano.Ledger.Constrained.Monad

Documentation

newtype Typed x

Constructors

Typed

Fields

runTyped ∷ Either [String] x

class LiftT x where

Methods

liftT ∷ x → Typed x

dropT ∷ Typed x → x

failT ∷ [String] → Typed a

explain ∷ String → Typed a → Typed a

mergeExplain ∷ (Monoid x, LiftT x) ⇒ String → x → x → x

data HasConstraint c t where

runTime evidence that the index i of an indexed type '(s i)' is constrained by the class c. If one has an un-indexed type t one can always use (Id t) instead. Eg With (Id x) <- hasOrd repT (Id t)

Constructors

With ∷ c t ⇒ s t → HasConstraint c (s t)

newtype Id x

Constructors

Id x

errorTyped ∷ HasCallStack ⇒ Typed t → t

Pushes the (Left msgs) into a call to error

monadTyped ∷ (HasCallStack, Monad m) ⇒ Typed t → m t

Pushes the (Left msgs) into a call to error, then injects into a Monad

requireAll ∷ [(Bool, [String])] → Typed a → Typed a

generateWithSeed ∷ Int → Gen a → IO a