Test.Cardano.Ledger.Constrained.Preds.Certs

Documentation

sRegKey ∷ Target era (StakeCredential (EraCrypto era) → ShelleyTxCert era)

sUnRegKey ∷ Target era (StakeCredential (EraCrypto era) → ShelleyTxCert era)

sDelegStake ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ShelleyTxCert era)

sRegPool ∷ Target era (PoolParams (EraCrypto era) → ShelleyTxCert era)

sRetirePool ∷ Target era (KeyHash 'StakePool (EraCrypto era) → EpochNo → ShelleyTxCert era)

sMirShift ∷ Target era (Coin → MIRPot → Coin → ShelleyTxCert era)

sGovern ∷ Target era (KeyHash 'Genesis (EraCrypto era) → KeyHash 'GenesisDelegate (EraCrypto era) → Hash (HASH (EraCrypto era)) (VerKeyVRF (VRF (EraCrypto era))) → ShelleyTxCert era)

cRegKey ∷ Target era (StakeCredential (EraCrypto era) → Maybe Coin → ConwayTxCert era)

cUnRegKey ∷ Target era (StakeCredential (EraCrypto era) → Maybe Coin → ConwayTxCert era)

cDelegStake ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ConwayTxCert era)

cDelegVote ∷ Target era (StakeCredential (EraCrypto era) → DRep (EraCrypto era) → a → ConwayTxCert era)

cDelegStakeVote ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → DRep (EraCrypto era) → a → ConwayTxCert era)

cRegDelegStake ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → Coin → ConwayTxCert era)

cRegDelegVote ∷ Target era (StakeCredential (EraCrypto era) → DRep (EraCrypto era) → Coin → a → ConwayTxCert era)

cRegDelegStakeVote ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → DRep (EraCrypto era) → Coin → a → ConwayTxCert era)

cRegPool ∷ Target era (PoolParams (EraCrypto era) → ConwayTxCert era)

cRetirePool ∷ Target era (KeyHash 'StakePool (EraCrypto era) → EpochNo → ConwayTxCert era)

partBfromPartA ∷ Ord k ⇒ Proof era → Map k Coin → Map k DeltaCoin

Transform some SubMap of instanReserves (or instanTreasury) into a partB map with invariants: (partB ! key) + (instanReserves ! key) >= (Coin 0) Note in Era before Alonzo, Negative transfers are not allowed.

makeDRepPred ∷ Term era (DRep (EraCrypto era)) → Term era (Credential 'DRepRole (EraCrypto era)) → Pred era

A user defined Predicate that Binds drep to a random DRep The parameter vote should be existentially bound in the surrounding context (inside a Choose Target perhaps)

minusCoinDeltaCoin ∷ Coin → DeltaCoin → DeltaCoin

availableForDistrC ∷ DeltaCoin → MIRPot → AccountState → InstantaneousRewards c → DeltaCoin

txCertMir ∷ Target era (MIRPot → Map (Credential 'Staking (EraCrypto era)) DeltaCoin → any → ShelleyTxCert era)

certsPreds ∷ ∀ era. Reflect era ⇒ Proof era → [Pred era]

certsStage ∷ Reflect era ⇒ Proof era → Subst era → Gen (Subst era)

main ∷ Int → IO ()

sMaybeL ∷ Lens' (StrictMaybe a) (Maybe a)

maybeSL ∷ Lens' (Maybe a) (StrictMaybe a)

poolMetaL ∷ Lens' (PoolParams era) (StrictMaybe PoolMetadata)

poolMetadata ∷ Proof era → Term era (Maybe PoolMetadata)