Test.Cardano.Ledger.Constrained.Preds.LedgerState

Documentation

prevGovActionIdsGenPreds ∷ Reflect era ⇒ Proof era → [Pred era]

prevGovActionIdsCheckPreds ∷ Proof era → [Pred era]

enactStateGenPreds ∷ Reflect era ⇒ Proof era → [Pred era]

enactStateCheckPreds ∷ Proof era → [Pred era]

ledgerStatePreds ∷ ∀ era. Reflect era ⇒ UnivSize → Proof era → [Pred era]

ledgerStateStage ∷ Reflect era ⇒ UnivSize → Proof era → Subst era → Gen (Subst era)

demo ∷ Reflect era ⇒ Proof era → ReplMode → IO ()

demoTest ∷ TestTree

main ∷ IO ()

genTree ∷ Ord a ⇒ [a] → Gen [(Maybe a, a)]

Generate the (parent,child) pairs in a Tree. Be sure the list of a are unique, because each node should appear in the Tree only once. The first one will be the root of the tree and have parent Nothing.

useTriples ∷ [(Maybe (GovActionId (EraCrypto era)), GovActionId (EraCrypto era))] → [GovAction era] → [GovActionState era] → [GovActionState era]

Tie together GovActionState and GovAction using the (parent,child) links that describe the shape of the Tree

govStatePreds ∷ ∀ era. (ConwayEraPParams era, Reflect era) ⇒ Proof era → [Pred era]

Pred era

that generate a valid (Map GovActionId GovActionState)

toProposalMap ∷ ∀ era. [GovActionState era] → Map (GovActionId (EraCrypto era)) (GovActionState era)

demoGov ∷ (ConwayEraPParams era, Reflect era) ⇒ Proof era → ReplMode → IO ()

mainGov ∷ IO ()

setActionId ∷ GovAction era → Maybe (GovActionId (EraCrypto era)) → GovAction era

actionIdL ∷ Lens' (GovAction era) (Maybe (GovActionId (EraCrypto era)))

children ∷ GovActionId c → [(Maybe (GovActionId c), GovActionId c)] → Set (GovActionId c)

genGovActionStates ∷ ∀ era. Era era ⇒ Proof era → Set (GovActionId (EraCrypto era)) → Gen (Map (GovActionId (EraCrypto era)) (GovActionState era))

genGovAction ∷ ∀ era. Era era ⇒ Proof era → GovActionPurpose → Maybe (GovActionId (EraCrypto era)) → Gen (GovAction era)

mapOMap ∷ HasOKey k v ⇒ (v → v) → OMap k v → OMap k v

updateProposals ∷ (GovAction era → GovAction era) → Proposals era → Proposals era

updateGovAction ∷ (PParamsUpdate era → PParamsUpdate era) → GovAction era → GovAction era

reasonable ∷ ConwayEraPParams era ⇒ Proposals era → Proposals era

What is reasonable for a Tx generated in Conway by drepCertTx Well parameters that affect the Fee, the Tx Size, TxOuts, and DRep Deposits