Test.Cardano.Ledger.Constrained.Preds.Universes

Documentation

variedCoin ∷ Gen Coin

noZeroCoin ∷ Gen Coin

genAddrPair ∷ Network → Gen (BootstrapAddress c, SigningKey)

Generate a pair, A Byron address, and the key that can sign it.

genByronUniv ∷ Crypto c ⇒ Network → Gen (Map (KeyHash 'Payment c) (Addr c, SigningKey))

Generate a Map, that maps the Hash of a Byron address to a pair of the actual Byron address and the key that can sign it.

bootWitness ∷ (Crypto c, DSIGN c ~ Ed25519DSIGN) ⇒ Hash c EraIndependentTxBody → [BootstrapAddress c] → Map (KeyHash 'Payment c) (Addr c, SigningKey) → Set (BootstrapWitness c)

Given a list of Byron addresses, compute BootStrap witnesses of all of those addresses Can only be used with StandardCrypto

genDatums ∷ Era era ⇒ Int → Map (DataHash (EraCrypto era)) (Data era) → Gen [Datum era]

The universe of non-empty Datums. i.e. There are no NoDatum Datums in this list

genDatum ∷ Era era ⇒ Map (DataHash (EraCrypto era)) (Data era) → Gen (Datum era)

Only generate non-empty Datums. I.e. There are no NoDatum Datums generated.

genValueF ∷ Proof era → Coin → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (Value era)

genTxOut ∷ Reflect era ⇒ Proof era → Coin → Set (Addr (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (DataHash (EraCrypto era)) (Data era) → Gen (TxOut era)

needsDatum ∷ EraScript era ⇒ Credential 'Payment (EraCrypto era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Bool

genTxOuts ∷ Reflect era ⇒ Proof era → Set (Addr (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (DataHash (EraCrypto era)) (Data era) → Gen [TxOutF era]

assets ∷ Set AssetName

genMultiAssetTriple ∷ Map (ScriptHash (EraCrypto era)) (ScriptF era) → Set AssetName → Gen Integer → Gen (PolicyID (EraCrypto era), AssetName, Integer)

multiAsset ∷ Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (MultiAsset (EraCrypto era))

pick1 ∷ [String] → Set t → Gen t

makeHashScriptMap ∷ Reflect era ⇒ Proof era → Int → Tag → Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era)) → ValidityInterval → Gen (Map (ScriptHash (EraCrypto era)) (ScriptF era))

genDataWits ∷ Era era ⇒ Proof era → Int → Gen (Map (DataHash (EraCrypto era)) (Data era))

genAddrWith ∷ Proof era → Network → Set (Credential 'Payment (EraCrypto era)) → Set Ptr → Set (Credential 'Staking (EraCrypto era)) → Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey) → Gen (Addr (EraCrypto era))

genPtr ∷ SlotNo → Gen Ptr

genStakeRefWith ∷ Set Ptr → Set (Credential 'Staking c) → Gen (StakeReference c)

noScripts ∷ Proof era → Addr (EraCrypto era) → Bool

txOutT ∷ Reflect era ⇒ Proof era → Addr (EraCrypto era) → Coin → TxOutF era

colTxOutT ∷ EraTxOut era ⇒ Proof era → Set (Addr (EraCrypto era)) → Gen (TxOutF era)

The collateral consists only of VKey addresses and the collateral outputs in the UTxO do not contain any non-ADA part

colTxOutSetT ∷ EraTxOut era ⇒ Proof era → Set (Addr (EraCrypto era)) → Gen (Set (TxOutF era))

The collateral consists only of VKey addresses and the collateral outputs in the UTxO do not contain any non-ADA part

scriptHashObjT ∷ Term era (ScriptHash (EraCrypto era)) → Target era (Credential k (EraCrypto era))

keyHashObjT ∷ Term era (KeyHash 'Witness (EraCrypto era)) → Target era (Credential k (EraCrypto era))

makeValidityT ∷ Term era SlotNo → Term era SlotNo → Term era SlotNo → Target era ValidityInterval

ptrUnivT ∷ Term era SlotNo → Target era (Gen (Set Ptr))

addrUnivT ∷ Proof era → Term era Network → Term era (Set (Credential 'Payment (EraCrypto era))) → Term era (Set Ptr) → Term era (Set (Credential 'Staking (EraCrypto era))) → Term era (Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey)) → Target era (Gen (Set (Addr (EraCrypto era))))

makeHashScriptMapT ∷ Proof era → Int → Tag → Term era (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era))) → Term era ValidityInterval → Target era (Gen (Map (ScriptHash (EraCrypto era)) (ScriptF era)))

cast ∷ ∀ c k. Set (KeyHash 'Witness c) → Set (KeyHash k c)

castCred ∷ Set (KeyHash 'Witness c) → Set (Credential 'ColdCommitteeRole c)

universePreds ∷ Reflect era ⇒ Proof era → [Pred era]

stakeToVote ∷ Credential 'Staking c → Credential 'DRepRole c

solveUniv ∷ Reflect era ⇒ Proof era → Gen (Subst era)

universeStage ∷ Reflect era ⇒ Proof era → Subst era → Gen (Subst era)

main ∷ IO ()