Test.Cardano.Ledger.Constrained.Scripts

Documentation

type KeyMap era = Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era))

genKeyMap ∷ Era era ⇒ Proof era → Gen (KeyMap era)

genMultiSig ∷ ∀ era. Era era ⇒ KeyMap era → Proof era → Gen (MultiSig era)

genTimelock ∷ ∀ era. Era era ⇒ KeyMap era → ValidityInterval → Proof era → Gen (Timelock era)

genPlutusScript ∷ ∀ era. PlutusPurposeTag → Proof era → Gen (Bool, Script era)

genCoreScript ∷ ∀ era. Proof era → PlutusPurposeTag → KeyMap era → ValidityInterval → Gen (Script era)

allPlutusScripts ∷ Reflect era ⇒ Proof era → Map (ScriptHash (EraCrypto era)) (IsValid, Script era)

For any given Era, there are only a finite number of Plutus scripts. This function computes all of them. There will be two failing scripts One for the Spend Tag, and another for all other Tags (Mint, Cert, Rewrd). The non-failing Spend scripts have varying number of arguments (0, 1, 2, 3) The non-failing (Mint Cert Rewrd) scripts are identical. Any Plutus script generated by genCoreScript will be in this Map.

spendPlutusScripts ∷ Reflect era ⇒ Proof era → Map (ScriptHash (EraCrypto era)) (IsValid, Script era)

There are only 5 plutus scripts that can be used in a Spend context

plutusByTag ∷ Proof era → PlutusPurposeTag → [(IsValid, Script era)]

sufficientMultiSig ∷ Era era ⇒ MultiSig era → Set (KeyHash 'Witness (EraCrypto era))

smallest ∷ [Set x] → Set x

sufficientTimelock ∷ Era era ⇒ Timelock era → Set (KeyHash 'Witness (EraCrypto era))

Return sufficient KeyHash to make the Timelock succeed. Note that some Timelock scripts need no KeyHashes to succeed (RequireTimeExpire, RequireTimeStart)

sufficientScript ∷ Proof era → Script era → Set (KeyHash 'Witness (EraCrypto era))