Test.Cardano.Ledger.Constrained.Size

Documentation

data Size

Constructors

SzNever [String]

SzAny

SzLeast Int

SzMost Int

SzRng Int Int
Size is in the range from i to j inclusive: SzRng i j = [i .. j]. Invariant i <= j

Bundled Patterns

pattern SzExact ∷ Int → Size

seps ∷ [String] → String

sepsP ∷ [String] → String

sepn ∷ [String] → String

runSize ∷ Int → Size → Bool

atLeastDelta ∷ Int

Used in tests so things don't get too large

atMostAny ∷ Int

Used in tests so things don't get too large If we can't find an era using things of size 10 using things of size 100, isn't going to help.

genFromSize ∷ Size → Gen Int

Use to generate real sizes, where there are no negative numbers, and the smallest possible size is 0. Use this only where you know it is NOT SzNever

genFromIntRange ∷ Size → Gen Int

Similar to genFromSize, but allows negative numbers (unlike size where the smallest Int is 0)

genFromNonNegIntRange ∷ Size → Gen Int

negateSize ∷ Size → Size

mergeSize ∷ Size → Size → Size