Test.Cardano.Ledger.Constrained.SpecClass

Description

Code which defines a class that encapsulates what it means to be a Spec Eventually this will repace the most of the tests in Spec.hs

Documentation

class Arbitrary t ⇒ HasRep t where

Methods

hasRep ∷ Rep era t

class (Monoid spec, LiftT spec) ⇒ Specification spec t | spec → t where

Associated Types

type Count spec ∷ Type

type Generators spec ∷ Type

type Reps spec ∷ Type

type Lenses spec ∷ Type

Methods

runS ∷ t → spec → Bool

genS ∷ [String] → Count spec → Generators spec → Reps spec → Lenses spec → Gen spec

sizeForS ∷ spec → Size

genFromS ∷ [String] → Int → Generators spec → spec → Gen t

testSound ∷ ∀ spec t. Specification spec t ⇒ Lenses spec → Reps spec → Count spec → Generators spec → Gen Bool

main ∷ IO ()