Test.Cardano.Ledger.Constrained.Stage

Documentation

data Stage era

Group together some Preds and OrderInfo about how to decide the order in which to solve the variables appearing in the Preds

Constructors

Stage OrderInfo [Pred era]

type Pipeline era = [Stage era]

ledgerPipeline ∷ Reflect era ⇒ UnivSize → Proof era → Pipeline era

A pipeline for specifying the LederState

stageToGraph ∷ Era era ⇒ Int → Stage era → HashSet (Name era) → Gen (Int, DependGraph era)

Translate a Stage into a DependGraph, given the set of variables that have aready been solved for.

mergePipeline ∷ Era era ⇒ Int → Pipeline era → HashSet (Name era) → DependGraph era → Gen (Int, DependGraph era)

Merge a Pipeline into an existing DependGraph, given the set of variables that have aready been solved for, to get a larger DependGraph

solvePipeline ∷ Reflect era ⇒ Pipeline era → Gen (Env era, Subst era, DependGraph era)

Solve a Pipeline to get a Env, Subst, and a DependGraph