Test.Cardano.Ledger.Constrained.Trace.DrepCertTx

Description

Generate a Simple Tx with 1 inout, 1 output, and 1 DRep related Cert

Documentation

fixOutput ∷ EraTxOut era ⇒ Coin → [TxBodyField era] → TraceM era [TxBodyField era]

Fix the first Outputs field in a [TxBodyField], by applying the delta Coin to the first Output in that Outputs field This is used to compensate for certificates which make deposits, and this Coin must come from somewhere, so it is added (subtracted if the Coin is negative) from the first TxOut. In rare occurrences, it is possible that delta might be too negative to subtract from the input in that case discard the trace. Note that in traces of length 150. this happens less than 1% of the time.

drepCert ∷ Reflect era ⇒ Proof era → TraceM era (Coin, [TxBodyField era])

Compute a valid Cert, and the change in the stored Deposits from that Cert.

drepCertTx ∷ Reflect era ⇒ Coin → Proof era → TraceM era (TxF era)

applyDRepCertActions ∷ Reflect era ⇒ Proof era → Tx era → TraceM era ()

drepCertTxForTrace ∷ Reflect era ⇒ Coin → Proof era → TraceM era (Tx era)

Generate a Tx that can be made into a Trace, because it applies the necessary Actions to update the Env. The Env must contain the Vars that are updated by applyDRepCertActions . It is best to intialize the whole LedgerState to do this.

drepTree ∷ TestTree

main ∷ IO ()

getpp ∷ EraGov era ⇒ NewEpochState era → PParams era

getProposals ∷ ConwayEraGov era ⇒ NewEpochState era → Map (GovActionId (EraCrypto era)) (GovActionState era)

allValidSignals ∷ Reflect era ⇒ Proof era → MockChainState era → MockBlock era → MockChainState era → Property

pulserWorks ∷ ConwayEraGov era ⇒ MockChainState era → MockChainState era → Property

bruteForceDRepDistr ∷ ∀ era. ConwayEraGov era ⇒ NewEpochState era → Map (DRep (EraCrypto era)) (CompactForm Coin)

extractPulsingDRepDistr ∷ ConwayEraGov era ⇒ NewEpochState era → Map (DRep (EraCrypto era)) (CompactForm Coin)

showCerts ∷ Proof era → [TxCert era] → String

certsOf ∷ EraTx era ⇒ Tx era → [TxCert era]

hashTx ∷ EraTx era ⇒ Proof era → Tx era → Hash (HASH (EraCrypto era)) EraIndependentTxBody

showMap ∷ (Show k, Show v) ⇒ String → Map k v → String

traceMap ∷ (Show k, Show v) ⇒ String → Map k v → a → a

showPotObl ∷ ConwayEraGov era ⇒ NewEpochState era → String