Test.Cardano.Ledger.Constrained.Trace.Tests

Documentation

genAndRunSimpleTx ∷ TraceM (ConwayEra StandardCrypto) Property

Construct and run one simpleTx, and run it through applySTS Check that the computed LedgerState is the same as the expected LedgerState Computed by using inputsAction , outputsAction , and feesAction

ledgerStateEqProp ∷ (Signal (EraRule "LEDGER" era) ~ Tx era, Reflect era, Show (State (EraRule "LEDGER" era)), Show (PredicateFailure (EraRule "LEDGER" era)), Eq (State (EraRule "LEDGER" era))) ⇒ Proof era → Env era → State (EraRule "LEDGER" era) → Environment (EraRule "LEDGER" era) → State (EraRule "LEDGER" era) → Signal (EraRule "LEDGER" era) → Property

Create a Property by testing that applying the STS LEDGER rule, succeeds and returns the expected LedgerState. If it fails, print out the failures and drop into the Repl, so that users can explore the inputs.

main1 ∷ IO ()

runOne ∷ (Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Signal (EraRule "LEDGER" era) ~ Tx era, Show (PredicateFailure (EraRule "LEDGER" era)), Reflect era) ⇒ Proof era → TxIx → TraceStep era (Signal (EraRule "LEDGER" era)) → Typed Property

oneTx ∷ Reflect era ⇒ Proof era → Int → TraceM era (Tx era)

testTrace ∷ (Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Signal (EraRule "LEDGER" era) ~ Tx era, Reflect era, Show (PredicateFailure (EraRule "LEDGER" era))) ⇒ Proof era → Int → TraceM era Property

conwayTrace ∷ TestTree

conwayTxwithDRepCertsTraceTests ∷ TestTree

main ∷ IO ()