Test.Cardano.Ledger.Constrained.TypeRep

Documentation

data Rep era t where

Constructors

RationalR ∷ Rep era Rational

CoinR ∷ Rep era Coin

EpochR ∷ Rep era EpochNo

EpochIntervalR ∷ Rep era EpochInterval

(:->) ∷ Rep era a → Rep era b → Rep era (a → b) infixr 0

MapR ∷ Ord a ⇒ Rep era a → Rep era b → Rep era (Map a b)

SetR ∷ Ord a ⇒ Rep era a → Rep era (Set a)

ListR ∷ Rep era a → Rep era [a]

AddrR ∷ Era era ⇒ Rep era (Addr (EraCrypto era))

CredR ∷ Era era ⇒ Rep era (Credential 'Staking (EraCrypto era))

VCredR ∷ Era era ⇒ Rep era (Credential 'DRepRole (EraCrypto era))

PoolHashR ∷ Era era ⇒ Rep era (KeyHash 'StakePool (EraCrypto era))

WitHashR ∷ Era era ⇒ Rep era (KeyHash 'Witness (EraCrypto era))

GenHashR ∷ Era era ⇒ Rep era (KeyHash 'Genesis (EraCrypto era))

GenDelegHashR ∷ Era era ⇒ Rep era (KeyHash 'GenesisDelegate (EraCrypto era))

VHashR ∷ Era era ⇒ Rep era (KeyHash 'DRepRole (EraCrypto era))

CommColdCredR ∷ Era era ⇒ Rep era (Credential 'ColdCommitteeRole (EraCrypto era))

CommHotCredR ∷ Era era ⇒ Rep era (Credential 'HotCommitteeRole (EraCrypto era))

PoolParamsR ∷ Era era ⇒ Rep era (PoolParams (EraCrypto era))

NewEpochStateR ∷ Era era ⇒ Rep era (NewEpochState era)

IntR ∷ Rep era Int

FloatR ∷ Rep era Float

NaturalR ∷ Rep era Natural

Word64R ∷ Rep era Word64

TxInR ∷ Era era ⇒ Rep era (TxIn (EraCrypto era))

TxIdR ∷ Era era ⇒ Rep era (TxId (EraCrypto era))

CharR ∷ Rep era Char

UnitR ∷ Rep era ()

PairR ∷ Rep era a → Rep era b → Rep era (a, b)

ProtVerR ∷ Era era ⇒ Proof era → Rep era ProtVer

ValueR ∷ Era era ⇒ Proof era → Rep era (ValueF era)

UTxOR ∷ Era era ⇒ Proof era → Rep era (UTxO era)

TxOutR ∷ Era era ⇒ Proof era → Rep era (TxOutF era)

PParamsR ∷ Era era ⇒ Proof era → Rep era (PParamsF era)

FuturePParamsR ∷ Era era ⇒ Proof era → Rep era (FuturePParams era)

PParamsUpdateR ∷ Era era ⇒ Proof era → Rep era (PParamsUpdateF era)

DeltaCoinR ∷ Rep era DeltaCoin

GenDelegPairR ∷ Era era ⇒ Rep era (GenDelegPair (EraCrypto era))

FutureGenDelegR ∷ Era era ⇒ Rep era (FutureGenDeleg (EraCrypto era))

PPUPStateR ∷ Era era ⇒ Proof era → Rep era (ShelleyGovState era)

PtrR ∷ Rep era Ptr

IPoolStakeR ∷ Era era ⇒ Rep era (IndividualPoolStake (EraCrypto era))

SnapShotsR ∷ Era era ⇒ Rep era (SnapShots (EraCrypto era))

RewardR ∷ Era era ⇒ Rep era (Reward (EraCrypto era))

MaybeR ∷ Rep era t → Rep era (Maybe t)

SlotNoR ∷ Rep era SlotNo

SizeR ∷ Rep era Size

MultiAssetR ∷ Era era ⇒ Rep era (MultiAsset (EraCrypto era))

PolicyIDR ∷ Era era ⇒ Rep era (PolicyID (EraCrypto era))

WitnessesFieldR ∷ Era era ⇒ Proof era → Rep era (WitnessesField era)

AssetNameR ∷ Rep era AssetName

TxCertR ∷ Era era ⇒ Proof era → Rep era (TxCertF era)

RewardAccountR ∷ Era era ⇒ Rep era (RewardAccount (EraCrypto era))

ValidityIntervalR ∷ Era era ⇒ Rep era ValidityInterval

KeyPairR ∷ Era era ⇒ Rep era (KeyPair 'Witness (EraCrypto era))

GenR ∷ Rep era x → Rep era (Gen x)

ScriptR ∷ Era era ⇒ Proof era → Rep era (ScriptF era)

ScriptHashR ∷ Era era ⇒ Rep era (ScriptHash (EraCrypto era))

NetworkR ∷ Rep era Network

RdmrPtrR ∷ Era era ⇒ Proof era → Rep era (PlutusPointerF era)

DataR ∷ Era era ⇒ Rep era (Data era)

DatumR ∷ Era era ⇒ Rep era (Datum era)

ExUnitsR ∷ Rep era ExUnits

DataHashR ∷ Era era ⇒ Rep era (DataHash (EraCrypto era))

PCredR ∷ Era era ⇒ Rep era (Credential 'Payment (EraCrypto era))

ShelleyTxCertR ∷ Era era ⇒ Rep era (ShelleyTxCert era)

ConwayTxCertR ∷ Era era ⇒ Rep era (ConwayTxCert era)

MIRPotR ∷ Rep era MIRPot

IsValidR ∷ Rep era IsValid

IntegerR ∷ Rep era Integer

ScriptsNeededR ∷ Era era ⇒ Proof era → Rep era (ScriptsNeededF era)

ScriptPurposeR ∷ Era era ⇒ Proof era → Rep era (PlutusPurposeF era)

TxBodyR ∷ Era era ⇒ Proof era → Rep era (TxBodyF era)

BootstrapWitnessR ∷ Era era ⇒ Rep era (BootstrapWitness (EraCrypto era))

SigningKeyR ∷ Rep era SigningKey

TxWitsR ∷ Era era ⇒ Proof era → Rep era (TxWitsF era)

PayHashR ∷ Era era ⇒ Rep era (KeyHash 'Payment (EraCrypto era))

TxR ∷ Era era ⇒ Proof era → Rep era (TxF era)

ScriptIntegrityHashR ∷ Era era ⇒ Rep era (SafeHash (EraCrypto era) EraIndependentScriptIntegrity)

AuxiliaryDataHashR ∷ Era era ⇒ Rep era (AuxiliaryDataHash (EraCrypto era))

GovActionR ∷ Era era ⇒ Rep era (GovAction era)

WitVKeyR ∷ Era era ⇒ Proof era → Rep era (WitVKey 'Witness (EraCrypto era))

TxAuxDataR ∷ Era era ⇒ Proof era → Rep era (TxAuxDataF era)

LanguageR ∷ Rep era Language

LedgerStateR ∷ Era era ⇒ Proof era → Rep era (LedgerState era)

StakeHashR ∷ Era era ⇒ Rep era (KeyHash 'Staking (EraCrypto era))

BoolR ∷ Rep era Bool

DRepR ∷ Era era ⇒ Rep era (DRep (EraCrypto era))

PoolMetadataR ∷ Era era ⇒ Proof era → Rep era PoolMetadata

DRepStateR ∷ Era era ⇒ Rep era (DRepState (EraCrypto era))

DStateR ∷ Era era ⇒ Rep era (DState era)

GovActionIdR ∷ Era era ⇒ Rep era (GovActionId (EraCrypto era))

GovActionIxR ∷ Rep era GovActionIx

GovActionStateR ∷ Era era ⇒ Rep era (GovActionState era)

ProposalsR ∷ Era era ⇒ Proof era → Rep era (Proposals era)

UnitIntervalR ∷ Rep era UnitInterval

CommitteeR ∷ Era era ⇒ Rep era (Committee era)

ConstitutionR ∷ Era era ⇒ Rep era (Constitution era)

PrevGovActionIdsR ∷ Era era ⇒ Rep era (GovRelation StrictMaybe era)

PrevPParamUpdateR ∷ Era era ⇒ Rep era (GovPurposeId 'PParamUpdatePurpose era)

PrevHardForkR ∷ Era era ⇒ Rep era (GovPurposeId 'HardForkPurpose era)

PrevCommitteeR ∷ Era era ⇒ Rep era (GovPurposeId 'CommitteePurpose era)

PrevConstitutionR ∷ Era era ⇒ Rep era (GovPurposeId 'ConstitutionPurpose era)

RatifyStateR ∷ Reflect era ⇒ Rep era (RatifyState era)

NumDormantEpochsR ∷ Era era ⇒ Rep era EpochNo

DRepHashR ∷ Era era ⇒ Rep era (KeyHash 'DRepRole (EraCrypto era))

AnchorR ∷ Era era ⇒ Rep era (Anchor (EraCrypto era))

CommitteeStateR ∷ Era era ⇒ Rep era (CommitteeState era)

CommitteeAuthorizationR ∷ Era era ⇒ Rep era (CommitteeAuthorization (EraCrypto era))

VStateR ∷ Era era ⇒ Rep era (VState era)

EnactStateR ∷ Reflect era ⇒ Rep era (EnactState era)

DRepPulserR ∷ (RunConwayRatify era, Reflect era) ⇒ Rep era (DRepPulser era Identity (RatifyState era))

DelegateeR ∷ Era era ⇒ Rep era (Delegatee (EraCrypto era))

VoteR ∷ Rep era Vote

data (a ∷ k) :~: (b ∷ k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl ∷ ∀ {k} (a ∷ k). a :~: a

class Singleton (t ∷ k → Type) where

Given (Singleton T), a value of type (T i) has exactly 0 or 1 inhabitants, so we can compare the structure of the type to get proofs that the indexes (i and j) are the same type, using testEql, at runtime. cmpIndex, it is like compare except we can have two different indexes (a and b).

Methods

testEql ∷ ∀ (i ∷ k) (j ∷ k). t i → t j → Maybe (i :~: j)

cmpIndex ∷ ∀ (a ∷ k) (b ∷ k). t a → t b → Ordering

type Eql (x ∷ k) (y ∷ k) = x :~: y

Type synonym, so we can use ( :~: ) without TypeOperators

typeRepOf ∷ Rep era t → TypeRep

synopsis ∷ ∀ e t. Rep e t → t → String

genSizedRep ∷ ∀ era t. Int → Rep era t → Gen t

genRep ∷ ∀ era b. Rep era b → Gen b

shrinkRep ∷ Rep era t → t → [t]

data TxOutF era where

Constructors

TxOutF ∷ Proof era → TxOut era → TxOutF era

unTxOut ∷ TxOutF era → TxOut era

data ValueF era where

Constructors

ValueF ∷ Proof era → Value era → ValueF era

unValue ∷ ValueF era → Value era

data PParamsF era where

Constructors

PParamsF ∷ Proof era → PParams era → PParamsF era

unPParams ∷ PParamsF era → PParams era

data PParamsUpdateF era where

Constructors

PParamsUpdateF ∷ Proof era → PParamsUpdate era → PParamsUpdateF era

unPParamsUpdate ∷ PParamsUpdateF era → PParamsUpdate era

liftUTxO ∷ Map (TxIn (EraCrypto era)) (TxOutF era) → UTxO era

data Proof era where

Proof of a valid (predefined) era

Constructors

Shelley ∷ Proof (ShelleyEra StandardCrypto)

Mary ∷ Proof (MaryEra StandardCrypto)

Allegra ∷ Proof (AllegraEra StandardCrypto)

Alonzo ∷ Proof (AlonzoEra StandardCrypto)

Babbage ∷ Proof (BabbageEra StandardCrypto)

Conway ∷ Proof (ConwayEra StandardCrypto)

stringR ∷ Rep era String

hasOrd ∷ Rep era t → s t → Typed (HasConstraint Ord (s t))

hasEq ∷ Rep era t → s t → Typed (HasConstraint Eq (s t))

format ∷ Rep era t → t → String

genSigningKey ∷ Gen SigningKey

Turn a random bytestring into a SigningKey