Test.Cardano.Ledger.Constrained.Vars

Documentation

field ∷ Rep era s → Term era t → AnyF era s

Used in Component constraints to turn a Var Term into a component (AnyF era s) E.g. (Component foo [ field fooRep fooPart1, field fooRep fooPart2]) Where fooPart1 :: Term era a, and fooPart2 :: Term era b And fooPart1 has an (Access foo a) And fooPart2 has an (Access foo b)

getName ∷ Term era t → Name era

type NELens era t = Lens' (NewEpochState era) t

currentEpoch ∷ Term era EpochNo

prevBlocksMade ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

currBlocksMade ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

poolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

mockPoolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Rational)

For tests only, Like PoolDistr but has a Rational (rather than a IndividualPoolStake).

poolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

unPoolDistrL ∷ Lens' (PoolDistr c) (Map (KeyHash 'StakePool c) (IndividualPoolStake c))

rewards ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

rewardsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

delegations ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

delegationsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

stakeDeposits ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

stakeDepositsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

ptrs ∷ Term era (Map Ptr (Credential 'Staking (EraCrypto era)))

ptrsL ∷ NELens era (Map Ptr (Credential 'Staking (EraCrypto era)))

futureGenDelegs ∷ Term era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

futureGenDelegsL ∷ NELens era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegs ∷ Term era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegsL ∷ NELens era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

instanReserves ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanReservesL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanReservesSum ∷ Term era Coin

instanTreasury ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasuryL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasurySum ∷ Term era Coin

deltaReserves ∷ Term era DeltaCoin

deltaReservesNEL ∷ NELens era DeltaCoin

deltaTreasury ∷ Term era DeltaCoin

deltaTreasuryNEL ∷ NELens era DeltaCoin

regPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

regPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

retiring ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

retiringL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

poolDeposits ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

poolDepositsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

dreps ∷ Term era (Set (Credential 'DRepRole (EraCrypto era)))

drepsL ∷ NELens era (Set (Credential 'DRepRole (EraCrypto era)))

ccHotKeys ∷ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))))

ccHotKeysL ∷ NELens era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))))

utxo ∷ Proof era → Term era (Map (TxIn (EraCrypto era)) (TxOutF era))

utxoL ∷ Proof era → NELens era (Map (TxIn (EraCrypto era)) (TxOutF era))

unUtxoL ∷ Proof era → Lens' (UTxO era) (Map (TxIn (EraCrypto era)) (TxOutF era))

deposits ∷ Term era Coin

depositsL ∷ NELens era Coin

fees ∷ Term era Coin

feesL ∷ NELens era Coin

ppup ∷ Proof era → Term era (ShelleyPPUPState era)

ppupsL ∷ Proof era → NELens era (ShelleyPPUPState era)

proposalsT ∷ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

futureProposalsT ∷ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

ppupStateT ∷ Proof era → Target era (ShelleyPPUPState era)

governanceStateT ∷ ∀ era. Proof era → Target era (GovernanceState era)

individualPoolStakeL ∷ Lens' (IndividualPoolStake c) Rational

isPtrMapT ∷ Term era (Map Ptr Coin)

ptrMapL ∷ Lens' (NewEpochState era) (Map Ptr Coin)

isCredMapT ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

credMapL ∷ Lens' (NewEpochState era) (Map (Credential 'Staking (EraCrypto era)) Coin)

treasury ∷ Term era Coin

treasuryL ∷ NELens era Coin

reserves ∷ Term era Coin

reservesL ∷ NELens era Coin

mirAvailTreasury ∷ Term era Coin

The Coin availabe for a MIR transfer to/from the Treasury Computed from treasury + deltaTreasury - sum(instanTreasury)

mirAvailReserves ∷ Term era Coin

The Coin availabe for a MIR transfer to/from the Reserves Computed from reserves + deltaReserves - sum(instanReserves)

snapshots ∷ Term era (SnapShots (EraCrypto era))

snapshotsL ∷ NELens era (SnapShots (EraCrypto era))

ppFL ∷ Proof era → Lens' (PParams era) (PParamsF era)

Lens' from the Core PParams to the Model PParamsF which embeds a (Proof era)

prevpparams ∷ Proof era → Term era (PParamsF era)

pparams ∷ Proof era → Term era (PParamsF era)

nmLikelihoodsT ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) [Float])

nmRewardPotT ∷ Term era Coin

stakeL ∷ Lens' (Stake c) (Map (Credential 'Staking c) Coin)

Helper lens that deals with the Stake newtype, and the shift from Map to VMap

vmapL ∷ Lens' (VMap VB VB k v) (Map k v)

Helper lens that deals with the shift from Map to VMap

markStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

markStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

markDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markSnapShotT ∷ Target era (SnapShot (EraCrypto era))

setStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

setStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

setDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setSnapShotT ∷ Target era (SnapShot (EraCrypto era))

goStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

goStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

goDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goSnapShotT ∷ Target era (SnapShot (EraCrypto era))

markPoolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

markPoolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

pooldistrHelpL ∷ Lens' (PoolDistr c) (Map (KeyHash 'StakePool c) (IndividualPoolStake c))

snapShotFee ∷ Term era Coin

snapShotsT ∷ Target era (SnapShots (EraCrypto era))

deltaT ∷ Term era (Maybe DeltaCoin)

deltaTL ∷ NELens era (Maybe DeltaCoin)

deltaR ∷ Term era (Maybe DeltaCoin)

deltaRL ∷ NELens era (Maybe DeltaCoin)

deltaF ∷ Term era (Maybe DeltaCoin)

deltaFL ∷ NELens era (Maybe DeltaCoin)

rewardSet ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

rewardSetL ∷ NELens era (Maybe (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))))

totalAda ∷ Term era Coin

utxoCoin ∷ Term era Coin

credsUniv ∷ Term era (Set (Credential 'Staking (EraCrypto era)))

The universe of Staking Credentials. A credential is either KeyHash of a ScriptHash Any Plutus scripts in this Universe are NOT Spending scripts, so they do not need a Redeemer

spendCredsUniv ∷ Term era (Set (Credential 'Payment (EraCrypto era)))

The universe of Staking Credentials. A credential is either KeyHash of a ScriptHash All Plutus scripts in this Universe are SPENDING scripts, so they will need a Redeemer Use this ONLY in the Pay-part of an Address (Do not use this in the Stake-part of an Address)

voteUniv ∷ Term era (Set (Credential 'DRepRole (EraCrypto era)))

The universe of Voting Credentials. A credential is either KeyHash of a ScriptHash

payUniv ∷ Term era (Set (Credential 'Payment (EraCrypto era)))

The universe of Payment Credentials. A credential is either KeyHash of a ScriptHash We only find payment credentials in the Payment part of an Addr.

spendscriptUniv ∷ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The universe of Scripts (and their hashes) useable in spending contexts That means if they are Plutus scripts then they will be passed an additional argument (the TxInfo context)

nonSpendScriptUniv ∷ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The universe of Scripts (and their hashes) useable in contexts other than Spending

allScriptUniv ∷ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The union of spendscriptUniv and nonSpendScriptUniv. All possible scripts in any context

dataUniv ∷ Era era ⇒ Term era (Map (DataHash (EraCrypto era)) (Data era))

The universe of Data (and their hashes)

poolHashUniv ∷ Term era (Set (KeyHash 'StakePool (EraCrypto era)))

The universe of StakePool key hashes. These hashes hash the cold key of the Pool operators.

stakeHashUniv ∷ Term era (Set (KeyHash 'Staking (EraCrypto era)))

The universe of StakePool key hashes. These hashes hash are hashes of the Owners of a PoolParam

genesisHashUniv ∷ Term era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

The universe of the Genesis key hashes and their signing and validating GenDelegPairs

voteCredUniv ∷ Term era (Set (Credential 'ColdCommitteeRole (EraCrypto era)))

txinUniv ∷ Term era (Set (TxIn (EraCrypto era)))

The universe of TxIns. Pairs of TxId: hashes of previously run transaction bodies, and TxIx: indexes of one of the bodies outputs.

txoutUniv ∷ Proof era → Term era (Set (TxOutF era))

The universe of TxOuts. It contains colTxoutUniv as a sublist and feeOutput as an element See also feeOutput which is defined by the universes, and is related.

colTxoutUniv ∷ Proof era → Term era (Set (TxOutF era))

The universe of TxOuts useable for collateral The collateral TxOuts consists only of VKey addresses and The collateral TxOuts do not contain any non-ADA part

feeTxOut ∷ Reflect era ⇒ Term era (TxOutF era)

A TxOut, guaranteed to have 1) no scripts in its Addr, and 2) It's Addr is in the addrUniv 3) bigCoin is stored in the Addr Value, and 4) the Addr Value has empty MutiAssets 5) be a member of the txoutUniv

feeTxIn ∷ Term era (TxIn (EraCrypto era))

A TxIn, guaranteed to have 1) be a member of the txinUniv

bigCoin ∷ Term era Coin

A Coin large enough to pay almost any fee. See also feeOutput which is related.

datumsUniv ∷ Era era ⇒ Term era [Datum era]

multiAssetUniv ∷ Era era ⇒ Term era [MultiAsset (EraCrypto era)]

keymapUniv ∷ Term era (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era)))

The universe of key hashes, and the signing and validating key pairs they represent.

currentSlot ∷ Term era SlotNo

endSlotDelta ∷ Term era SlotNo

beginSlotDelta ∷ Term era SlotNo

network ∷ Term era Network

From Globals

quorumConstant ∷ Word64

This not really a variable, But a constant that is set by the testGlobals we reflect this into a Term, so we can refer to it in the Preds.

quorum ∷ Term era Int

From Globals. Reflected here at type Int, This is set to quorumConstant in CertState. because is is used to compare the Size of things, which are computed as Int

addrUniv ∷ Term era (Set (Addr (EraCrypto era)))

ptrUniv ∷ Term era (Set Ptr)

plutusUniv ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (IsValid, ScriptF era))

spendPlutusUniv ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (IsValid, ScriptF era))

byronAddrUniv ∷ Term era (Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey))

The universe of all Byron addresses. In Eras, Babbage, Conway we avoid these Adresses, as they do not play well with Plutus Scripts.

newEpochStateConstr ∷ Proof era → EpochNo → Map (KeyHash 'StakePool (EraCrypto era)) Natural → Map (KeyHash 'StakePool (EraCrypto era)) Natural → EpochState era → Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)) → NewEpochState era

Abstract constuctor function for NewEpochState

newEpochStateT ∷ Proof era → Target era (NewEpochState era)

Target for NewEpochState

epochStateT ∷ Proof era → Target era (EpochState era)

Target for EpochState

accountStateT ∷ Target era AccountState

Target for AccountState

ledgerStateT ∷ Proof era → Target era (LedgerState era)

Target for LedgerState

ledgerState ∷ Reflect era ⇒ Term era (LedgerState era)

utxoStateT ∷ Proof era → Target era (UTxOState era)

Target for UTxOState

certstateT ∷ Target era (CertState era)

Target for CertState

vstateT ∷ Target era (VState era)

Target for VState

pstateT ∷ Target era (PState era)

Target for PState

dstateT ∷ Target era (DState era)

Target for DState

dstate ∷ Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)) → Map Ptr (Credential 'Staking (EraCrypto era)) → Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)) → Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → InstantaneousRewards (EraCrypto era) → DState era

Abstract construcor function for DState

instantaneousRewardsT ∷ Target era (InstantaneousRewards (EraCrypto era))

allvars ∷ String

A String that pretty prints the complete set of variables

withEraPParams ∷ ∀ era a. Proof era → (EraPParams era ⇒ a) → a

protVer ∷ Proof era → Term era ProtVer

ProtVer in pparams

prevProtVer ∷ Proof era → Term era ProtVer

ProtVer in prevpparams

minFeeA ∷ Proof era → Term era Coin

minFeeB ∷ Proof era → Term era Coin

maxBBSize ∷ Proof era → Term era Natural

Max Block Body Size

maxTxSize ∷ Proof era → Term era Natural

Max Tx Size

maxBHSize ∷ Proof era → Term era Natural

Max Block Header Size

poolDepAmt ∷ Proof era → Term era Coin

keyDepAmt ∷ Proof era → Term era Coin

maxTxExUnits ∷ AlonzoEraPParams era ⇒ Proof era → Term era ExUnits

collateralPercentage ∷ AlonzoEraPParams era ⇒ Proof era → Term era Natural

maxEpoch ∷ Proof era → Term era EpochNo

txbodyterm ∷ Reflect era ⇒ Term era (TxBodyF era)

inputs ∷ Term era (Set (TxIn (EraCrypto era)))

collateral ∷ Term era (Set (TxIn (EraCrypto era)))

refInputs ∷ Term era (Set (TxIn (EraCrypto era)))

outputs ∷ Proof era → Term era [TxOutF era]

collateralReturn ∷ Proof era → Term era (TxOutF era)

totalCol ∷ Term era Coin

The sum of all the collateral inputs. The Tx is constucted by SNothing or wrapping SJust around this value.

certs ∷ Reflect era ⇒ Term era [TxCertF era]

withdrawals ∷ ∀ era. Term era (Map (RewardAcnt (EraCrypto era)) Coin)

txfee ∷ Term era Coin

ttl ∷ Term era SlotNo

validityInterval ∷ Term era ValidityInterval

mint ∷ Term era (Map (ScriptHash (EraCrypto era)) (Map AssetName Integer))

reqSignerHashes ∷ Term era (Set (KeyHash 'Witness (EraCrypto era)))

networkID ∷ Term era (Maybe Network)

adHash ∷ Term era (Maybe (AuxiliaryDataHash (EraCrypto era)))

wppHash ∷ Term era (Maybe (SafeHash (EraCrypto era) EraIndependentScriptIntegrity))

liftMultiAsset ∷ Map (ScriptHash c) (Map AssetName Integer) → MultiAsset c

lift the model type of mint into a MultiAsset

scriptsNeeded ∷ Reflect era ⇒ Term era (ScriptsNeededF era)

smNeededL ∷ ScriptsNeeded era ~ ShelleyScriptsNeeded era ⇒ Lens' (ScriptsNeededF era) (Set (ScriptHash (EraCrypto era)))

acNeededL ∷ ScriptsNeeded era ~ AlonzoScriptsNeeded era ⇒ Lens' (ScriptsNeededF era) [(ScriptPurpose era, ScriptHash (EraCrypto era))]

extraCol ∷ Term era Coin

A Coin that needs to be added to the range of the colInputs in the UtxO that will make sure the collateral is large enough to pay the fees if needed

sumCol ∷ Term era Coin

The sum of all the collateral inputs, total colateral of the Tx is computed by adding (SJust _) to this value.

colRetAddr ∷ Term era (Addr (EraCrypto era))

colRetCoin ∷ Term era Coin

The Coin in the collateralReturn TxOut

owed ∷ Term era Coin

The amount that the collateral must cover if there is a two phase error. This is roughly the collateralPercentage * txfee . The calculation deals with rounding, but you don't need those details to understand what is going on.

txbody ∷ Reflect era ⇒ Term era (TxBodyF era)

txwits ∷ Reflect era ⇒ Term era (TxWitsF era)

txauxdata ∷ Reflect era ⇒ Term era (Maybe (TxAuxDataF era))

txisvalid ∷ Term era IsValid

valids ∷ Term era [IsValid]

txterm ∷ Reflect era ⇒ Term era (TxF era)

getRwdCredL ∷ Lens' (RewardAcnt c) (Credential 'Staking c)

txOutFL ∷ Lens' (TxOutF era) (TxOut era)

valueFL ∷ Reflect era ⇒ Lens' (Value era) (ValueF era)

lensVC ∷ Val t ⇒ Lens' t Coin

valueFCoinL ∷ (HasCallStack, Reflect era) ⇒ Lens' (ValueF era) Coin

outputCoinL ∷ (HasCallStack, Reflect era) ⇒ Lens' (TxOutF era) Coin

valCoinF ∷ (HasCallStack, Reflect era) ⇒ Field era (ValueF era) Coin

a Field from (ValueF era) to Coin

valCoin ∷ (HasCallStack, Reflect era) ⇒ Term era Coin

maryValueMultiAssetL ∷ Lens' (MaryValue c) (MultiAsset c)

valueFMultiAssetL ∷ Lens' (ValueF era) (MultiAsset (EraCrypto era))

valueFMultiAssetF ∷ Reflect era ⇒ Field era (ValueF era) (MultiAsset (EraCrypto era))

a Field from (ValueF era) to MultiAsset

valueFMultiAsset ∷ Reflect era ⇒ Term era (MultiAsset (EraCrypto era))

txoutAddressF ∷ Reflect era ⇒ Field era (TxOutF era) (Addr (EraCrypto era))

a Field from (TxOut era) to (Addr era)

txoutAddress ∷ Reflect era ⇒ Term era (Addr (EraCrypto era))

txoutCoinF ∷ (HasCallStack, Reflect era) ⇒ Field era (TxOutF era) Coin

a Field from (TxOutF era) to Coin

txoutCoin ∷ (HasCallStack, Reflect era) ⇒ Term era Coin

txoutAmountF ∷ Reflect era ⇒ Field era (TxOutF era) (ValueF era)

a Field from (TxOutF era) to (ValueF era)

txoutAmount ∷ Reflect era ⇒ Term era (ValueF era)

scriptWits ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

redeemers ∷ Reflect era ⇒ Term era (Map RdmrPtr (Data era, ExUnits))

bootWits ∷ ∀ era. Reflect era ⇒ Term era (Set (BootstrapWitness (EraCrypto era)))

dataWits ∷ Reflect era ⇒ Term era (Map (DataHash (EraCrypto era)) (Data era))

keyWits ∷ Reflect era ⇒ Term era (Set (WitVKey 'Witness (EraCrypto era)))

witsTarget ∷ Reflect era ⇒ Term era (Set (BootstrapWitness (EraCrypto era))) → Term era (Set (WitVKey 'Witness (EraCrypto era))) → Target era (TxWits era)

txTarget ∷ Reflect era ⇒ Term era (TxBodyF era) → Term era (Set (BootstrapWitness (EraCrypto era))) → Term era (Set (WitVKey 'Witness (EraCrypto era))) → Target era (TxF era)

txbodyTarget ∷ Reflect era ⇒ Term era Coin → Term era (Maybe (ScriptIntegrityHash (EraCrypto era))) → Term era Coin → Target era (TxBodyF era)

Need to build the TxBody with different terms that control the fee and wppHash so we parameterise this target over those two terms

allowMIRTransfer ∷ Proof era → Term era Bool