Test.Cardano.Ledger.Examples.STSTestUtils

Documentation

class EraTest era ⇒ EraModel era where

Minimal complete definition

applyTx, applyCert, always, never, collateralReturnTxBodyT, validTxOut

Methods

applyTx ∷ Int → SlotNo → Model era → Tx era → Model era

applyCert ∷ Model era → TxCert era → Model era

mkRedeemersFromTags ∷ [((PlutusPurposeTag, Word32), (Data era, ExUnits))] → Redeemers era

mkRedeemers ∷ [(PlutusPurpose AsIx era, (Data era, ExUnits))] → Redeemers era

newScriptIntegrityHash ∷ PParams era → [Language] → Redeemers era → TxDats era → StrictMaybe ScriptIntegrityHash

mkPlutusPurposePointer ∷ PlutusPurposeTag → Word32 → PlutusPurpose AsIx era

always ∷ Natural → Script era

never ∷ Natural → Script era

collateralReturnTxBodyT ∷ Lens' (TxBody era) (StrictMaybe (TxOut era))

validTxOut ∷ Map ScriptHash (Script era) → TxOut era → Bool

data PlutusPurposeTag

Constructors

Spending

Minting

Certifying

Rewarding

Voting

Proposing

initUTxO ∷ ∀ era. (AllegraEraScript era, AlonzoEraTxOut era, EraModel era) ⇒ UTxO era

mkGenesisTxIn ∷ HasCallStack ⇒ Integer → TxIn

mkTxDats ∷ Era era ⇒ Data era → TxDats era

mkSingleRedeemer ∷ ∀ era. AlonzoEraScript era ⇒ PlutusPurpose AsIx era → Data era → Redeemers era

someAddr ∷ Addr

someKeys ∷ KeyPair 'Payment

someScriptAddr ∷ ∀ era. EraScript era ⇒ Script era → Addr

testBBODY ∷ ∀ era. (HasCallStack, Eq (State (EraRule "LEDGERS" era)), ToExpr (PredicateFailure (EraRule "BBODY" era)), ToExpr (State (EraRule "LEDGERS" era)), BaseM (EraRule "BBODY" era) ~ ShelleyBase, Environment (EraRule "BBODY" era) ~ BbodyEnv era, State (EraRule "BBODY" era) ~ ShelleyBbodyState era, Signal (EraRule "BBODY" era) ~ Block BHeaderView era, STS (EraRule "BBODY" era)) ⇒ ShelleyBbodyState era → Block BHeaderView era → Either (NonEmpty (PredicateFailure (EraRule "BBODY" era))) (ShelleyBbodyState era) → PParams era → Assertion

runLEDGER ∷ ∀ era. (BaseM (EraRule "LEDGER" era) ~ ShelleyBase, STS (EraRule "LEDGER" era), Environment (EraRule "LEDGER" era) ~ LedgerEnv era, State (EraRule "LEDGER" era) ~ LedgerState era, Tx era ~ Signal (EraRule "LEDGER" era)) ⇒ LedgerState era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era))

testUTXOW ∷ ∀ era. (Reflect era, HasCallStack, ShelleyEraTest era, BaseM (EraRule "UTXOW" era) ~ ShelleyBase, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, STS (EraRule "UTXOW" era), Tx era ~ Signal (EraRule "UTXOW" era), State (EraRule "UTXOW" era) ~ UTxOState era, ToExpr (PredicateFailure (EraRule "UTXOW" era))) ⇒ UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use an equality test on the expected and computed [PredicateFailure]

testUTXOWsubset ∷ ∀ era. (Reflect era, BaseM (EraRule "UTXOW" era) ~ ShelleyBase, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Tx era ~ Signal (EraRule "UTXOW" era), STS (EraRule "UTXOW" era), ToExpr (PredicateFailure (EraRule "UTXOW" era)), ShelleyEraTest era) ⇒ UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use a subset test on the expected and computed [PredicateFailure]

testUTXOspecialCase ∷ ∀ era. (Reflect era, HasCallStack, BaseM (EraRule "UTXOW" era) ~ ShelleyBase, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Tx era ~ Signal (EraRule "UTXOW" era), STS (EraRule "UTXOW" era)) ⇒ UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use a test where any two (ValidationTagMismatch x y) failures match regardless of x and y

alwaysFailsHash ∷ ∀ era. (ShelleyEraScript era, EraModel era) ⇒ Natural → ScriptHash

alwaysSucceedsHash ∷ ∀ era. (ShelleyEraScript era, EraModel era) ⇒ Natural → ScriptHash

timelockScript ∷ AllegraEraScript era ⇒ SlotNo → Script era

timelockHash ∷ ∀ era. AllegraEraScript era ⇒ SlotNo → ScriptHash

timelockStakeCred ∷ ∀ era. AllegraEraScript era ⇒ StakeCredential

genericCont ∷ (Eq (t x), Eq y, ToExpr y, HasCallStack, ToExpr (t x)) ⇒ String → Either (t x) y → Either (t x) y → Assertion

A small example of what a continuation for runSTS might look like