Test.Cardano.Ledger.Generic.AggPropTests

Documentation

aggProp ∷ agg → (agg → Signal sts → agg) → (State sts → State sts → agg → prop) → Trace sts → prop

consistentUtxoSizeProp ∷ EraTx era ⇒ Proof era → Trace (MOCKCHAIN era) → Property

The aggregate sizes of (outputs - inputs) is consistent with the change in size of the UTxO. Be carefull to choose the correct outputs and inputs, depending on if the Tx validates.

aggUTxO ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), Reflect era) ⇒ Proof era → Gen Property

aggTests ∷ TestTree

forAllChainTrace ∷ (Testable prop, Reflect era) ⇒ Proof era → Int → (Trace (MOCKCHAIN era) → prop) → Property

depositInvariant ∷ SourceSignalTarget (MOCKCHAIN era) → Property

Check that the sum of Key Deposits and the Pool Depoits are equal to the utxosDeposits

rewardDepositDomainInvariant ∷ SourceSignalTarget (MOCKCHAIN era) → Property

itemPropToTraceProp ∷ (SourceSignalTarget (MOCKCHAIN era) → Property) → Trace (MOCKCHAIN era) → Property

depositEra ∷ ∀ era. Reflect era ⇒ Proof era → TestTree

testEras ∷ String → [Some Proof] → (∀ era. Reflect era ⇒ Proof era → TestTree) → TestTree

Build a TestTree that tests f at all the Eras listed in ps

depositTests ∷ TestTree