Test.Cardano.Ledger.Generic.ApplyTx

Documentation

defaultPPs ∷ [PParamsField era]

pparams ∷ EraPParams era ⇒ Proof era → PParams era

hasValid ∷ [TxField era] → Maybe Bool

applyTx ∷ Reflect era ⇒ Proof era → Int → SlotNo → Model era → Tx era → Model era

epochBoundary ∷ ∀ era. Proof era → EpochNo → EpochNo → Model era → Model era

applyTxSimple ∷ Reflect era ⇒ Proof era → Int → Model era → TxField era → Model era

applyTxBody ∷ Reflect era ⇒ Proof era → Int → Model era → TxBody era → Model era

applyField ∷ Reflect era ⇒ Proof era → Int → Model era → TxBodyField era → Model era

applyWithdrawals ∷ Proof era → Model era → RewardAccount (EraCrypto era) → Coin → Model era

applyCert ∷ ∀ era. Reflect era ⇒ Model era → TxCert era → Model era

applyShelleyCert ∷ ∀ era. EraPParams era ⇒ Model era → ShelleyTxCert era → Model era

data CollInfo era

Constructors

CollInfo

Fields

ciBal ∷ Coin

ciRet ∷ Coin

ciDelset ∷ Set (TxIn (EraCrypto era))

ciAddmap ∷ Map (TxIn (EraCrypto era)) (TxOut era)

emptyCollInfo ∷ CollInfo era

collInfo ∷ (Reflect era, HasCallStack) ⇒ Int → TxIx → Model era → CollInfo era → TxBodyField era → CollInfo era

Collect information about how to process Collateral, in a second phase failure.

updateInfo ∷ CollInfo era → Model era → Model era

applyTxFail ∷ Reflect era ⇒ Proof era → Int → TxIx → Model era → TxField era → Model era

additions ∷ SafeHash (EraCrypto era) EraIndependentTxBody → TxIx → [TxOut era] → Map (TxIn (EraCrypto era)) (TxOut era)

go ∷ IO ()

This is a template of how we might create unit tests that run both the real STS rules and the model to see that they agree. collateralOutputTx and initUTxO are from the BabbageFeatures.hs unit test file.

filterRewards ∷ EraPParams era ⇒ PParams era → Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))) → (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))), Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

filterAllRewards ∷ EraPParams era ⇒ Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))) → Model era → (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))), Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))), Set (Credential 'Staking (EraCrypto era)), Coin)

applyRUpd ∷ ∀ era. RewardUpdateOld (EraCrypto era) → Model era → Model era

notValidatingTx ∷ (Scriptic era, EraTx era, GoodCrypto (EraCrypto era)) ⇒ Proof era → Tx era