Test.Cardano.Ledger.Generic.Fields

Documentation

data TxField era

Constructors

Body (TxBody era)

BodyI [TxBodyField era]

TxWits (TxWits era)

WitnessesI [WitnessesField era]

AuxData (StrictMaybe (TxAuxData era))

Valid IsValid

Bundled Patterns

pattern AuxData' ∷ [TxAuxData era] → TxField era

pattern Valid' ∷ Bool → TxField era

data TxBodyField era

Constructors

Inputs (Set TxIn)

Collateral (Set TxIn)

RefInputs (Set TxIn)

Outputs (StrictSeq (TxOut era))

CollateralReturn (StrictMaybe (TxOut era))

TotalCol (StrictMaybe Coin)

Certs (StrictSeq (TxCert era))

Withdrawals' Withdrawals

Txfee Coin

Vldt ValidityInterval

TTL SlotNo

Update (StrictMaybe (Update era))

ReqSignerHashes (Set (KeyHash 'Witness))

Mint MultiAsset

WppHash (StrictMaybe ScriptIntegrityHash)

AdHash (StrictMaybe TxAuxDataHash)

Txnetworkid (StrictMaybe Network)

ProposalProc (OSet (ProposalProcedure era))

VotingProc (VotingProcedures era)

CurrentTreasuryValue (StrictMaybe Coin)

TreasuryDonation Coin

Bundled Patterns

pattern Inputs' ∷ [TxIn] → TxBodyField era

pattern Collateral' ∷ [TxIn] → TxBodyField era

pattern RefInputs' ∷ [TxIn] → TxBodyField era

pattern Outputs' ∷ [TxOut era] → TxBodyField era

pattern Certs' ∷ [TxCert era] → TxBodyField era

pattern CollateralReturn' ∷ [TxOut era] → TxBodyField era

pattern Update' ∷ [Update era] → TxBodyField era

pattern ReqSignerHashes' ∷ [KeyHash 'Witness] → TxBodyField era

pattern WppHash' ∷ [ScriptIntegrityHash] → TxBodyField era

pattern AdHash' ∷ [TxAuxDataHash] → TxBodyField era

pattern Txnetworkid' ∷ [Network] → TxBodyField era

data WitnessesField era

Constructors

AddrWits (Set (WitVKey 'Witness))

BootWits (Set BootstrapWitness)

ScriptWits (Map ScriptHash (Script era))

DataWits (TxDats era)

RdmrWits (Redeemers era)

Bundled Patterns

pattern AddrWits' ∷ Era era ⇒ [WitVKey 'Witness] → WitnessesField era

pattern BootWits' ∷ Era era ⇒ [BootstrapWitness] → WitnessesField era

pattern ScriptWits' ∷ ∀ era. EraScript era ⇒ [Script era] → WitnessesField era

pattern DataWits' ∷ Era era ⇒ [Data era] → WitnessesField era

data PParamsField era

Constructors

MinfeeA Coin

MinfeeB Coin
The constant factor for the minimum fee calculation

MaxBBSize Word32
Maximal block body size

MaxTxSize Word32
Maximal transaction size

MaxBHSize Word16
Maximal block header size

KeyDeposit Coin
The amount of a key registration deposit

PoolDeposit Coin
The amount of a pool registration deposit

EMax EpochInterval
epoch bound on pool retirement

NOpt Word16
Desired number of pools

A0 NonNegativeInterval
Pool influence

Rho UnitInterval
Monetary expansion

Tau UnitInterval
Treasury expansion

D UnitInterval
Decentralization parameter

ExtraEntropy Nonce
Extra entropy

ProtocolVersion ProtVer
Protocol version

MinPoolCost Coin
Minimum Stake Pool Cost

MinUTxOValue Coin
Minimum Lovelace in a UTxO deprecated by AdaPerUTxOWord

CoinPerUTxOWord CoinPerWord
Cost in ada per 8 bytes of UTxO storage instead of _minUTxOValue

CoinPerUTxOByte CoinPerByte
Cost in ada per 1 byte of UTxO storage instead of _coinsPerUTxOWord

Costmdls CostModels
Cost models for non-native script languages

Prices Prices
Prices of execution units for non-native script languages

MaxTxExUnits ExUnits
Max total script execution resources units allowed per tx

MaxBlockExUnits ExUnits
Max total script execution resources units allowed per block

MaxValSize Natural
Max size of a Value in an output

CollateralPercentage Natural
The scaling percentage of the collateral relative to the fee

MaxCollateralInputs Natural
Maximum number of collateral inputs allowed in a transaction

PoolVotingThreshold PoolVotingThresholds
These are new to Conway

DRepVotingThreshold DRepVotingThresholds

MinCommitteeSize Natural

CommitteeTermLimit EpochInterval

GovActionExpiration EpochInterval

GovActionDeposit Coin

DRepDeposit Coin

DRepActivity EpochInterval

data TxOutField era

Constructors

Address Addr

Amount (Value era)

DHash (StrictMaybe DataHash)

FDatum (Datum era)

RefScript (StrictMaybe (Script era))

Bundled Patterns

pattern DHash' ∷ [DataHash] → TxOutField era

pattern RefScript' ∷ [Script era] → TxOutField era

initVI ∷ ValidityInterval

initWithdrawals ∷ Withdrawals

initialTxBody ∷ Proof era → TxBody era

initialWitnesses ∷ Proof era → TxWits era

initialTx ∷ ∀ era. Proof era → Tx era

initialTxOut ∷ Proof era → TxOut era

valid ∷ IsValid → Bool

abstractTx ∷ Proof era → Tx era → [TxField era]

abstractTxBody ∷ Proof era → TxBody era → [TxBodyField era]

abstractTxOut ∷ Era era ⇒ Proof era → TxOut era → [TxOutField era]

abstractWitnesses ∷ Proof era → TxWits era → [WitnessesField era]

abstractPParams ∷ Proof era → PParams era → [PParamsField era]

abstractPParamsUpdate ∷ Proof era → PParamsUpdate era → [Maybe (PParamsField era)]

abstractPPUpdate ∷ Proof era → PParamsUpdate era → [PParamsField era]