Test.Cardano.Ledger.Generic.Functions

Description

Functions in this module take a (Proof era) as their first parameter and do something potentially different in each Era.

Documentation

depositsAndRefunds ∷ (EraAccounts era, EraPParams era, ShelleyEraTxCert era) ⇒ PParams era → [TxCert era] → Accounts era → Coin

Positive numbers are "deposits owed", negative amounts are "refunds gained"

scriptWitsNeeded' ∷ Proof era → MUtxo era → TxBody era → Set ScriptHash

Compute the set of ScriptHashes for which there should be ScriptWitnesses. In Babbage Era and later, where inline Scripts are allowed, they should not appear in this set.

scriptsNeeded' ∷ EraUTxO era ⇒ MUtxo era → TxBody era → Set ScriptHash

txInBalance ∷ ∀ era. EraTxOut era ⇒ Set TxIn → MUtxo era → Coin

injectFee ∷ EraTxOut era ⇒ Coin → TxOut era → TxOut era

getTxOutRefScript ∷ Proof era → TxOut era → StrictMaybe (Script era)

emptyPPUPstate ∷ ∀ era. Proof era → ShelleyGovState era

maxRefInputs ∷ Proof era → Int

isValid' ∷ Proof era → Tx era → IsValid

txoutEvidence ∷ ∀ era. Proof era → TxOut era → ([Credential 'Payment], Maybe DataHash)

Does the TxOut have evidence of credentials and data. Evidence of data is either ScriptHash or (in Babbage) an inline Datum Evidence of credentials can come from the Addr

addrCredentials ∷ Addr → [Credential 'Payment]

paymentCredAddr ∷ Addr → Maybe (Credential 'Payment)

stakeCredAddr ∷ Addr → Maybe (Credential 'Staking)

getBody ∷ EraTx era ⇒ Proof era → Tx era → TxBody era

getCollateralInputs ∷ Proof era → TxBody era → Set TxIn

getCollateralOutputs ∷ Proof era → TxBody era → [TxOut era]

alwaysSucceedsLang' ∷ ∀ era. EraPlutusContext era ⇒ Language → Natural → Script era

alwaysFailsLang' ∷ ∀ era. EraPlutusContext era ⇒ Language → Natural → Script era

alwaysTrue ∷ ∀ era. EraPlutusContext era ⇒ Maybe Language → Natural → Script era

alwaysFalse ∷ ∀ era. EraPlutusContext era ⇒ Maybe Language → Natural → Script era

certs ∷ (ShelleyEraTxBody era, EraTx era) ⇒ Proof era → Tx era → [TxCert era]

createRUpdNonPulsing' ∷ ∀ era. (EraPParams era, EraAccounts era) ⇒ Model era → RewardUpdateOld

Create an old style RewardUpdate to be used in tests, in any Era.

languagesUsed ∷ ∀ era. Proof era → Tx era → UTxO era → Set ScriptHash → Set Language

languages ∷ ∀ era. (EraUTxO era, AlonzoEraScript era) ⇒ Tx era → UTxO era → Set ScriptHash → Set Language

Compute the Set of Languages in an era, where AlonzoScripts are used

class TotalAda t where

Compute the total Ada from Ada pots within t

Methods

totalAda ∷ t → Coin

govStateTotalAda ∷ ∀ era. Reflect era ⇒ GovState era → Coin

certStateTotalAda ∷ ∀ era. Reflect era ⇒ CertState era → Coin

adaPots ∷ Proof era → EpochState era → AdaPots