Test.Cardano.Ledger.Generic.ModelState

Description

The data types in this file constitute a Model of the NewEpochState sufficient for generating transactions that can run in the MockChain STS instance. This is a model so we drop some details of the real NewEpochState, and add some additional details specific to Tx generation. Dropped details include 1) Efficiency concerns a) Pulsing Reward updates b) Inverse of the Rewards Map c) esNonMyopic field d) _stakeDistro, Incremental Stake distribution 2) Transaction features that make changes to the Protocol Parameters 3) Using Hashes of the TxBody as the Ix, instead we maintain an index of an arbitrary Hash to the sequence (Count) in which the TxBody was generated. Additional details 1) Utxo entries to pay fees. It is incredibly hard to generate Txs with the correct fees. So we keep a small set of UtxoEntrys, where it is allowed to mutate the value field of the TxOut. We have to be sure these entries are carefully managed, as they do not follow the rules of the real world. Additional comments 1) We include data in the Model for Epoch boundary computations, but we do not do anything with them at this time.

Documentation

type MUtxo era = Map TxIn (TxOut era)

MUtxo = Model UTxO. In the Model we represent the UTxO as a Map (not a newtype around a Map)

data ModelNewEpochState era

Constructors

ModelNewEpochState

Fields

mStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

mAccounts ∷ !(Accounts era)

mUTxO ∷ !(Map TxIn (TxOut era))

mMutFee ∷ !(Map TxIn (TxOut era))

The current UTxO

mChainAccountState ∷ !ChainAccountState

mPoolDistr ∷ !(Map (KeyHash 'StakePool) IndividualPoolStake)

mPParams ∷ !(PParams era)

mDeposited ∷ !Coin

mFees ∷ !Coin

mCount ∷ !Int

mIndex ∷ !(Map Int TxId)

mFStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

mRetiring ∷ !(Map (KeyHash 'StakePool) EpochNo)

mSnapshots ∷ !SnapShots

mEL ∷ !EpochNo

mBprev ∷ !(Map (KeyHash 'StakePool) Natural)

mBcur ∷ !(Map (KeyHash 'StakePool) Natural)

mRu ∷ !(StrictMaybe RewardUpdate)

type UtxoEntry era = (TxIn, TxOut era)

type Model era = ModelNewEpochState era

blocksMadeZero ∷ BlocksMade

poolDistrZero ∷ PoolDistr

accountStateZero ∷ ChainAccountState

utxoZero ∷ UTxO era

genDelegsZero ∷ GenDelegs

instantaneousRewardsZero ∷ InstantaneousRewards

dStateZero ∷ EraCertState era ⇒ DState era

pStateZero ∷ PState era

dPStateZero ∷ EraCertState era ⇒ CertState era

nonMyopicZero ∷ NonMyopic

uTxOStateZero ∷ ∀ era. Reflect era ⇒ UTxOState era

ledgerStateZero ∷ ∀ era. Reflect era ⇒ LedgerState era

epochStateZero ∷ Reflect era ⇒ EpochState era

newEpochStateZero ∷ ∀ era. Reflect era ⇒ NewEpochState era

stashedAVVMAddressesZero ∷ Proof era → StashedAVVMAddresses era

mNewEpochStateZero ∷ (EraPParams era, EraAccounts era) ⇒ ModelNewEpochState era

testNES ∷ NewEpochState BabbageEra

testMNES ∷ ModelNewEpochState BabbageEra

class Extract t era where

Methods

extract ∷ ModelNewEpochState era → t

extractCertState ∷ ∀ era. Reflect era ⇒ ModelNewEpochState era → CertState era

abstract ∷ (EraGov era, EraCertState era) ⇒ NewEpochState era → ModelNewEpochState era

complete ∷ PulsingRewUpdate → RewardUpdate

pcModelNewEpochState ∷ ShelleyEraTest era ⇒ ModelNewEpochState era → Expr