Test.Cardano.Ledger.Generic.PrettyCore

Documentation

class PrettyA t where

Methods

prettyA ∷ t → PDoc

isEmpty ∷ Doc ann → Bool

putDoc ∷ Doc ann → IO ()

newtype PrettyAnn

Constructors

Width Int

type Ann = [PrettyAnn]

type PDoc = Doc Ann

text ∷ Text → Doc ann

trim ∷ PDoc → PDoc

ppString ∷ String → Doc a

ppDouble ∷ Double → Doc a

ppInteger ∷ Integer → Doc a

ppRational ∷ Rational → Doc a

ppFloat ∷ Float → Doc a

ppNatural ∷ Natural → Doc a

ppWord64 ∷ Word64 → Doc a

ppWord32 ∷ Word32 → Doc a

ppWord8 ∷ Word8 → Doc a

ppWord16 ∷ Word16 → Doc a

ppFixedPoint ∷ FixedPoint → Doc a

ppPair ∷ (t1 → PDoc) → (t2 → PDoc) → (t1, t2) → PDoc

ppSignedDSIGN ∷ Show a ⇒ a → PDoc

ppBool ∷ Bool → Doc a

ppInt ∷ Int → Doc a

long_bech32 ∷ ByteString → Text

lazy_bech32 ∷ ByteString → Text

ppLong ∷ ByteString → PDoc

ppLazy ∷ ByteString → PDoc

equate ∷ Doc a → Doc a → Doc a

x == y

arrow ∷ (Doc a, Doc a) → Doc a

x -> y

ppSexp ∷ Text → [PDoc] → PDoc

ppSexp x [w,y,z] --> (x w y z)

ppSexp' ∷ PDoc → [PDoc] → PDoc

ppRecord ∷ Text → [(Text, PDoc)] → PDoc

ppRecord name [("a",x),("b",y),("c",z)] --> name { a = x, b = y, c = z }

ppRecord' ∷ PDoc → [(Text, PDoc)] → PDoc

puncLeft ∷ Doc ann → [Doc ann] → Doc ann → Doc ann → Doc ann

Vertical layout with commas aligned on the left hand side

ppSet ∷ (x → Doc ann) → Set x → Doc ann

ppSeq ∷ (a → Doc ann) → Seq a → Doc ann

ppList ∷ (x → Doc ann) → [x] → Doc ann

ppStrictSeq ∷ (a → Doc ann) → StrictSeq a → Doc ann

ppOSet ∷ (a → Doc ann) → OSet a → Doc ann

ppStrictMaybe ∷ (x → Doc ann) → StrictMaybe x → Doc ann

ppMaybe ∷ (x → Doc ann) → Maybe x → Doc ann

ppAssocList ∷ PDoc → (k → PDoc) → (v → PDoc) → [(k, v)] → PDoc

ppMap' ∷ PDoc → (k → PDoc) → (v → PDoc) → Map k v → PDoc

ppMap ∷ (k → PDoc) → (v → PDoc) → Map k v → PDoc

ppVMap ∷ (Vector kv k, Vector vv v) ⇒ (k → PDoc) → (v → PDoc) → VMap kv vv k v → PDoc

ppPolicyID ∷ PolicyID c → PDoc

ppLogWeight ∷ LogWeight → PDoc

ppPrices ∷ Prices → PDoc

ppRewardType ∷ RewardType → PDoc

ppLikelihood ∷ Likelihood → PDoc

ppNonMyopic ∷ NonMyopic c → PDoc

ppRewardUpdate ∷ RewardUpdate c → PDoc

ppLanguage ∷ Language → PDoc

ppTxWitness ∷ ∀ era. Reflect era ⇒ AlonzoTxWits era → PDoc

ppIsValid ∷ IsValid → PDoc

ppVersion ∷ Version → PDoc

ppNetwork ∷ Network → PDoc

ppValidityInterval ∷ ValidityInterval → PDoc

ppAuxiliaryDataHash ∷ AuxiliaryDataHash c → PDoc

ppSafeHash ∷ SafeHash c index → PDoc

ppEpochNo ∷ EpochNo → Doc ann

ppEpochInterval ∷ EpochInterval → Doc ann

ppHash ∷ Hash a b → PDoc

ppUnitInterval ∷ UnitInterval → PDoc

ppNonce ∷ Nonce → PDoc

ppPtr ∷ Ptr → PDoc

ppProtVer ∷ ProtVer → PDoc

ppVKey ∷ Crypto c ⇒ VKey r c → PDoc

ppBootstrapWitness ∷ Crypto c ⇒ BootstrapWitness c → PDoc

ppWitnessSetHKD ∷ ∀ era. Reflect era ⇒ ShelleyTxWits era → PDoc

ppPParamsUpdate ∷ (ProtVerAtMost era 4, ProtVerAtMost era 6, ProtVerAtMost era 8, EraPParams era) ⇒ PParamsUpdate era → PDoc

ppUpdate ∷ PrettyA (PParamsUpdate era) ⇒ Update era → PDoc

ppProposedPPUpdates ∷ PrettyA (PParamsUpdate era) ⇒ ProposedPPUpdates era → PDoc

ppMetadatum ∷ Metadatum → PDoc

ppShelleyTxAuxData ∷ Era era ⇒ ShelleyTxAuxData era → PDoc

ppAllegraTxAuxData ∷ (AllegraEraScript era, Reflect era, NativeScript era ~ Timelock era) ⇒ AllegraTxAuxData era → PDoc

ppAlonzoTxAuxData ∷ Reflect era ⇒ AlonzoTxAuxData era → PDoc

extractAlonzoTxAuxDataScripts ∷ ∀ era. Reflect era ⇒ AlonzoTxAuxData era → StrictSeq (Script era)

pcAuxData ∷ Proof era → TxAuxData era → PDoc

ppWithdrawals ∷ Withdrawals c → PDoc

ppWitHashes ∷ Set (KeyHash 'Witness c) → PDoc

ppShelleyTx ∷ Reflect era ⇒ Tx era → PDoc

ppAlonzoTx ∷ ∀ era. (Reflect era, Tx era ~ AlonzoTx era) ⇒ AlonzoTx era → PDoc

ppShelleyTxBody ∷ (Reflect era, PrettyA (PParamsUpdate era)) ⇒ ShelleyTxBody era → PDoc

ppAllegraTxBody ∷ ∀ era. (TxBody era ~ AllegraTxBody era, Reflect era) ⇒ AllegraTxBody era → PDoc

ppAlonzoTxBody ∷ ∀ era. (TxBody era ~ AlonzoTxBody era, Reflect era) ⇒ AlonzoTxBody era → PDoc

ppMaryTxBody ∷ ∀ era. (TxBody era ~ MaryTxBody era, Reflect era) ⇒ MaryTxBody era → PDoc

ppCoreWitnesses ∷ Proof era → TxWits era → PDoc

pcTxOut ∷ Reflect era ⇒ Proof era → TxOut era → PDoc

pcScript ∷ ∀ era. Reflect era ⇒ Proof era → Script era → PDoc

pcWitnesses ∷ Reflect era ⇒ Proof era → TxWits era → PDoc

pcTx ∷ Proof era → Tx era → PDoc

Pretty print a Tx. Also see Test.Cardano.Ledger.Constrained.Preds.Tx(pcTxWithUTxO) which expands the inputs (given the UTxO)

pcTxBody ∷ Proof era → TxBody era → PDoc

pcPParams ∷ Proof era → PParams era → PDoc

pcTxBodyField ∷ Proof era → TxBodyField era → [(Text, PDoc)]

pcTxField ∷ ∀ era. Reflect era ⇒ Proof era → TxField era → [(Text, PDoc)]

pcWitnessesField ∷ ∀ era. Reflect era ⇒ Proof era → WitnessesField era → [(Text, PDoc)]

pcPParamsField ∷ PParamsField era → [(Text, PDoc)]

ppUTXOW ∷ Reflect era ⇒ Proof era → PredicateFailure (EraRule "UTXOW" era) → PDoc

ppUTXOS ∷ Proof era → PredicateFailure (EraRule "UTXOS" era) → PDoc

ppDELEGS ∷ Proof era → PredicateFailure (EraRule "DELEGS" era) → PDoc

ppDELEG ∷ Proof era → PredicateFailure (EraRule "DELEG" era) → PDoc

ppPOOL ∷ Proof era → PredicateFailure (EraRule "POOL" era) → PDoc

ppLEDGER ∷ Reflect era ⇒ Proof era → PredicateFailure (EraRule "LEDGER" era) → PDoc

ppUTXO ∷ Proof era → PredicateFailure (EraRule "UTXO" era) → PDoc

ppLEDGERS ∷ Proof era → PredicateFailure (EraRule "LEDGERS" era) → PDoc

ppDELPL ∷ Proof era → PredicateFailure (EraRule "DELPL" era) → PDoc

ppNEWEPOCH ∷ Proof era → PredicateFailure (EraRule "NEWEPOCH" era) → PDoc

ppEPOCH ∷ Proof era → PredicateFailure (EraRule "EPOCH" era) → PDoc

ppUPEC ∷ Proof era → UpecPredFailure era → PDoc

A bit different since it is NOT of the form: PredicateFailure (EraRule UPEC era) but instead, the type family UpecPredFailurePV pv era where, type UpecPredFailure era = UpecPredFailurePV (ProtVerLow era) era But the effect is still the same. The Proof era, fixes the type family result.

ppStateLEDGERS ∷ Proof era → State (EraRule "LEDGERS" era) → PDoc

A bit different since it is NOT of the form: PredicateFailure (EraRule LEDGERS era) but instead: State (EraRule LEDGERS era) But the effect is still the same. The Proof era, fixes the type family result.

ppShelleyDelegsPredFailure ∷ ∀ era. Proof era → ShelleyDelegsPredFailure era → PDoc

ppConwayUtxoPredFailure ∷ ∀ era. Reflect era ⇒ ConwayUtxoPredFailure era → PDoc

ppBabbageUtxoPredFailure ∷ Reflect era ⇒ BabbageUtxoPredFailure era → PDoc

ppShelleyLedgersPredFailure ∷ Reflect era ⇒ Proof era → ShelleyLedgersPredFailure era → PDoc

ppConwayLedgerPredFailure ∷ Reflect era ⇒ Proof era → ConwayLedgerPredFailure era → PDoc

ppConwayCertPredFailure ∷ Proof era → ConwayCertPredFailure era → PDoc

ppConwayGovCertPredFailure ∷ ConwayGovCertPredFailure era → PDoc

ppConwayCertsPredFailure ∷ Proof era → ConwayCertsPredFailure era → PDoc

ppConwayGovPredFailure ∷ ConwayGovPredFailure era → PDoc

ppShelleyLedgerPredFailure ∷ Reflect era ⇒ Proof era → ShelleyLedgerPredFailure era → PDoc

ppConwayUtxowPredFailure ∷ Reflect era ⇒ Proof era → ConwayUtxowPredFailure era → PDoc

ppBabbageUtxowPredFailure ∷ Reflect era ⇒ Proof era → BabbageUtxowPredFailure era → PDoc

ppBbodyPredicateFailure ∷ ∀ era. Reflect era ⇒ ShelleyBbodyPredFailure era → PDoc

ppConwayBbodyPredFail ∷ ∀ era. Reflect era ⇒ ConwayBbodyPredFailure era → PDoc

ppAlonzoBbodyPredFail ∷ Reflect era ⇒ AlonzoBbodyPredFailure era → PDoc

ppTickPredicateFailure ∷ ∀ era. Reflect era ⇒ ShelleyTickPredFailure era → PDoc

ppShelleyNewEpochPredicateFailure ∷ ∀ era. Reflect era ⇒ ShelleyNewEpochPredFailure era → PDoc

ppConwayNewEpochPredFailure ∷ ConwayNewEpochPredFailure era → PDoc

ppShelleyEpochPredFailure ∷ ∀ era. Reflect era ⇒ ShelleyEpochPredFailure era → PDoc

ppUpecPredicateFailure ∷ ShelleyUpecPredFailure era → PDoc

ppNewppPredicateFailure ∷ ShelleyNewppPredFailure era → PDoc

ppAlonzoUtxowPredFailure ∷ Reflect era ⇒ AlonzoUtxowPredFailure era → PDoc

ppShelleyUtxowPredFailure ∷ ∀ era. Reflect era ⇒ ShelleyUtxowPredFailure era → PDoc

ppAlonzoUtxoPredFailure ∷ ∀ era. Reflect era ⇒ AlonzoUtxoPredFailure era → PDoc

ppShelleyDelegPredFailure ∷ ShelleyDelegPredFailure era → PDoc

ppShelleyDelplPredFailure ∷ Proof era → ShelleyDelplPredFailure era → PDoc

ppConwayDelegPredFailure ∷ ConwayDelegPredFailure era → PDoc

ppShelleyPoolPredFailure ∷ ShelleyPoolPredFailure era → PDoc

ppUtxosPredicateFailure ∷ ∀ era. Reflect era ⇒ AlonzoUtxosPredFailure era → PDoc

ppCollectError ∷ Reflect era ⇒ CollectError era → PDoc

ppContextError ∷ ∀ era. Reflect era ⇒ ContextError era → PDoc

ppPlutusPurposeAsIx ∷ Reflect era ⇒ PlutusPurpose AsIx era → PDoc

ppPlutusPurposeAsItem ∷ Reflect era ⇒ PlutusPurpose AsItem era → PDoc

ppPlutusPurposeAsIxItem ∷ Reflect era ⇒ PlutusPurpose AsIxItem era → PDoc

ppPlutusPurpose ∷ ∀ f era. (Reflect era, ∀ ix it. (PrettyA ix, PrettyA it) ⇒ PrettyA (f ix it)) ⇒ PlutusPurpose f era → PDoc

ppTagMismatchDescription ∷ TagMismatchDescription → PDoc

ppFailureDescription ∷ FailureDescription → PDoc

ppShelleyUtxoPredFailure ∷ ∀ era. Reflect era ⇒ ShelleyUtxoPredFailure era → PDoc

ppPpupPredicateFailure ∷ ShelleyPpupPredFailure era → PDoc

ppPPUPPredFailure ∷ a → PDoc

ppAllegraUtxoPredFailure ∷ ∀ era. Reflect era ⇒ AllegraUtxoPredFailure era → PDoc

ppBbodyState ∷ ∀ era. Reflect era ⇒ ShelleyBbodyState era → PDoc

txBodyFieldSummary ∷ EraTxBody era ⇒ TxBodyField era → [(Text, PDoc)]

bodySummary ∷ EraTxBody era ⇒ Proof era → TxBody era → PDoc

witnessFieldSummary ∷ WitnessesField era → (Text, PDoc)

witnessSummary ∷ Proof era → TxWits era → PDoc

txFieldSummary ∷ EraTxBody era ⇒ Proof era → TxField era → [PDoc]

txSummary ∷ EraTx era ⇒ Proof era → Tx era → PDoc

txInSummary ∷ TxIn era → PDoc

txOutSummary ∷ Proof era → TxOut era → PDoc

datumSummary ∷ Era era ⇒ Datum era → PDoc

dataSummary ∷ Era era ⇒ Data era → PDoc

plutusDataSummary ∷ Data → PDoc

multiAssetSummary ∷ MultiAsset c → PDoc

vSummary ∷ MaryValue c → PDoc

scriptSummary ∷ ∀ era. Proof era → Script era → PDoc

networkSummary ∷ Network → PDoc

addrSummary ∷ Addr c → PDoc

credSummary ∷ Credential keyrole c → PDoc

stakeSummary ∷ StakeReference c → PDoc

utxoSummary ∷ Proof era → UTxO era → PDoc

utxoString ∷ Proof era → UTxO era → String

scriptHashSummary ∷ ScriptHash c → PDoc

keyHashSummary ∷ KeyHash keyrole c → PDoc

dataHashSummary ∷ DataHash era → PDoc

keyPairSummary ∷ Crypto c ⇒ KeyPair r c → PDoc

vKeySummary ∷ Crypto c ⇒ VKey r c → PDoc

timelockSummary ∷ (AllegraEraScript era, NativeScript era ~ Timelock era) ⇒ Timelock era → PDoc

multiSigSummary ∷ (ShelleyEraScript era, NativeScript era ~ MultiSig era) ⇒ MultiSig era → PDoc

plutusSummary ∷ ∀ era. Proof era → AlonzoScript era → PDoc

dStateSummary ∷ DState c → PDoc

instantSummary ∷ InstantaneousRewards c → PDoc

uMapSummary ∷ UMap c → PDoc

pStateSummary ∷ PState c → PDoc

dpStateSummary ∷ CertState era → PDoc

pcTxId ∷ TxId c → PDoc

pcTxIn ∷ TxIn c → PDoc

pcNetwork ∷ Network → PDoc

pcKeyHash ∷ KeyHash discriminator c → PDoc

pcCredential ∷ Credential keyrole c → PDoc

pcStakeReference ∷ StakeReference c → PDoc

pcAddr ∷ Addr c → PDoc

pcCoreValue ∷ Proof era → Value era → PDoc

Value is a type family, so it has no PrettyA instance.

pcCoin ∷ Coin → PDoc

pcValue ∷ MaryValue c → PDoc

pcVal ∷ Proof era → Value era → PDoc

pcDatum ∷ Era era ⇒ Datum era → PDoc

pcData ∷ ∀ era. Era era ⇒ Data era → PDoc

pcTimelock ∷ ∀ era. (AllegraEraScript era, Reflect era, NativeScript era ~ Timelock era) ⇒ Timelock era → PDoc

pcMultiSig ∷ (ShelleyEraScript era, Reflect era, NativeScript era ~ MultiSig era) ⇒ PDoc → MultiSig era → PDoc

pcScriptHash ∷ ScriptHash era → PDoc

pcHashScript ∷ ∀ era. Reflect era ⇒ Proof era → Script era → PDoc

pcDataHash ∷ DataHash era → PDoc

pcUTxO ∷ Proof era → UTxO era → PDoc

pcPoolParams ∷ PoolParams era → PDoc

pcDelegCert ∷ ShelleyDelegCert c → PDoc

pcPoolCert ∷ PoolCert c → PDoc

pcShelleyTxCert ∷ ShelleyTxCert c → PDoc

pcConwayTxCert ∷ ConwayTxCert c → PDoc

pcConwayGovCert ∷ ConwayGovCert c → PDoc

pcConwayDelegCert ∷ ConwayDelegCert c → PDoc

pcDelegatee ∷ Delegatee c → PDoc

pcTxCert ∷ Proof era → TxCert era → PDoc

pcGovSignal ∷ ∀ era. Reflect era ⇒ GovSignal era → PDoc

pcVotingProcedures ∷ VotingProcedures era → PDoc

pcProposalProcedure ∷ ProposalProcedure era → PDoc

pcVoter ∷ Voter c → PDoc

pcVotingProcedure ∷ VotingProcedure era → PDoc

pcRewardAccount ∷ RewardAccount c → PDoc

pcExUnits ∷ ExUnits → PDoc

pcPair ∷ (t1 → PDoc) → (t2 → PDoc) → (t1, t2) → PDoc

pcWitVKey ∷ ∀ era keyrole. (Reflect era, Typeable keyrole) ⇒ Proof era → WitVKey keyrole (EraCrypto era) → PDoc

pcGovState ∷ Proof era → GovState era → PDoc

GovState is a type family, No PrettyA instance

pcShelleyGovState ∷ Proof era → ShelleyGovState era → PDoc

pcFuturePParams ∷ Proof era → FuturePParams era → PDoc

pcEnactState ∷ Proof era → EnactState era → PDoc

pcGovActionId ∷ GovActionId c → PDoc

pcGovPurposeId ∷ GovPurposeId p era → PDoc

pcPrevGovActionIds ∷ GovRelation StrictMaybe era → PDoc

pcConwayGovState ∷ Reflect era ⇒ Proof era → ConwayGovState era → PDoc

pcPulsingSnapshot ∷ PulsingSnapshot era → PDoc

pcDRepPulsingState ∷ Proof era → DRepPulsingState era → PDoc

pcRatifyState ∷ Proof era → RatifyState era → PDoc

pcProposals ∷ Proposals era → PDoc

pcPEdges ∷ PEdges (GovPurposeId purpose era) → PDoc

pcForest ∷ GovRelation PGraph era → PDoc

pcGovActionState ∷ GovActionState era → PDoc

pcVote ∷ Vote → PDoc

pcCommittee ∷ Committee era → PDoc

pcGovAction ∷ GovAction era → PDoc

pcConstitution ∷ Constitution c → PDoc

ppCommitteeAuthorization ∷ CommitteeAuthorization c → PDoc

pcCommitteeState ∷ CommitteeState era → PDoc

pcReward ∷ Reward c → PDoc

pcFutureGenDeleg ∷ FutureGenDeleg c → PDoc

pcCertState ∷ CertState era → PDoc

pcVState ∷ VState era → PDoc

pcAnchor ∷ Anchor c → PDoc

pcDRepState ∷ DRepState c → PDoc

pcDRep ∷ DRep c → PDoc

pcSnapShotL ∷ Text → SnapShot c → [(Text, PDoc)]

pcIndividualPoolStake ∷ IndividualPoolStake c → PDoc

pcSnapShots ∷ SnapShots c → PDoc

pcPoolDistr ∷ PoolDistr c → PDoc

withEraPParams ∷ ∀ era a. Proof era → (EraPParams era ⇒ a) → a

pcPParamsSynopsis ∷ ∀ era. Proof era → PParams era → PDoc

Print just a few of the PParams fields

showProtver ∷ ProtVer → String

pcEpochState ∷ Reflect era ⇒ Proof era → EpochState era → PDoc

pcAccountState ∷ AccountState → PDoc

psEpochState ∷ Reflect era ⇒ Proof era → EpochState era → PDoc

Like pcEpochState.but it only prints a summary of the UTxO

pcNewEpochState ∷ Reflect era ⇒ Proof era → NewEpochState era → PDoc

psNewEpochState ∷ Reflect era ⇒ Proof era → NewEpochState era → PDoc

Like pcEpochState.but it only prints a summary of the UTxO

pcUTxOState ∷ Proof era → UTxOState era → PDoc

psUTxOState ∷ ∀ era. Reflect era ⇒ Proof era → UTxOState era → PDoc

Like pcUTxOState, except it prints only a summary of the UTxO

pcLedgerState ∷ Proof era → LedgerState era → PDoc

psLedgerState ∷ Reflect era ⇒ Proof era → LedgerState era → PDoc

Like pcLedgerState, except it prints only a summary of the UTxO

pcPState ∷ PState era → PDoc

pcDState ∷ DState c → PDoc

pcGenDelegPair ∷ GenDelegPair c → PDoc

pcIRewards ∷ InstantaneousRewards c → PDoc

pcDeltaCoin ∷ DeltaCoin → PDoc

pcSlotNo ∷ SlotNo → PDoc

pcAdaPot ∷ (EraTxOut era, EraGov era) ⇒ EpochState era → PDoc

pcPolicyID ∷ PolicyID c → PDoc

pcAssetName ∷ AssetName → PDoc

pcMultiAsset ∷ MultiAsset c → PDoc

pcScriptsNeeded ∷ Reflect era ⇒ Proof era → ScriptsNeeded era → PDoc

pcDRepPulser ∷ DRepPulser era Identity (RatifyState era) → PDoc

summaryMapCompact ∷ Map a (CompactForm Coin) → PDoc

pcConwayGovCertEnv ∷ ∀ era. Reflect era ⇒ ConwayGovCertEnv era → PDoc

pcPoolEnv ∷ Reflect era ⇒ PoolEnv era → PDoc

pcEnactSignal ∷ EnactSignal era → PDoc

pcRatifySignal ∷ RatifySignal era → PDoc

pcRatifyEnv ∷ RatifyEnv era → PDoc

pcGovEnv ∷ Reflect era ⇒ GovEnv era → PDoc