Test.Cardano.Ledger.Generic.Properties

Documentation

genTxAndUTXOState ∷ (Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, State (EraRule "UTXOW" era) ~ UTxOState era, Tx era ~ Signal (EraRule "UTXOW" era), EraGenericGen era) ⇒ GenSize → Gen (TRC (EraRule "UTXOW" era), GenState era)

genTxAndLEDGERState ∷ ∀ era. (Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era, EraGenericGen era) ⇒ GenSize → Gen (TRC (EraRule "LEDGER" era), GenState era)

testTxValidForLEDGER ∷ ∀ era. (Reflect era, Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, ToExpr (PredicateFailure (EraRule "LEDGER" era)), EraTest era, BaseM (EraRule "LEDGER" era) ~ ShelleyBase, STS (EraRule "LEDGER" era), ToExpr (Environment (EraRule "LEDGER" era))) ⇒ (TRC (EraRule "LEDGER" era), GenState era) → Property

txPreserveAda ∷ GenSize → TestTree

A single Tx preserves Ada

adaIsPreserved ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → TestTree

Ada is preserved over a trace of length 100

tracePreserveAda ∷ Int → GenSize → TestTree

stakeInvariant ∷ EraStake era ⇒ MockChainState era → MockChainState era → Property

The incremental Stake invaraint is preserved over a trace of length 100=

incrementStakeInvariant ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ GenSize → TestTree

incrementalStake ∷ GenSize → TestTree

genericProperties ∷ GenSize → TestTree

epochPreserveAda ∷ GenSize → TestTree

adaIsPreservedInEachEpoch ∷ ∀ era. (ShelleyEraAccounts era, State (EraRule "NEWEPOCH" era) ~ NewEpochState era, State (EraRule "RUPD" era) ~ StrictMaybe PulsingRewUpdate, State (EraRule "TICK" era) ~ NewEpochState era, State (EraRule "LEDGER" era) ~ LedgerState era, State (EraRule "LEDGERS" era) ~ LedgerState era, Environment (EraRule "NEWEPOCH" era) ~ (), Environment (EraRule "RUPD" era) ~ RupdEnv era, Environment (EraRule "LEDGERS" era) ~ ShelleyLedgersEnv era, Environment (EraRule "TICK" era) ~ (), Environment (EraRule "LEDGER" era) ~ LedgerEnv era, Signal (EraRule "NEWEPOCH" era) ~ EpochNo, Signal (EraRule "RUPD" era) ~ SlotNo, Signal (EraRule "LEDGERS" era) ~ Seq (Tx era), Signal (EraRule "TICK" era) ~ SlotNo, Signal (EraRule "LEDGER" era) ~ Tx era, BaseM (EraRule "NEWEPOCH" era) ~ ShelleyBase, Embed (EraRule "TICK" era) (MOCKCHAIN era), Embed (EraRule "NEWEPOCH" era) (ShelleyTICK era), Embed (EraRule "RUPD" era) (ShelleyTICK era), Embed (EraRule "LEDGERS" era) (MOCKCHAIN era), EraGenericGen era, ToExpr (PredicateFailure (EraRule "NEWEPOCH" era)), ToExpr (PredicateFailure (EraRule "RUPD" era)), ToExpr (PredicateFailure (EraRule "LEDGER" era)), Eq (PredicateFailure (EraRule "LEDGER" era)), Show (PredicateFailure (EraRule "LEDGER" era))) ⇒ GenSize → TestTree

twiddleInvariantHolds ∷ ∀ a. (Arbitrary a, Show a, Twiddle a) ⇒ String → TestTree

twiddleInvariantHoldsEras ∷ TestTree