Test.Cardano.Ledger.Generic.Same

Description

A Proof parameterized equality test, that records the sameness of individual record fields, this way it is possible to know where the equality failed.

Documentation

extendLabel ∷ String → [(String, Maybe x)] → [(String, Maybe x)]

Relabel by appending s to the front of the path

notEq ∷ Doc a → Doc a → Doc a

Dispays a difference vertically as x =/= y

eqByShow ∷ (Eq t, Show t) ⇒ t → t → Maybe PDoc

Compare for equality, and display difference using show

eqVia ∷ Eq t ⇒ (t → PDoc) → t → t → Maybe PDoc

Compare for equality, and display differences using pcf

class Same era t where

The Same class is similar to Eq but returns descriptions (paths and non-matching components) for each difference in a large structure.

Methods

same ∷ Proof era → t → t → [(String, Maybe PDoc)]

sameUTxO ∷ Proof era → UTxO era → UTxO era → Maybe PDoc

samePPUP ∷ Proof era → ShelleyGovState era → ShelleyGovState era → Maybe PDoc

sameStashedAVVMAddresses ∷ Proof era → StashedAVVMAddresses era → StashedAVVMAddresses era → Maybe PDoc

data SomeDepend where

Constructors

SomeD ∷ String → (x → x → Maybe PDoc) → x → x → SomeDepend

SomeM ∷ String → (x → x → [(String, Maybe PDoc)]) → x → x → SomeDepend

sameWithDependency ∷ [SomeDepend] → [(String, Maybe PDoc)]

If x occurs in y then a difference in x forces a difference in y, so only return the information on x if it has differences.

samePParams ∷ Proof era → PParams era → PParams era → Maybe PDoc

samePParamsUpdate ∷ Proof era → PParamsUpdate era → PParamsUpdate era → Maybe PDoc

sameTxOut ∷ Proof era → TxOut era → TxOut era → Maybe PDoc

sameLedgerFail ∷ Proof era → ApplyTxError era → ApplyTxError era → Maybe PDoc

sameTransCtx ∷ Proof era → TranslationContext era → TranslationContext era → Maybe PDoc

sameShelleyTxWits ∷ ∀ era. Reflect era ⇒ Proof era → ShelleyTxWits era → ShelleyTxWits era → [(String, Maybe PDoc)]

sameAlonzoTxWits ∷ ∀ era. (Reflect era, AlonzoEraScript era) ⇒ Proof era → AlonzoTxWits era → AlonzoTxWits era → [(String, Maybe PDoc)]

sameTxWits ∷ Reflect era ⇒ Proof era → TxWits era → TxWits era → [(String, Maybe PDoc)]

sameShelleyTxBody ∷ Reflect era ⇒ Proof era → ShelleyTxBody era → ShelleyTxBody era → [(String, Maybe PDoc)]

sameAllegraTxBody ∷ Reflect era ⇒ Proof era → AllegraTxBody era → AllegraTxBody era → [(String, Maybe PDoc)]

sameMaryTxBody ∷ Reflect era ⇒ Proof era → MaryTxBody era → MaryTxBody era → [(String, Maybe PDoc)]

sameAlonzoTxBody ∷ Reflect era ⇒ Proof era → AlonzoTxBody era → AlonzoTxBody era → [(String, Maybe PDoc)]

sameBabbageTxBody ∷ (Reflect era, BabbageEraTxBody era) ⇒ Proof era → BabbageTxBody era → BabbageTxBody era → [(String, Maybe PDoc)]

sameConwayTxBody ∷ (ConwayEraTxBody era, Reflect era) ⇒ Proof era → ConwayTxBody era → ConwayTxBody era → [(String, Maybe PDoc)]

sameTxBody ∷ Reflect era ⇒ Proof era → TxBody era → TxBody era → [(String, Maybe PDoc)]

sameShelleyTx ∷ (Reflect era, TxWits era ~ ShelleyTxWits era) ⇒ Proof era → ShelleyTx era → ShelleyTx era → [(String, Maybe PDoc)]

sameAlonzoTx ∷ (Reflect era, AlonzoEraScript era, TxWits era ~ AlonzoTxWits era) ⇒ Proof era → AlonzoTx era → AlonzoTx era → [(String, Maybe PDoc)]

sameTx ∷ Reflect era ⇒ Proof era → Tx era → Tx era → [(String, Maybe PDoc)]

ints ∷ [Int]

sameShelleyTxSeq ∷ (Reflect era, Tx era ~ ShelleyTx era, SafeToHash (TxWits era)) ⇒ Proof era → ShelleyTxSeq era → ShelleyTxSeq era → [(String, Maybe PDoc)]

sameAlonzoTxSeq ∷ (Reflect era, AlonzoEraTx era, SafeToHash (TxWits era)) ⇒ Proof era → AlonzoTxSeq era → AlonzoTxSeq era → [(String, Maybe PDoc)]

sameTxSeq ∷ Reflect era ⇒ Proof era → TxSeq era → TxSeq era → [(String, Maybe PDoc)]

Orphan instances

Terse (PParamsUpdate (AllegraEra c))

Instance details

Methods

terse ∷ PParamsUpdate (AllegraEra c) → String

Terse (PParamsUpdate (AlonzoEra c))

Instance details

Methods

terse ∷ PParamsUpdate (AlonzoEra c) → String

Terse (PParamsUpdate (BabbageEra c))

Instance details

Methods

terse ∷ PParamsUpdate (BabbageEra c) → String

Terse (PParamsUpdate (ConwayEra c))

Instance details

Methods

terse ∷ PParamsUpdate (ConwayEra c) → String

Terse (PParamsUpdate (MaryEra c))

Instance details

Methods

terse ∷ PParamsUpdate (MaryEra c) → String

Terse (PParamsUpdate (ShelleyEra c))

Instance details

Methods

terse ∷ PParamsUpdate (ShelleyEra c) → String

Terse (KeyHash 'Genesis c)

Instance details

Methods

terse ∷ KeyHash 'Genesis c → String