Test.Cardano.Ledger.Generic.TxGen

Documentation

genAlonzoTx ∷ ∀ era. Reflect era ⇒ Proof era → SlotNo → GenRS era (UTxO era, Tx era)

data Box era

Constructors

Box (Proof era) (TRC (EraRule "LEDGER" era)) (GenState era)

applySTSByProof ∷ ∀ era. Era era ⇒ Proof era → RuleContext 'Transition (EraRule "LEDGER" era) → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era))

assembleWits ∷ Proof era → [WitnessesField era] → TxWits era

This uses merging semantics, it expects duplicate fields, and merges them together

coreTx ∷ Proof era → [TxField era] → Tx era

coreTxBody ∷ EraTxBody era ⇒ Proof era → [TxBodyField era] → TxBody era

coreTxOut ∷ Proof era → [TxOutField era] → TxOut era

genUTxO ∷ Reflect era ⇒ GenRS era (MUtxo era, Maybe (UtxoEntry era))

Generate a somewhat arbitrary MUtxo. Occasionally add a bit of the MUtxo in the Model to the one generated. This way the Tx we generate may spend some of the old UTxo. The result has at most 1 entry from the old MUtxo, and If it has only one entry, that entry is not from the old MUtxo