Test.Cardano.Ledger.Tickf

Description

Test that modifications to the calculatePoolDistr function made when building the Tickf benchmarks behave the same as the code that was replaced.

Documentation

oldCalculatePoolDistr ∷ ∀ c. (KeyHash 'StakePool c → Bool) → SnapShot c → PoolDistr c

The original version of calculatePoolDistr

calcPoolDistOldEqualsNew ∷ TestTree