Test.Cardano.Ledger.Constrained.Ast

Description

The types that make up the Abstract Syntax Trees of the Language

Documentation

data Term era t where

Constructors

Lit ∷ Rep era t → t → Term era t

Var ∷ V era t → Term era t

Dom ∷ Ord a ⇒ Term era (Map a b) → Term era (Set a)

Rng ∷ (Ord a, Ord b) ⇒ Term era (Map a b) → Term era (Set b)

ProjM ∷ Ord a ⇒ Lens' b t → Rep era t → Term era (Map a b) → Term era (Map a t)

ProjS ∷ (Ord b, Ord t) ⇒ Lens' b t → Rep era t → Term era (Set b) → Term era (Set t)

Delta ∷ Term era Coin → Term era DeltaCoin

Negate ∷ Term era DeltaCoin → Term era DeltaCoin

data Pred era where

Constructors

Sized ∷ Sizeable t ⇒ Term era Size → Term era t → Pred era

(:=:) ∷ Eq a ⇒ Term era a → Term era a → Pred era infix 4

Subset ∷ Ord a ⇒ Term era (Set a) → Term era (Set a) → Pred era

Disjoint ∷ Ord a ⇒ Term era (Set a) → Term era (Set a) → Pred era

SumsTo ∷ Adds c ⇒ c → Term era c → OrdCond → [Sum era c] → Pred era

Random ∷ Term era t → Pred era

Component ∷ Term era t → [AnyF era t] → Pred era

CanFollow ∷ Count n ⇒ Term era n → Term era n → Pred era

data Sum era c where

Constructors

SumMap ∷ Adds c ⇒ Term era (Map a c) → Sum era c

SumList ∷ Adds c ⇒ Term era [c] → Sum era c

One ∷ Term era c → Sum era c

Project ∷ ∀ x c a era. Sums x c ⇒ Rep era c → Term era (Map a x) → Sum era c

pattern (:⊆:) ∷ ∀ era. ∀ a. Ord a ⇒ Term era (Set a) → Term era (Set a) → Pred era infix 4

pattern ExactSize ∷ Int → Term era Size

pattern AtLeast ∷ Int → Term era Size

pattern Size ∷ Size → Term era Size

pattern AtMost ∷ Int → Term era Size

pattern Range ∷ Int → Int → Term era Size

sameRng ∷ Term era Size → Maybe Int

pattern Word64 ∷ Word64 → Term era Word64

data Target era t where

Constructors

Simple ∷ Term era t → Target era t

(:$) ∷ Target era (a → b) → Target era a → Target era b infixl 0

Constr ∷ String → (a → b) → Target era (a → b)

(^$) ∷ Target era (a → t) → Term era a → Target era t infixl 0

Version of (:$) That takes a Term on the right, rather than a Target

constTarget ∷ t → Target era t

showL ∷ (t → String) → [Char] → [t] → [Char]

args ∷ Target era t → [Any (Target era)]

ppTarget ∷ Target era t → PDoc

Print a Target as a record showing the struture and names of all the variables involved. This documents what is in scope where the Target value was defined.

targetRecord ∷ Target era t → [(Text, PDoc)] → PDoc

nameOf ∷ Target era t → Text

targetPair ∷ Target era t → (Text, PDoc)

varsOfTerm ∷ Set (Name era) → Term era t → Set (Name era)

vars ∷ Term era t → Set (Name era)

varsOfTarget ∷ Set (Name era) → Target era t → Set (Name era)

varsOfPred ∷ Set (Name era) → Pred era → Set (Name era)

varsOfSum ∷ Set (Name era) → Sum era r → Set (Name era)

substToEnv ∷ [SubItem era] → Env era → Typed (Env era)

data SubItem era where

Constructors

SubItem ∷ V era t → Term era t → SubItem era

pad ∷ Int → String → String

type Subst era = [SubItem era]

extend ∷ V era t → Term era t → Subst era → Subst era

findV ∷ Subst era → V era t → Term era t

substTerm ∷ Subst era → Term era t → Term era t

substPred ∷ Subst era → Pred era → Pred era

substSum ∷ Subst era → Sum era t → Sum era t

substTarget ∷ Subst era → Target era t → Target era t

simplify ∷ Term era t → Typed t

Simplify Terms that only contain Literals (or constant) sub-Terms

simplifySum ∷ Sum era c → Typed c

Simplify constant Sum's

runTerm ∷ Env era → Term era t → Typed t

Fully evaluate a Term, looking up the variables in the Env.

runPred ∷ Env era → Pred era → Typed Bool

runComp ∷ Env era → s → AnyF era s → Typed Bool

termRep ∷ Term era t → Rep era t

runSum ∷ Env era → Sum era c → Typed c

makeTest ∷ Env era → Pred era → Typed (String, Bool, Pred era)