Test.Cardano.Ledger.Constrained.Classes

Documentation

gauss ∷ Floating a ⇒ a → a → a → a

class (Ord x, Show x) ⇒ Adds x where

Methods

add ∷ x → x → x

minus ∷ [String] → x → x → x

zero ∷ x

partition ∷ x → [String] → Int → x → Gen [x]

'partition 7 trace 4 235', generate a list of length 4 that adds up t0 235, where the smallest number is >= 7

genAdds ∷ [String] → AddsSpec x → Gen x

'genAdds trace spec' generate an x in the range specified by spec

fromI ∷ [String] → Int → x

Analogous to fromIntegral, translate an Int to an appropriate x

toI ∷ x → Int

translate an x Int to an appropriate Int

genSmall ∷ Gen Int

Used in testing to get appropriate smallest values to test 'partition smallest trace count total'. The generator should choose from several values appropriate for the type x. choose [0,1,2] would be appropriate for Natural, since there are no negative Natural numbers.

sumAdds ∷ (Foldable t, Adds c) ⇒ t c → c

projAdds ∷ (Foldable t, Sums a b) ⇒ t a → b

genFromAddsSpec ∷ [String] → AddsSpec c → Gen Int

genFromNonNegAddsSpec ∷ [String] → AddsSpec c → Gen Int

class (Show x, Adds x) ⇒ Sums t x | t → x where

Methods

getSum ∷ t → x

genT ∷ [String] → x → Gen t

genTxOutX ∷ Reflect era ⇒ Proof era → Coin → Gen (TxOutF era)

genValueX ∷ Reflect era ⇒ Proof era → Coin → Gen (ValueF era)

updateRew ∷ ∀ c. Coin → Reward c → Reward c

class Show t ⇒ Sizeable t where

Methods

getSize ∷ t → Int

extract the size of t

class Count t where

Methods

canFollow ∷ t → t → Bool

'canFollow x y', is x an appropriate successor to y

genPred ∷ t → Gen t

Generate the predecessor, given the successor

genSucc ∷ t → Gen t

Generate the successor, given the predecessor

data TxOutF era where

Constructors

TxOutF ∷ Proof era → TxOut era → TxOutF era

unTxOut ∷ TxOutF era → TxOut era

data ValueF era where

Constructors

ValueF ∷ Proof era → Value era → ValueF era

unValue ∷ ValueF era → Value era

data PParamsF era where

Constructors

PParamsF ∷ Proof era → PParams era → PParamsF era

unPParams ∷ PParamsF era → PParams era

pparamsWrapperL ∷ Lens' (PParamsF era) (PParams era)

data PParamsUpdateF era where

Constructors

PParamsUpdateF ∷ Proof era → PParamsUpdate era → PParamsUpdateF era

unPParamsUpdate ∷ PParamsUpdateF era → PParamsUpdate era

pparamsUpdateWrapperL ∷ Lens' (PParamsUpdateF era) (PParamsUpdate era)

data ProposedPPUpdatesF era where

Constructors

ProposedPPUpdatesF ∷ Proof era → ProposedPPUpdates era → ProposedPPUpdatesF era

unProposedPPUpdates ∷ ProposedPPUpdatesF era → ProposedPPUpdates era

proposedCoreL ∷ Lens' (ProposedPPUpdates era) (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era))

proposedWrapperL ∷ Lens' (ProposedPPUpdatesF era) (ProposedPPUpdates era)

coreMapL ∷ Proof era → Lens' (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdate era)) (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

proposedMapL ∷ Lens' (ProposedPPUpdatesF era) (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

data GovernanceState era

Constructors

GovernanceState (Proof era) (GovernanceState era)

unGovernanceState ∷ GovernanceState era → GovernanceState era

governanceProposedL ∷ Lens' (GovernanceState era) (ShelleyPPUPState era)

getPPUP ∷ ∀ era. Proof era → GovernanceState era → ShelleyPPUPState era

putPPUP ∷ ∀ era. Proof era → ShelleyPPUPState era → GovernanceState era

liftUTxO ∷ Map (TxIn (EraCrypto era)) (TxOutF era) → UTxO era

genValue ∷ Proof era → Gen (ValueF era)

genTxOut ∷ Proof era → Gen (TxOutF era)

genPParams ∷ Proof era → Gen (PParamsF era)

genPParamsUpdate ∷ Proof era → Gen (PParamsUpdateF era)

genProposedPPUpdates ∷ Proof era → Gen (ProposedPPUpdatesF era)

genGovernanceState ∷ Proof era → Gen (GovernanceState era)

genUTxO ∷ Proof era → Gen (UTxO era)

zeroCount ∷ Show a ⇒ [Char] → a → [Char]

legalCallPartition ∷ [String] → String → Integer → Int → Integer → Maybe [String]

integerPartition ∷ [String] → String → Integer → Int → Integer → Gen [Integer]

Generate a list of length size that sums to total, where the minimum element is (>= smallest)

partitionRational ∷ Rational → [String] → Int → Rational → Gen [Rational]

partitionCoin ∷ Coin → [String] → Int → Coin → Gen [Coin]

partitionDeltaCoin ∷ DeltaCoin → [String] → Int → DeltaCoin → Gen [DeltaCoin]

partitionInt ∷ Int → [String] → Int → Int → Gen [Int]

partitionWord64 ∷ Word64 → [String] → Int → Word64 → Gen [Word64]

partitionNatural ∷ Natural → [String] → Int → Natural → Gen [Natural]