Test.Cardano.Ledger.Constrained.Conway.Certs

Description

Specs necessary to generate, environment, state, and signal for the CERTS rule

Documentation

setMapMaybe ∷ Ord a ⇒ (t → Maybe a) → Set t → Set a

txZero ∷ EraTx era ⇒ Tx era

certsEnvSpec ∷ (EraSpecPParams era, HasSpec (Tx era)) ⇒ Specification (CertsEnv era)

projectEnv ∷ CertsEnv era → CertEnv era

Project a CertEnv out of a CertsEnv (i.e drop the Tx)

txCertsSpec ∷ ∀ era. EraSpecCert era ⇒ WitUniv era → CertsEnv era → CertState era → Specification (Seq (TxCert era))

noSameKeys ∷ ∀ era. EraSpecCert era ⇒ [TxCert era] → [TxCert era]

listSeqCertPairSpec ∷ ∀ era. EraSpecCert era ⇒ WitUniv era → CertEnv era → CertState era → Specification ([TxCert era], Seq (TxCert era))

Specify a pair of List and Seq, where they have essentially the same elements EXCEPT, the Seq has duplicate keys filtered out.