Test.Cardano.Ledger.Constrained.Conway.Deleg

Description

Specs necessary to generate, environment, state, and signal for the DELEG rule

Documentation

someZeros ∷ ∀ fn. IsConwayUniv fn ⇒ Specification fn RDPair

Specify that some of the rewards in the RDPair's are zero. without this in the DState, it is hard to generate the ConwayUnRegCert certificate, since it requires a rewards balance of 0.

dStateSpec ∷ ∀ fn era. (IsConwayUniv fn, EraSpecDeleg era) ⇒ Specification fn (DState era)

conwayDelegCertSpec ∷ ∀ fn era. (EraPParams era, IsConwayUniv fn) ⇒ ConwayDelegEnv era → CertState era → Specification fn (ConwayDelegCert (EraCrypto era))

delegEnvSpec ∷ (EraSpecPParams era, IsConwayUniv fn) ⇒ Specification fn (ConwayDelegEnv era)

shelleyDelegCertSpec ∷ ∀ fn era. (EraPParams era, IsConwayUniv fn) ⇒ ConwayDelegEnv era → DState era → Specification fn (ShelleyDelegCert (EraCrypto era))

class (Era era, EraPParams era) ⇒ EraSpecDeleg era where

Methods

hasGenDelegs ∷ proxy era → Bool