Test.Cardano.Ledger.Constrained.Conway.Deleg

Description

Specs necessary to generate, environment, state, and signal for the DELEG rule

Documentation

coinToWord64 ∷ Coin → Word64

wdrlCredentials ∷ Map RewardAccount Coin → Set (Credential 'Staking)

keyHashWdrl ∷ Map RewardAccount Coin → Set (Credential 'Staking)

isKeyHash ∷ Credential 'Staking → Bool

accountBalanceSpec ∷ (HasSpec a, Monoid a) ⇒ Term a → Pred

dRepMembershipPred ∷ Map (Credential 'DRepRole) a → Term DRep → Pred

witnessedKeyHashStakePoolMapSpec ∷ Era era ⇒ WitUniv era → Specification (Map (KeyHash StakePool) StakePoolState)

The DState needs a witnessed set of delegations to be usefull. Use this Spec to obtain a random one

conwayAccountsSpec ∷ Era era ⇒ WitUniv era → Term (Map (KeyHash 'StakePool) PoolParams) → Specification (ConwayAccounts era)

stakePoolDelegationsSpec ∷ Era era ⇒ WitUniv era → Specification (Map (Credential 'Staking) (KeyHash 'StakePool))

dRepDelegationsSpec ∷ Era era ⇒ WitUniv era → Specification (Map (Credential 'Staking) DRep)

dStateSpec ∷ (Era era, HasSpec (Accounts era)) ⇒ WitUniv era → Map RewardAccount Coin → Specification (DState era)

conwayDelegCertSpec ∷ ∀ era. (EraPParams era, ConwayEraCertState era) ⇒ ConwayDelegEnv era → CertState era → Specification ConwayDelegCert

delegEnvSpec ∷ EraSpecPParams era ⇒ Specification (ConwayDelegEnv era)

shelleyDelegCertSpec ∷ ∀ era. (EraPParams era, EraAccounts era) ⇒ WitUniv era → ConwayDelegEnv era → DState era → Specification ShelleyDelegCert

isZeroAccountBalance ∷ EraAccounts era ⇒ AccountState era → Bool

class (Era era, EraPParams era, HasSpec (Accounts era), IsNormalType (Accounts era)) ⇒ EraSpecDeleg era where

Methods

hasGenDelegs ∷ proxy era → Bool