Test.Cardano.Ledger.Constrained.Conway.GovCert

Description

Specs necessary to generate, environment, state, and signal for the GOVCERT rule

Documentation

vStateSpec ∷ ∀ era. Era era ⇒ WitUniv era → Set (Credential 'DRepRole) → Specification (VState era)

There are no hard constraints on VState, other than witnessing, and delegatee conformance which must align with the conwayCertExecContextSpec

govCertSpec ∷ EraCertState era ⇒ WitUniv era → ConwayGovCertEnv ConwayEra → CertState ConwayEra → Specification ConwayGovCert

notYetResigned ∷ (HasSpec x, Ord x, IsNormalType x) ⇒ Map x CommitteeAuthorization → Term x → Pred

Operations for authenticating a HotKey, or resigning a ColdKey are illegal if that key has already resigned.

govCertEnvSpec ∷ WitUniv ConwayEra → Specification (ConwayGovCertEnv ConwayEra)