Test.Cardano.Ledger.Constrained.Conway.Instances

Documentation

data SimplePParams era

Use this as the SimpleRep of (PParams era). It is polymorphic enough to encode PParams in EVERY Era. The EraPParams instances remove the fields that do not appear in that Era.

Constructors

SimplePParams

Fields

minFeeA ∷ Coin

minFeeB ∷ Coin

maxBBSize ∷ Word32

maxTxSize ∷ Word32

maxBHSize ∷ Word32

keyDeposit ∷ Coin

poolDeposit ∷ Coin

eMax ∷ EpochInterval

nOpt ∷ Word16

a0 ∷ NonNegativeInterval

rho ∷ UnitInterval

tau ∷ UnitInterval

decentral ∷ UnitInterval

protocolVersion ∷ ProtVer

minUTxOValue ∷ Coin

minPoolCost ∷ Coin

Alonzo

coinsPerUTxOWord ∷ Coin

costModels ∷ CostModels

prices ∷ Prices

maxTxExUnits ∷ ExUnits

maxBlockExUnits ∷ ExUnits

maxValSize ∷ Natural

collateralPercentage ∷ Natural

maxCollateralInputs ∷ Natural

Babbage

coinsPerUTxOByte ∷ Coin

Conway

poolVotingThresholds ∷ PoolVotingThresholds

drepVotingThresholds ∷ DRepVotingThresholds

committeeMinSize ∷ Natural

committeeMaxTermLength ∷ EpochInterval

govActionLifetime ∷ EpochInterval

govActionDeposit ∷ Coin

dRepDeposit ∷ Coin

dRepActivity ∷ EpochInterval

minFeeRefScriptCostPerByte ∷ NonNegativeInterval

data SimplePPUpdate

Use this as the SimpleRep of (PParamsUpdate era)

Constructors

SimplePPUpdate

Fields

uminFeeA ∷ StrictMaybe Coin

uminFeeB ∷ StrictMaybe Coin

umaxBBSize ∷ StrictMaybe Word32

umaxTxSize ∷ StrictMaybe Word32

umaxBHSize ∷ StrictMaybe Word32

ukeyDeposit ∷ StrictMaybe Coin

upoolDeposit ∷ StrictMaybe Coin

ueMax ∷ StrictMaybe EpochInterval

unOpt ∷ StrictMaybe Word16

ua0 ∷ StrictMaybe NonNegativeInterval

urho ∷ StrictMaybe UnitInterval

utau ∷ StrictMaybe UnitInterval

udecentral ∷ StrictMaybe UnitInterval

uprotocolVersion ∷ StrictMaybe ProtVer

uminUTxOValue ∷ StrictMaybe Coin

uminPoolCost ∷ StrictMaybe Coin

ucoinsPerUTxOWord ∷ StrictMaybe Coin

ucostModels ∷ StrictMaybe CostModels

uprices ∷ StrictMaybe Prices

umaxTxExUnits ∷ StrictMaybe ExUnits

umaxBlockExUnits ∷ StrictMaybe ExUnits

umaxValSize ∷ StrictMaybe Natural

ucollateralPercentage ∷ StrictMaybe Natural

umaxCollateralInputs ∷ StrictMaybe Natural

ucoinsPerUTxOByte ∷ StrictMaybe Coin

upoolVotingThresholds ∷ StrictMaybe PoolVotingThresholds

udrepVotingThresholds ∷ StrictMaybe DRepVotingThresholds

ucommitteeMinSize ∷ StrictMaybe Natural

ucommitteeMaxTermLength ∷ StrictMaybe EpochInterval

ugovActionLifetime ∷ StrictMaybe EpochInterval

ugovActionDeposit ∷ StrictMaybe Coin

udRepDeposit ∷ StrictMaybe Coin

udRepActivity ∷ StrictMaybe EpochInterval

uminFeeRefScriptCostPerByte ∷ StrictMaybe NonNegativeInterval

class (Eq (PParamsHKD Identity era), Show (PParamsHKD Identity era), ToExpr (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), EraPParams era, EraTxOut era, EraGov era, EraTx era) ⇒ EraSpecPParams era where

Methods

subsetToPP ∷ SimplePParams era → PParams era

ppToSubset ∷ PParams era → SimplePParams era

updateToPPU ∷ SimplePPUpdate → PParamsUpdate era

ppuToUpdate ∷ PParamsUpdate era → SimplePPUpdate

data StringW ∷ [Type] → Type → Type

type ProposalTree era = (StrictMaybe GovActionId, [Tree (GAS era)])

data ProposalsSplit

Constructors

ProposalsSplit

Fields

psPPChange ∷ Integer

psHFInitiation ∷ Integer

psUpdateCommittee ∷ Integer

psNewConstitution ∷ Integer

psOthers ∷ Integer

type ShelleyTxBodyTypes = '[Set TxIn, [TxOut ShelleyEra], [TxCert ShelleyEra], Map RewardAccount Coin, Coin, SlotNo, Maybe (Update ShelleyEra), Maybe TxAuxDataHash]

This is an abstraction of the Pattern ShelleyTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AllegraTxBodyTypes = '[Set TxIn, [TxOut AllegraEra], [TxCert AllegraEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update AllegraEra), Maybe TxAuxDataHash]

This is an abstraction of the Pattern AllegraTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type MaryTxBodyTypes = '[Set TxIn, [TxOut MaryEra], [TxCert MaryEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update MaryEra), Maybe TxAuxDataHash, MultiAsset]

This is an abstraction of the Pattern MaryTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AlonzoTxBodyTypes = '[Set TxIn, Set TxIn, [TxOut AlonzoEra], [TxCert AlonzoEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update AlonzoEra), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern AlonzoTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type BabbageTxBodyTypes = '[Set TxIn, Set TxIn, Set TxIn, [Sized (TxOut BabbageEra)], Maybe (Sized (TxOut BabbageEra)), Maybe Coin, [TxCert BabbageEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update BabbageEra), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern BabbageTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

cSNothing_ ∷ (HasSpec a, IsNormalType a) ⇒ Term (StrictMaybe a)

cSJust_ ∷ (HasSpec a, IsNormalType a) ⇒ Term a → Term (StrictMaybe a)

succV_ ∷ Term Version → Term Version

makePrices ∷ Integer → Integer → Prices

makeUnitInterval ∷ Integer → Integer → UnitInterval

makeNonNegativeInterval ∷ Integer → Integer → NonNegativeInterval

prettyE ∷ ToExpr x ⇒ x → Doc

onJust' ∷ (HasSpec a, IsNormalType a, IsPred p) ⇒ Term (StrictMaybe a) → (Term a → p) → Pred

onSized ∷ (HasSpec a, IsPred p) ⇒ Term (Sized a) → (Term a → p) → Pred

cKeyHashObj ∷ Typeable r ⇒ Term (KeyHash r) → Term (Credential r)

cScriptHashObj ∷ Typeable r ⇒ Term ScriptHash → Term (Credential r)

maryValueCoin_ ∷ Term MaryValue → Term Coin

strLen_ ∷ (HasSpec s, StringLike s) ⇒ Term s → Term Int

sizedValue_ ∷ (HasSpec (Sized a), HasSpec a) ⇒ Term (Sized a) → Term a

sizedSize_ ∷ (HasSpec (Sized a), HasSpec a) ⇒ Term (Sized a) → Term Int64

txOutVal_ ∷ (HasSpec (Value era), Era era, HasSpec (Data era), Val (Value era), HasSpec (Script era), HasSpec (BabbageTxOut era), IsNormalType (Script era)) ⇒ Term (BabbageTxOut era) → Term (Value era)

pProcDeposit_ ∷ Term (ProposalProcedure ConwayEra) → Term Coin

pProcGovAction_ ∷ Term (ProposalProcedure ConwayEra) → Term (GovAction ConwayEra)

gasId_ ∷ Term (GovActionState ConwayEra) → Term GovActionId

gasCommitteeVotes_ ∷ Term (GovActionState ConwayEra) → Term (Map (Credential 'HotCommitteeRole) Vote)

gasDRepVotes_ ∷ Term (GovActionState ConwayEra) → Term (Map (Credential 'DRepRole) Vote)

gasProposalProcedure_ ∷ Term (GovActionState ConwayEra) → Term (ProposalProcedure ConwayEra)

psPParamUpdate_ ∷ (EraSpecPParams era, Arbitrary (Proposals era)) ⇒ Term (Proposals era) → Term (ProposalTree era)

genProposalsSplit ∷ Integer → Gen ProposalsSplit

proposalSplitSum ∷ ProposalsSplit → Integer

coerce_ ∷ ∀ a b. (HasSpec a, HasSpec b, CoercibleLike a b) ⇒ Term a → Term b

toDelta_ ∷ Term Coin → Term DeltaCoin

simplePParamsSpec ∷ ∀ era. EraSpecPParams era ⇒ Specification (SimplePParams era)

A sample of how to constrain (PParams era) for every Era, by writing a specification for SimplePParams. Constrained but not applicable fields (for that era) are elided in the result. Missing fields are left unconstrained and will appear as random values in the result. This can easily be lifted to PParams: see Test.Cardano.Ledger.Constrained.Conway.PParams(pparamsSpec)

minFeeA_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

minFeeB_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

maxBBSize_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Word32

maxTxSize_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Word32

maxBHSize_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Word32

keyDeposit_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

poolDeposit_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

eMax_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term EpochInterval

nOpt_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Word16

a0_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term NonNegativeInterval

rho_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term UnitInterval

tau_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term UnitInterval

decentral_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term UnitInterval

protocolVersion_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term ProtVer

minUTxOValue_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

minPoolCost_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

coinsPerUTxOWord_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

costModels_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term CostModels

prices_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Prices

maxTxExUnits_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term ExUnits

maxBlockExUnits_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term ExUnits

maxValSize_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Natural

collateralPercentage_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Natural

maxCollateralInputs_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Natural

coinsPerUTxOByte_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

poolVotingThresholds_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term PoolVotingThresholds

drepVotingThresholds_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term DRepVotingThresholds

committeeMinSize_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Natural

committeeMaxTermLength_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term EpochInterval

govActionLifetime_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term EpochInterval

govActionDeposit_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

dRepDeposit_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term Coin

dRepActivity_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term EpochInterval

minFeeRefScriptCostPerByte_ ∷ EraSpecPParams era ⇒ Term (SimplePParams era) → Term NonNegativeInterval