Test.Cardano.Ledger.Constrained.Conway.Instances.TxBody

Documentation

type ShelleyTxBodyTypes = '[Set TxIn, [TxOut ShelleyEra], [TxCert ShelleyEra], Map RewardAccount Coin, Coin, SlotNo, Maybe (Update ShelleyEra), Maybe TxAuxDataHash]

This is an abstraction of the Pattern ShelleyTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AllegraTxBodyTypes = '[Set TxIn, [TxOut AllegraEra], [TxCert AllegraEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update AllegraEra), Maybe TxAuxDataHash]

This is an abstraction of the Pattern AllegraTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms bewtween the two, in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type MaryTxBodyTypes = '[Set TxIn, [TxOut MaryEra], [TxCert MaryEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update MaryEra), Maybe TxAuxDataHash, MultiAsset]

This is an abstraction of the Pattern MaryTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type AlonzoTxBodyTypes = '[Set TxIn, Set TxIn, [TxOut AlonzoEra], [TxCert AlonzoEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update AlonzoEra), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern AlonzoTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

type BabbageTxBodyTypes = '[Set TxIn, Set TxIn, Set TxIn, [Sized (TxOut BabbageEra)], Maybe (Sized (TxOut BabbageEra)), Maybe Coin, [TxCert BabbageEra], Map RewardAccount Coin, Coin, ValidityInterval, Maybe (Update BabbageEra), Set (KeyHash 'Witness), MultiAsset, Maybe ScriptIntegrityHash, Maybe TxAuxDataHash, Maybe Network]

This is an abstraction of the Pattern BabbageTxBody, that uses [x] instead of (StrictSeq x) and (Maybe x) instead of (StrictMaybe x). It transforms between the abstractions and the real types in the toSimpleRep and fromSimpleRep methods. This makes it much easier to write Specifications, because the Constrained packaage knows about Lists and Maybe.

Orphan instances

HasSpec (TxBody AllegraEra)

Instance details

Associated Types

type TypeSpec (TxBody AllegraEra)

type Prerequisites (TxBody AllegraEra)

Methods

emptySpec ∷ TypeSpec (TxBody AllegraEra)

combineSpec ∷ TypeSpec (TxBody AllegraEra) → TypeSpec (TxBody AllegraEra) → Specification (TxBody AllegraEra)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxBody AllegraEra) → GenT m (TxBody AllegraEra)

conformsTo ∷ TxBody AllegraEra → TypeSpec (TxBody AllegraEra) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxBody AllegraEra) → TxBody AllegraEra → [TxBody AllegraEra]

toPreds ∷ Term (TxBody AllegraEra) → TypeSpec (TxBody AllegraEra) → Pred

cardinalTypeSpec ∷ TypeSpec (TxBody AllegraEra) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxBody AllegraEra) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxBody AllegraEra) → BinaryShow

monadConformsTo ∷ TxBody AllegraEra → TypeSpec (TxBody AllegraEra) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxBody AllegraEra) → [TxBody AllegraEra] → Specification (TxBody AllegraEra)

guardTypeSpec ∷ [String] → TypeSpec (TxBody AllegraEra) → Specification (TxBody AllegraEra)

prerequisites ∷ Evidence (Prerequisites (TxBody AllegraEra))

HasSpec (TxBody AlonzoEra)

Instance details

Associated Types

type TypeSpec (TxBody AlonzoEra)

type Prerequisites (TxBody AlonzoEra)

Methods

emptySpec ∷ TypeSpec (TxBody AlonzoEra)

combineSpec ∷ TypeSpec (TxBody AlonzoEra) → TypeSpec (TxBody AlonzoEra) → Specification (TxBody AlonzoEra)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxBody AlonzoEra) → GenT m (TxBody AlonzoEra)

conformsTo ∷ TxBody AlonzoEra → TypeSpec (TxBody AlonzoEra) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxBody AlonzoEra) → TxBody AlonzoEra → [TxBody AlonzoEra]

toPreds ∷ Term (TxBody AlonzoEra) → TypeSpec (TxBody AlonzoEra) → Pred

cardinalTypeSpec ∷ TypeSpec (TxBody AlonzoEra) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxBody AlonzoEra) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxBody AlonzoEra) → BinaryShow

monadConformsTo ∷ TxBody AlonzoEra → TypeSpec (TxBody AlonzoEra) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxBody AlonzoEra) → [TxBody AlonzoEra] → Specification (TxBody AlonzoEra)

guardTypeSpec ∷ [String] → TypeSpec (TxBody AlonzoEra) → Specification (TxBody AlonzoEra)

prerequisites ∷ Evidence (Prerequisites (TxBody AlonzoEra))

HasSpec (TxBody BabbageEra)

Instance details

Associated Types

type TypeSpec (TxBody BabbageEra)

type Prerequisites (TxBody BabbageEra)

Methods

emptySpec ∷ TypeSpec (TxBody BabbageEra)

combineSpec ∷ TypeSpec (TxBody BabbageEra) → TypeSpec (TxBody BabbageEra) → Specification (TxBody BabbageEra)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxBody BabbageEra) → GenT m (TxBody BabbageEra)

conformsTo ∷ TxBody BabbageEra → TypeSpec (TxBody BabbageEra) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxBody BabbageEra) → TxBody BabbageEra → [TxBody BabbageEra]

toPreds ∷ Term (TxBody BabbageEra) → TypeSpec (TxBody BabbageEra) → Pred

cardinalTypeSpec ∷ TypeSpec (TxBody BabbageEra) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxBody BabbageEra) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxBody BabbageEra) → BinaryShow

monadConformsTo ∷ TxBody BabbageEra → TypeSpec (TxBody BabbageEra) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxBody BabbageEra) → [TxBody BabbageEra] → Specification (TxBody BabbageEra)

guardTypeSpec ∷ [String] → TypeSpec (TxBody BabbageEra) → Specification (TxBody BabbageEra)

prerequisites ∷ Evidence (Prerequisites (TxBody BabbageEra))

HasSpec (TxBody MaryEra)

Instance details

Associated Types

type TypeSpec (TxBody MaryEra)

type Prerequisites (TxBody MaryEra)

Methods

emptySpec ∷ TypeSpec (TxBody MaryEra)

combineSpec ∷ TypeSpec (TxBody MaryEra) → TypeSpec (TxBody MaryEra) → Specification (TxBody MaryEra)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxBody MaryEra) → GenT m (TxBody MaryEra)

conformsTo ∷ TxBody MaryEra → TypeSpec (TxBody MaryEra) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxBody MaryEra) → TxBody MaryEra → [TxBody MaryEra]

toPreds ∷ Term (TxBody MaryEra) → TypeSpec (TxBody MaryEra) → Pred

cardinalTypeSpec ∷ TypeSpec (TxBody MaryEra) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxBody MaryEra) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxBody MaryEra) → BinaryShow

monadConformsTo ∷ TxBody MaryEra → TypeSpec (TxBody MaryEra) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxBody MaryEra) → [TxBody MaryEra] → Specification (TxBody MaryEra)

guardTypeSpec ∷ [String] → TypeSpec (TxBody MaryEra) → Specification (TxBody MaryEra)

prerequisites ∷ Evidence (Prerequisites (TxBody MaryEra))

HasSpec (TxBody ShelleyEra)

Instance details

Associated Types

type TypeSpec (TxBody ShelleyEra)

type Prerequisites (TxBody ShelleyEra)

Methods

emptySpec ∷ TypeSpec (TxBody ShelleyEra)

combineSpec ∷ TypeSpec (TxBody ShelleyEra) → TypeSpec (TxBody ShelleyEra) → Specification (TxBody ShelleyEra)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (TxBody ShelleyEra) → GenT m (TxBody ShelleyEra)

conformsTo ∷ TxBody ShelleyEra → TypeSpec (TxBody ShelleyEra) → Bool

shrinkWithTypeSpec ∷ TypeSpec (TxBody ShelleyEra) → TxBody ShelleyEra → [TxBody ShelleyEra]

toPreds ∷ Term (TxBody ShelleyEra) → TypeSpec (TxBody ShelleyEra) → Pred

cardinalTypeSpec ∷ TypeSpec (TxBody ShelleyEra) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (TxBody ShelleyEra) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (TxBody ShelleyEra) → BinaryShow

monadConformsTo ∷ TxBody ShelleyEra → TypeSpec (TxBody ShelleyEra) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (TxBody ShelleyEra) → [TxBody ShelleyEra] → Specification (TxBody ShelleyEra)

guardTypeSpec ∷ [String] → TypeSpec (TxBody ShelleyEra) → Specification (TxBody ShelleyEra)

prerequisites ∷ Evidence (Prerequisites (TxBody ShelleyEra))

EraSpecPParams era ⇒ HasSpec (Update era)

Instance details

Associated Types

type TypeSpec (Update era)

type Prerequisites (Update era)

Methods

emptySpec ∷ TypeSpec (Update era)

combineSpec ∷ TypeSpec (Update era) → TypeSpec (Update era) → Specification (Update era)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Update era) → GenT m (Update era)

conformsTo ∷ Update era → TypeSpec (Update era) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Update era) → Update era → [Update era]

toPreds ∷ Term (Update era) → TypeSpec (Update era) → Pred

cardinalTypeSpec ∷ TypeSpec (Update era) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Update era) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Update era) → BinaryShow

monadConformsTo ∷ Update era → TypeSpec (Update era) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Update era) → [Update era] → Specification (Update era)

guardTypeSpec ∷ [String] → TypeSpec (Update era) → Specification (Update era)

prerequisites ∷ Evidence (Prerequisites (Update era))

HasSimpleRep (TxBody AllegraEra)

Instance details

Associated Types

type SimpleRep (TxBody AllegraEra)

type TheSop (TxBody AllegraEra) ∷ [Type]

Methods

toSimpleRep ∷ TxBody AllegraEra → SimpleRep (TxBody AllegraEra)

fromSimpleRep ∷ SimpleRep (TxBody AllegraEra) → TxBody AllegraEra

HasSimpleRep (TxBody AlonzoEra)

Instance details

Associated Types

type SimpleRep (TxBody AlonzoEra)

type TheSop (TxBody AlonzoEra) ∷ [Type]

Methods

toSimpleRep ∷ TxBody AlonzoEra → SimpleRep (TxBody AlonzoEra)

fromSimpleRep ∷ SimpleRep (TxBody AlonzoEra) → TxBody AlonzoEra

HasSimpleRep (TxBody BabbageEra)

Instance details

Associated Types

type SimpleRep (TxBody BabbageEra)

type TheSop (TxBody BabbageEra) ∷ [Type]

Methods

toSimpleRep ∷ TxBody BabbageEra → SimpleRep (TxBody BabbageEra)

fromSimpleRep ∷ SimpleRep (TxBody BabbageEra) → TxBody BabbageEra

HasSimpleRep (TxBody MaryEra)

Instance details

Associated Types

type SimpleRep (TxBody MaryEra)

type TheSop (TxBody MaryEra) ∷ [Type]

Methods

toSimpleRep ∷ TxBody MaryEra → SimpleRep (TxBody MaryEra)

fromSimpleRep ∷ SimpleRep (TxBody MaryEra) → TxBody MaryEra

HasSimpleRep (TxBody ShelleyEra)

Instance details

Associated Types

type SimpleRep (TxBody ShelleyEra)

type TheSop (TxBody ShelleyEra) ∷ [Type]

Methods

toSimpleRep ∷ TxBody ShelleyEra → SimpleRep (TxBody ShelleyEra)

fromSimpleRep ∷ SimpleRep (TxBody ShelleyEra) → TxBody ShelleyEra

Typeable era ⇒ HasSimpleRep (Update era)

Instance details

Associated Types

type SimpleRep (Update era)

type TheSop (Update era) ∷ [Type]

Methods

toSimpleRep ∷ Update era → SimpleRep (Update era)

fromSimpleRep ∷ SimpleRep (Update era) → Update era