Test.Cardano.Ledger.Constrained.Conway.TxBodySpec

Documentation

subMap ∷ (Ord k, IsNormalType v, IsNormalType k, HasSpec k, HasSpec v) ⇒ Specification (Map k v, Map k v)

subMapSubDependsOnSuper ∷ (Ord k, IsNormalType v, IsNormalType k, HasSpec k, HasSpec v) ⇒ Term (Map k v) → Term (Map k v) → Pred

subMapSuperDependsOnSub ∷ (Ord k, IsNormalType v, IsNormalType k, HasSpec k, HasSpec v) ⇒ Term (Map k v) → Term (Map k v) → Pred

sumTxOut_ ∷ ∀ era. EraSpecTxOut era ⇒ Term [TxOut era] → Term Coin

sumCoin_ ∷ Term [Coin] → Term Coin

adjustTxOutCoin ∷ EraTxOut era ⇒ DeltaCoin → TxOut era → TxOut era

getDepositRefund ∷ ∀ era. (EraTxCert era, ConwayEraCertState era) ⇒ PParams era → CertState era → [TxCert era] → (DeltaCoin, DeltaCoin)

Extract the total deposits and refunds from a list of TxCerts. This a kind of AdaPot relative to the Certs in a Transaction body It depends on the PParams (deposit ammounts for registering a staking key, a ppol, and registering a Drep) and on the CertState (what deposits were made in the past)

putPretty ∷ ToExpr t ⇒ [Char] → t → IO ()