Test.Cardano.Ledger.Constrained.Env

Description

Provides variables (V era t), and mappings of them to objects of type t

Documentation

data V era t where

A proto variable. May or may not contain a Lens (encoded as Access)

Constructors

V ∷ String → Rep era t → Access era s t → V era t

data Field era s t where

Fields are like V, except they expose the type of the Lens

Constructors

Field ∷ Eq t ⇒ String → Rep era t → Access era s t → Field era s t

FConst ∷ Eq t ⇒ Rep era t → t → Access era s t → Field era s t

data AnyF era s where

Hide the type of t in a Field.

Constructors

AnyF ∷ (Show t, Eq t) ⇒ Field era s t → AnyF era s

vToField ∷ Eq t ⇒ Rep era s → V era t → Typed (Field era s t)

fieldToV ∷ Field era s t → Typed (V era t)

newtype Env era

Constructors

Env (Map String (Payload era))

data Payload era where

Constructors

Payload ∷ Rep era t → t → Access era s t → Payload era

emptyEnv ∷ Env era

findVar ∷ V era t → Env era → Typed t

storeVar ∷ V era t → t → Env era → Env era

data P era where

Constructors

P ∷ V era t → t → P era

bulkStore ∷ [P era] → Env era → Env era

data Name era where

An existentially quantified (V era t), hiding the t Usefull because unlike (V era t), it has both Eq and Ord instances

Constructors

Name ∷ V era t → Name era

data Access era s t where

Constructors

Yes ∷ Rep era s → Lens' s t → Access era s t

No ∷ Access era s t