Test.Cardano.Ledger.Constrained.Preds.Certs

Documentation

sRegKey ∷ ∀ era. Typeable era ⇒ RootTarget era (ShelleyTxCert era) (StakeCredential (EraCrypto era) → ShelleyTxCert era)

sUnRegKey ∷ ∀ era. Typeable era ⇒ RootTarget era (ShelleyTxCert era) (StakeCredential (EraCrypto era) → ShelleyTxCert era)

sDelegStake ∷ ∀ era. Typeable era ⇒ RootTarget era (ShelleyTxCert era) (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ShelleyTxCert era)

sRegPool ∷ Target era (PoolParams (EraCrypto era) → ShelleyTxCert era)

sRetirePool ∷ Target era (KeyHash 'StakePool (EraCrypto era) → EpochNo → ShelleyTxCert era)

sMirShift ∷ ∀ era. Era era ⇒ RootTarget era (ShelleyTxCert era) (Coin → MIRPot → Coin → ShelleyTxCert era)

sGovern ∷ Target era (KeyHash 'Genesis (EraCrypto era) → KeyHash 'GenesisDelegate (EraCrypto era) → Hash (HASH (EraCrypto era)) (VerKeyVRF (VRF (EraCrypto era))) → ShelleyTxCert era)

cRegKey ∷ Target era (StakeCredential (EraCrypto era) → Maybe Coin → ConwayTxCert era)

cUnRegKey ∷ Target era (StakeCredential (EraCrypto era) → Maybe Coin → ConwayTxCert era)

cDelegStake ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → ConwayTxCert era)

cDelegVote ∷ Target era (StakeCredential (EraCrypto era) → DRep (EraCrypto era) → a → ConwayTxCert era)

cDelegStakeVote ∷ Target era (StakeCredential (EraCrypto era) → KeyHash 'StakePool (EraCrypto era) → DRep (EraCrypto era) → a → ConwayTxCert era)

cRegDeleg ∷ ∀ era. ConwayEraTxCert era ⇒ Target era (StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → Coin → TxCert era)

cDelegateeStake ∷ ∀ era. Era era ⇒ RootTarget era (Delegatee (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era) → Delegatee (EraCrypto era))

cDelegateeVote ∷ ∀ era. Era era ⇒ RootTarget era (Delegatee (EraCrypto era)) (DRep (EraCrypto era) → Delegatee (EraCrypto era))

cDelegateeStakeVote ∷ ∀ era. Era era ⇒ RootTarget era (Delegatee (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era) → DRep (EraCrypto era) → Delegatee (EraCrypto era))

cRegPool ∷ Target era (PoolParams (EraCrypto era) → ConwayTxCert era)

cRetirePool ∷ Target era (KeyHash 'StakePool (EraCrypto era) → EpochNo → ConwayTxCert era)

partBfromPartA ∷ Ord k ⇒ Proof era → Map k Coin → Map k DeltaCoin

Transform some SubMap of instanReserves (or instanTreasury) into a partB map with invariants: (partB ! key) + (instanReserves ! key) >= (Coin 0) Note in Era before Alonzo, Negative transfers are not allowed.

makeDRepPred ∷ ∀ era. Era era ⇒ Term era (DRep (EraCrypto era)) → Term era (Credential 'DRepRole (EraCrypto era)) → Pred era

A user defined Predicate that Binds drep to a random DRep The parameter vote should be existentially bound in the surrounding context (inside a Choose Target perhaps)

minusCoinDeltaCoin ∷ Coin → DeltaCoin → DeltaCoin

availableForDistrC ∷ DeltaCoin → MIRPot → AccountState → InstantaneousRewards c → DeltaCoin

txCertMir ∷ ∀ era any. Era era ⇒ RootTarget era (ShelleyTxCert era) (MIRPot → Map (Credential 'Staking (EraCrypto era)) DeltaCoin → any → ShelleyTxCert era)

certsPreds ∷ ∀ era. Reflect era ⇒ UnivSize → Proof era → [Pred era]

certsStage ∷ Reflect era ⇒ UnivSize → Proof era → Subst era → Gen (Subst era)

demo ∷ ReplMode → Int → IO ()

demoTest ∷ TestTree

main ∷ Int → IO ()

sMaybeL ∷ Lens' (StrictMaybe a) (Maybe a)

maybeSL ∷ Lens' (Maybe a) (StrictMaybe a)

poolMetaL ∷ Lens' (PoolParams era) (StrictMaybe PoolMetadata)

poolMetadata ∷ Era era ⇒ Proof era → Term era (Maybe PoolMetadata)