Test.Cardano.Ledger.Constrained.Preds.Universes

Documentation

data UnivSize

Constructors

UnivSize

Fields

usNumTxOuts ∷ Int

usMaxAssets ∷ Int

usMaxPolicyID ∷ Int

usNumMultiAsset ∷ Int

usNumPtr ∷ Int

usNumAddr ∷ Int

usNumKeys ∷ Int

usNumPools ∷ Int

usNumStakeKeys ∷ Int

usNumGenesisKeys ∷ Int

usNumVoteKeys ∷ Int

usNumCredentials ∷ Int

usNumDatums ∷ Int

usNumTxIn ∷ Int

usNumPreUtxo ∷ Int

usNumColUtxo ∷ Int

usNumDReps ∷ Int

usMinCerts ∷ Int

usMaxCerts ∷ Int

usDatumFreq ∷ Int

usGenerateWithdrawals ∷ Bool

usMinInputs ∷ Int

usMaxInputs ∷ Int

usMinCollaterals ∷ Int

usMaxCollaterals ∷ Int

usRegKeyFreq ∷ Int

usUnRegKeyFreq ∷ Int

usAllowReRegisterPool ∷ Bool

usSpendScriptFreq ∷ Int

usCredScriptFreq ∷ Int

variedCoin ∷ Gen Coin

noZeroCoin ∷ Gen Coin

genAddrPair ∷ Network → Gen (BootstrapAddress c, SigningKey)

Generate a pair, A Byron address, and the key that can sign it.

genByronUniv ∷ Crypto c ⇒ Network → Gen (Map (KeyHash 'Payment c) (Addr c, SigningKey))

Generate a Map, that maps the Hash of a Byron address to a pair of the actual Byron address and the key that can sign it.

bootWitness ∷ (Crypto c, DSIGN c ~ Ed25519DSIGN) ⇒ Hash c EraIndependentTxBody → [BootstrapAddress c] → Map (KeyHash 'Payment c) (Addr c, SigningKey) → Set (BootstrapWitness c)

Given a list of Byron addresses, compute BootStrap witnesses of all of those addresses Can only be used with StandardCrypto

genDatums ∷ Era era ⇒ UnivSize → Int → Map (DataHash (EraCrypto era)) (Data era) → Gen [Datum era]

The universe of non-empty Datums. i.e. There are no NoDatum Datums in this list

genDatum ∷ Era era ⇒ UnivSize → Map (DataHash (EraCrypto era)) (Data era) → Gen (Datum era)

Only generate non-empty Datums. I.e. There are no NoDatum Datums generated.

genTxOut ∷ Reflect era ⇒ UnivSize → (Coin → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (Value era)) → Proof era → Coin → Set (Addr (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (DataHash (EraCrypto era)) (Data era) → Gen (TxOut era)

needsDatum ∷ EraScript era ⇒ Credential 'Payment (EraCrypto era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Bool

genTxOuts ∷ Reflect era ⇒ UnivSize → (Coin → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (Value era)) → Proof era → Int → Set (Addr (EraCrypto era)) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Map (DataHash (EraCrypto era)) (Data era) → Gen [TxOutF era]

genMultiAssetTriple ∷ Map (ScriptHash (EraCrypto era)) (ScriptF era) → Set AssetName → Gen Integer → Gen (PolicyID (EraCrypto era), AssetName, Integer)

pick1 ∷ [String] → Set t → Gen t

makeHashScriptMap ∷ Reflect era ⇒ Proof era → Int → PlutusPurposeTag → Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era)) → ValidityInterval → Gen (Map (ScriptHash (EraCrypto era)) (ScriptF era))

genDataWits ∷ Era era ⇒ Proof era → Int → Gen (Map (DataHash (EraCrypto era)) (Data era))

genAddrWith ∷ Proof era → Network → Set (Credential 'Payment (EraCrypto era)) → Set Ptr → Set (Credential 'Staking (EraCrypto era)) → Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey) → Gen (Addr (EraCrypto era))

genPtr ∷ SlotNo → Gen Ptr

genStakeRefWith ∷ ∀ era. Proof era → Set Ptr → Set (Credential 'Staking (EraCrypto era)) → Gen (StakeReference (EraCrypto era))

noScripts ∷ Proof era → Addr (EraCrypto era) → Bool

genDReps ∷ Set (Credential 'Staking c) → Gen [DRep c]

Make some candidate DReps. The 'Always...' and one from each Credential.

genDRepsT ∷ UnivSize → Term era (Set (Credential 'Staking (EraCrypto era))) → Target era (Gen (Set (DRep (EraCrypto era))))

txOutT ∷ Reflect era ⇒ Proof era → Addr (EraCrypto era) → Coin → TxOutF era

colTxOutT ∷ EraTxOut era ⇒ Proof era → Set (Addr (EraCrypto era)) → Gen (TxOutF era)

The collateral consists only of VKey addresses and the collateral outputs in the UTxO do not contain any non-ADA part

colTxOutSetT ∷ EraTxOut era ⇒ Proof era → Set (Addr (EraCrypto era)) → Gen (Set (TxOutF era))

The collateral consists only of VKey addresses and the collateral outputs in the UTxO do not contain any non-ADA part

scriptHashObjT ∷ Term era (ScriptHash (EraCrypto era)) → Target era (Credential k (EraCrypto era))

keyHashObjT ∷ Term era (KeyHash 'Witness (EraCrypto era)) → Target era (Credential k (EraCrypto era))

makeValidityT ∷ Term era SlotNo → Term era SlotNo → Term era SlotNo → Target era ValidityInterval

ptrUnivT ∷ Int → Term era SlotNo → Target era (Gen (Set Ptr))

addrUnivT ∷ Proof era → Int → Term era Network → Term era (Set (Credential 'Payment (EraCrypto era))) → Term era (Set Ptr) → Term era (Set (Credential 'Staking (EraCrypto era))) → Term era (Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey)) → Target era (Gen (Set (Addr (EraCrypto era))))

makeHashScriptMapT ∷ Proof era → Int → PlutusPurposeTag → Term era (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era))) → Term era ValidityInterval → Target era (Gen (Map (ScriptHash (EraCrypto era)) (ScriptF era)))

cast ∷ ∀ c k. Set (KeyHash 'Witness c) → Set (KeyHash k c)

castCredCold ∷ Set (KeyHash 'Witness c) → Set (Credential 'ColdCommitteeRole c)

castCredHot ∷ Set (KeyHash 'Witness c) → Set (Credential 'HotCommitteeRole c)

txinToGovactionId ∷ TxIn c → GovActionId c

universePreds ∷ Reflect era ⇒ UnivSize → Proof era → [Pred era]

multiAsset ∷ UnivSize → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (MultiAsset (EraCrypto era))

genValueF ∷ UnivSize → Proof era → Coin → Map (ScriptHash (EraCrypto era)) (ScriptF era) → Gen (Value era)

stakeToDRepRole ∷ Credential 'Staking c → Credential 'DRepRole c

stakeToHotCommittee ∷ Credential 'Staking c → Credential 'HotCommitteeRole c

stakeToColdCommittee ∷ Credential 'Staking c → Credential 'ColdCommitteeRole c

solveUniv ∷ Reflect era ⇒ UnivSize → Proof era → Gen (Subst era)

universeStage ∷ Reflect era ⇒ UnivSize → Proof era → Subst era → Gen (Subst era)

demo ∷ ReplMode → IO ()

demoTest ∷ TestTree

main ∷ IO ()