Test.Cardano.Ledger.Constrained.Scripts

Documentation

allPlutusScripts ∷ Reflect era ⇒ Proof era → Map (ScriptHash (EraCrypto era)) (IsValid, Script era)

For any given Era, there are only a finite number of Plutus scripts. This function computes all of them. There will be two failing scripts One for the Spend Tag, and another for all other Tags (Mint, Cert, Rewrd). The non-failing Spend scripts have varying number of arguments (0, 1, 2, 3) The non-failing (Mint Cert Rewrd) scripts are identical. Any Plutus script generated by genCoreScript will be in this Map.

genCoreScript ∷ ∀ era. Proof era → PlutusPurposeTag → KeyMap era → ValidityInterval → Gen (Script era)

spendPlutusScripts ∷ Reflect era ⇒ Proof era → Map (ScriptHash (EraCrypto era)) (IsValid, Script era)

There are only 5 plutus scripts that can be used in a Spend context

sufficientScript ∷ Proof era → Script era → Set (KeyHash 'Witness (EraCrypto era))