Test.Cardano.Ledger.Constrained.Size

Documentation

data Size

Constructors

SzNever [String]

SzAny

SzLeast Int

SzMost Int

SzRng Int Int
Size is in the range from i to j inclusive: SzRng i j = [i .. j]. Invariant i <= j

Bundled Patterns

pattern SzExact ∷ Int → Size

data AddsSpec c where

A specification of summation. like: lhs = ∑ rhs The idea is that the rhs can contain multiple terms: rhs = ∑ r1 + r2 + r3 Other example conditions: (lhs rhs), and (lhs= ∑ rhs) The invariant is that only a single variable appears in the summation. It can appear on either side. If it appears in the rhs then there may be other, constant terms, in the rhs: 7 = ∑ 3 + v + 9 We always do the sums and solving at type Int, and cast back and forth to accommodate other types with (Adds c) instances, using the methods 'fromI" and toI This allows the instance to deal with special conditions. There are two (non-failure) possibilities 1) Var on the left, 2) Var on the right We supply functions vLeft :: String -> OrdCond -> Integer -> AddsSpec c SumsTo _ x 4 + 6 + 9 === (vLeft x LTE 19) == (AddsSpecSize x (AtMost 19)) vRight :: Integer -> OrdCond -> Integer -> String -> AddsSpec c SumsTo _ 8 + x + 3 === (vRight 8 LTH 5 x) == (AddsSpecSize x (AtLeast 4)) But internally we store the information as a String and a Size (I.e. a range of Int)

Constructors

AddsSpecSize

Fields

∷ String

name

→ Size

total (range like (4 .. 12))

→ AddsSpec c

AddsSpecAny ∷ AddsSpec c

AddsSpecNever ∷ [String] → AddsSpec c

vLeft ∷ String → OrdCond → Int → AddsSpec c

vRight ∷ Int → OrdCond → Int → String → AddsSpec c

vLeftSize ∷ String → OrdCond → Int → Size

vRightSize ∷ Int → OrdCond → Int → String → Size

data OrdCond

First order representation of the Ord comparisons

Constructors

EQL

LTH

LTE

GTH

GTE

reverseOrdCond ∷ OrdCond → OrdCond

This function reverseOrdCond has been defined to handle the Pred SumsTo when the variable is on the right-hand-side (rhs) of the OrdCond operator. In order to do that we must multiply both sides of the inequality by (-1). For example consider [SumsTo (DeltaCoin 1) ▵₳ -2 > ∑ ▵₳ -1 + x] Note variable x on the rhs ^ To solve we subtract x from both sides, and add '▵₳ -2' from bothsides getting (-x) > ∑ (▵₳ -1) + (▵₳ -2) reduced to (-x) > ∑ (▵₳ -3) to solve we must multiply both sides by (-1) x ?? ∑ (▵₳ 3) What operator do we replace ?? by to make the original (▵₳ -2 > ∑ ▵₳ -1 + x) True? The change in the operator is called "reversing" the operator. See https://www.mathsisfun.com/algebra/inequality-solving.html for one explantion.

seps ∷ [String] → String

sepsP ∷ [String] → String

sepn ∷ [String] → String

runOrdCond ∷ Ord c ⇒ OrdCond → c → c → Bool

runSize ∷ Int → Size → Bool

atLeastDelta ∷ Int

Used in tests so things don't get too large

atMostAny ∷ Int

Used in tests so things don't get too large If we can't find an era using things of size 10 using things of size 100, isn't going to help.

genFromSize ∷ Size → Gen Int

Use to generate real sizes, where there are no negative numbers, and the smallest possible size is 0. Use this only where you know it is NOT SzNever

genFromIntRange ∷ Size → Gen Int

Similar to genFromSize, but allows negative numbers (unlike size where the smallest Int is 0)

genFromNonNegIntRange ∷ Size → Gen Int

vLeftNeg ∷ String → OrdCond → Int → AddsSpec c

vRightNeg ∷ Int → OrdCond → Int → String → AddsSpec c

negateSize ∷ Size → Size