Test.Cardano.Ledger.Constrained.TypeRep

Documentation

data Rep era t where

Constructors

RationalR ∷ Rep era Rational

CoinR ∷ Rep era Coin

EpochR ∷ Rep era EpochNo

(:->) ∷ Rep era a → Rep era b → Rep era (a → b) infixr 0

MapR ∷ Ord a ⇒ Rep era a → Rep era b → Rep era (Map a b)

SetR ∷ Ord a ⇒ Rep era a → Rep era (Set a)

ListR ∷ Rep era a → Rep era [a]

CredR ∷ Rep era (Credential 'Staking (EraCrypto era))

VCredR ∷ Rep era (Credential 'Voting (EraCrypto era))

PoolHashR ∷ Rep era (KeyHash 'StakePool (EraCrypto era))

WitHashR ∷ Rep era (KeyHash 'Witness (EraCrypto era))

GenHashR ∷ Rep era (KeyHash 'Genesis (EraCrypto era))

GenDelegHashR ∷ Rep era (KeyHash 'GenesisDelegate (EraCrypto era))

VHashR ∷ Rep era (KeyHash 'Voting (EraCrypto era))

PoolParamsR ∷ Rep era (PoolParams (EraCrypto era))

NewEpochStateR ∷ Rep era (NewEpochState era)

IntR ∷ Rep era Int

FloatR ∷ Rep era Float

NaturalR ∷ Rep era Natural

Word64R ∷ Rep era Word64

TxInR ∷ Rep era (TxIn (EraCrypto era))

StringR ∷ Rep era String

UnitR ∷ Rep era ()

PairR ∷ Rep era a → Rep era b → Rep era (a, b)

ProtVerR ∷ Proof era → Rep era ProtVer

ValueR ∷ Proof era → Rep era (ValueF era)

UTxOR ∷ Proof era → Rep era (UTxO era)

TxOutR ∷ Proof era → Rep era (TxOutF era)

PParamsR ∷ Proof era → Rep era (PParamsF era)

PParamsUpdateR ∷ Proof era → Rep era (PParamsUpdateF era)

DeltaCoinR ∷ Rep era DeltaCoin

GenDelegPairR ∷ Rep era (GenDelegPair (EraCrypto era))

FutureGenDelegR ∷ Rep era (FutureGenDeleg (EraCrypto era))

PPUPStateR ∷ Proof era → Rep era (ShelleyPPUPState era)

PtrR ∷ Rep era Ptr

IPoolStakeR ∷ Rep era (IndividualPoolStake (EraCrypto era))

SnapShotsR ∷ Rep era (SnapShots (EraCrypto era))

RewardR ∷ Rep era (Reward (EraCrypto era))

MaybeR ∷ Rep era t → Rep era (Maybe t)

SlotNoR ∷ Rep era SlotNo

SizeR ∷ Rep era Size

data (a ∷ k) :~: (b ∷ k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl ∷ ∀ {k} (a ∷ k). a :~: a

data Shape (rep ∷ Type → Type) where

Constructors

Nullary ∷ ∀ (rep ∷ Type → Type). Int → Shape rep

Nary ∷ ∀ (rep ∷ Type → Type). Int → [Shape rep] → Shape rep

Esc ∷ ∀ (rep ∷ Type → Type) t. (Singleton rep, Ord t) ⇒ rep t → t → Shape rep

class Singleton (t ∷ k → Type) where

Given (Singleton T), a value of type (T i) has exactly 0 or 1 inhabitants, so we can compare the structure of the type to get proofs that the indexes (i and j) are the same type, using testEql, at runtime. cmpIndex, it is like compare except we can have two different indexes (a and b).

Methods

testEql ∷ ∀ (i ∷ k) (j ∷ k). t i → t j → Maybe (i :~: j)

cmpIndex ∷ ∀ (a ∷ k) (b ∷ k). t a → t b → Ordering

type Eql (x ∷ k) (y ∷ k) = x :~: y

Type synonym, so we can use ( :~: ) without TypeOperators

synopsis ∷ ∀ e t. Rep e t → t → String

compareRep ∷ ∀ era t s. Rep era t → Rep era s → Ordering

genSizedRep ∷ Era era ⇒ Int → Rep era t → Gen t

genRep ∷ Era era ⇒ Rep era b → Gen b

data TxOutF era where

Constructors

TxOutF ∷ Proof era → TxOut era → TxOutF era

unTxOut ∷ TxOutF era → TxOut era

data ValueF era where

Constructors

ValueF ∷ Proof era → Value era → ValueF era

unValue ∷ ValueF era → Value era

data PParamsF era where

Constructors

PParamsF ∷ Proof era → PParams era → PParamsF era

unPParams ∷ PParamsF era → PParams era

data PParamsUpdateF era where

Constructors

PParamsUpdateF ∷ Proof era → PParamsUpdate era → PParamsUpdateF era

unPParamsUpdate ∷ PParamsUpdateF era → PParamsUpdate era

liftUTxO ∷ Map (TxIn (EraCrypto era)) (TxOutF era) → UTxO era

data Proof era where

Proof of a valid (predefined) era

Constructors

Shelley ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (ShelleyEra c)

Mary ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (MaryEra c)

Allegra ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (AllegraEra c)

Alonzo ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (AlonzoEra c)

Babbage ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (BabbageEra c)

Conway ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (ConwayEra c)

data Evidence c where

Constructors

Standard ∷ Evidence Standard

Mock ∷ Evidence Mock