Test.Cardano.Ledger.Constrained.Vars

Documentation

field ∷ Era era ⇒ Rep era s → Term era t → AnyF era s

Used in Component constraints to turn a Var Term into a component (AnyF era s) E.g. (Component foo [ field fooRep fooPart1, field fooRep fooPart2]) Where fooPart1 :: Term era a, and fooPart2 :: Term era b And fooPart1 has an (Access foo a) And fooPart2 has an (Access foo b)

getName ∷ Term era t → Name era

type NELens era t = Lens' (NewEpochState era) t

currentEpoch ∷ Era era ⇒ Term era EpochNo

prevBlocksMade ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

currBlocksMade ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

poolDistr ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

mockPoolDistr ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Rational)

For tests only, Like PoolDistr but has a Rational (rather than a IndividualPoolStake).

poolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

rewards ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

rewardsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

delegations ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

delegationsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

stakeDeposits ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

stakeDepositsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

ptrs ∷ Era era ⇒ Term era (Map Ptr (Credential 'Staking (EraCrypto era)))

ptrsL ∷ NELens era (Map Ptr (Credential 'Staking (EraCrypto era)))

currentDRepState ∷ Era era ⇒ Term era (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

drepsL ∷ NELens era (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

drepDelegation ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)))

drepDelegationL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)))

futureGenDelegs ∷ Era era ⇒ Term era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

futureGenDelegsL ∷ NELens era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegs ∷ Era era ⇒ Term era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegsL ∷ NELens era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

instanReserves ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanReservesL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanReservesSum ∷ Era era ⇒ Term era Coin

instanTreasury ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasuryL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasurySum ∷ Era era ⇒ Term era Coin

deltaReserves ∷ Era era ⇒ Term era DeltaCoin

deltaReservesNEL ∷ NELens era DeltaCoin

deltaTreasury ∷ Era era ⇒ Term era DeltaCoin

deltaTreasuryNEL ∷ NELens era DeltaCoin

regPools ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

regPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPools ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

retiring ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

retiringL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

poolDeposits ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

poolDepositsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

committeeState ∷ Era era ⇒ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)))

committeeStateL ∷ NELens era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)))

numDormantEpochs ∷ Era era ⇒ Term era EpochNo

numDormantEpochsL ∷ NELens era EpochNo

utxo ∷ Era era ⇒ Proof era → Term era (Map (TxIn (EraCrypto era)) (TxOutF era))

utxoL ∷ Proof era → NELens era (Map (TxIn (EraCrypto era)) (TxOutF era))

unUtxoL ∷ Proof era → Lens' (UTxO era) (Map (TxIn (EraCrypto era)) (TxOutF era))

deposits ∷ Era era ⇒ Term era Coin

depositsL ∷ NELens era Coin

fees ∷ Era era ⇒ Term era Coin

feesL ∷ NELens era Coin

donation ∷ Era era ⇒ Term era Coin

donationL ∷ NELens era Coin

ppup ∷ Era era ⇒ Proof era → Term era (ShelleyGovState era)

ppupsL ∷ Proof era → NELens era (ShelleyGovState era)

pparamProposals ∷ Era era ⇒ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

futurePParamProposals ∷ Era era ⇒ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

currPParams ∷ Era era ⇒ Proof era → Term era (PParamsF era)

futurePParams ∷ Era era ⇒ Proof era → Term era (FuturePParams era)

prevPParams ∷ EraGov era ⇒ Proof era → Term era (PParamsF era)

ppupStateT ∷ ∀ era. (GovState era ~ ShelleyGovState era, EraGov era) ⇒ Proof era → RootTarget era (ShelleyGovState era) (ShelleyGovState era)

govL ∷ Lens' (GovState era) (GovState era)

govStateT ∷ ∀ era. Era era ⇒ Proof era → RootTarget era (GovState era) (GovState era)

individualPoolStakeL ∷ Lens' (IndividualPoolStake c) Rational

isPtrMapT ∷ Era era ⇒ Term era (Map Ptr Coin)

ptrMapL ∷ Lens' (NewEpochState era) (Map Ptr Coin)

isCredMapT ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

credMapL ∷ Lens' (NewEpochState era) (Map (Credential 'Staking (EraCrypto era)) Coin)

incrementalStake ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

This variable is computed from the UTxO and the PParams, It represents the incremental stake that is computed by smartUTxO in the UTxOState Target UTxOStateT The domain of this map is the complete set of credentials used to delegate Coin in the TxOuts in the UTxO.

incrementalStakeT ∷ Reflect era ⇒ Proof era → Target era (Map (Credential 'Staking (EraCrypto era)) Coin)

treasury ∷ Era era ⇒ Term era Coin

treasuryL ∷ NELens era Coin

reserves ∷ Era era ⇒ Term era Coin

reservesL ∷ NELens era Coin

mirAvailTreasury ∷ Era era ⇒ Term era Coin

The Coin availabe for a MIR transfer to/from the Treasury Computed from treasury + deltaTreasury - sum(instanTreasury)

mirAvailReserves ∷ Era era ⇒ Term era Coin

The Coin availabe for a MIR transfer to/from the Reserves Computed from reserves + deltaReserves - sum(instanReserves)

snapshots ∷ Era era ⇒ Term era (SnapShots (EraCrypto era))

snapshotsL ∷ NELens era (SnapShots (EraCrypto era))

ppFL ∷ Proof era → Lens' (PParams era) (PParamsF era)

Lens' from the Core PParams to the Model PParamsF which embeds a (Proof era)

pparamsVar ∷ EraGov era ⇒ Proof era → V era (PParamsF era)

pparams ∷ EraGov era ⇒ Proof era → Term era (PParamsF era)

nmLikelihoodsT ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) [Float])

nmRewardPotT ∷ Era era ⇒ Term era Coin

stakeL ∷ Lens' (Stake c) (Map (Credential 'Staking c) Coin)

Helper lens that deals with the Stake newtype, and the shift from Map to VMap

vmapL ∷ Lens' (VMap VB VB k v) (Map k v)

Helper lens that deals with the shift from Map to VMap

markStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

markStake ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

markDelegs ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markPools ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markSnapShotT ∷ ∀ era. Era era ⇒ RootTarget era (SnapShot (EraCrypto era)) (SnapShot (EraCrypto era))

setStake ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

setStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

setDelegs ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setPools ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setSnapShotT ∷ ∀ era. Era era ⇒ RootTarget era (SnapShot (EraCrypto era)) (SnapShot (EraCrypto era))

goStake ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

goStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

goDelegs ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goPools ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goSnapShotT ∷ ∀ era. Era era ⇒ RootTarget era (SnapShot (EraCrypto era)) (SnapShot (EraCrypto era))

markPoolDistr ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

markPoolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

snapShotFee ∷ Era era ⇒ Term era Coin

snapShotsT ∷ ∀ era. Era era ⇒ RootTarget era (SnapShots (EraCrypto era)) (SnapShots (EraCrypto era))

deltaT ∷ Era era ⇒ Term era (Maybe DeltaCoin)

deltaTL ∷ NELens era (Maybe DeltaCoin)

deltaR ∷ Era era ⇒ Term era (Maybe DeltaCoin)

deltaRL ∷ NELens era (Maybe DeltaCoin)

deltaF ∷ Era era ⇒ Term era (Maybe DeltaCoin)

deltaFL ∷ NELens era (Maybe DeltaCoin)

rewardSet ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

rewardSetL ∷ NELens era (Maybe (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))))

totalAda ∷ Era era ⇒ Term era Coin

utxoCoin ∷ Era era ⇒ Term era Coin

credsUniv ∷ Era era ⇒ Term era (Set (Credential 'Staking (EraCrypto era)))

The universe of Staking Credentials. A credential is either KeyHash of a ScriptHash Any Plutus scripts in this Universe are NOT Spending scripts, so they do not need a Redeemer

spendCredsUniv ∷ Era era ⇒ Term era (Set (Credential 'Payment (EraCrypto era)))

The universe of Staking Credentials. A credential is either KeyHash of a ScriptHash All Plutus scripts in this Universe are SPENDING scripts, so they will need a Redeemer Use this ONLY in the Pay-part of an Address (Do not use this in the Stake-part of an Address)

voteUniv ∷ Era era ⇒ Term era (Set (Credential 'DRepRole (EraCrypto era)))

The universe of Voting Credentials. A credential is either KeyHash of a ScriptHash

drepUniv ∷ Era era ⇒ Term era (Set (DRep (EraCrypto era)))

The universe of DReps

hotCommitteeCredsUniv ∷ Era era ⇒ Term era (Set (Credential 'HotCommitteeRole (EraCrypto era)))

The universe of Credentials used in voting for constitutional committee changes.

coldCommitteeCredsUniv ∷ Era era ⇒ Term era (Set (Credential 'ColdCommitteeRole (EraCrypto era)))

The universe of Credentials used in voting for constitutional committee changes.

payUniv ∷ Era era ⇒ Term era (Set (Credential 'Payment (EraCrypto era)))

The universe of Payment Credentials. A credential is either KeyHash of a ScriptHash We only find payment credentials in the Payment part of an Addr.

spendscriptUniv ∷ Era era ⇒ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The universe of Scripts (and their hashes) useable in spending contexts That means if they are Plutus scripts then they will be passed an additional argument (the TxInfo context)

nonSpendScriptUniv ∷ Era era ⇒ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The universe of Scripts (and their hashes) useable in contexts other than Spending

allScriptUniv ∷ Era era ⇒ Proof era → Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

The union of spendscriptUniv and nonSpendScriptUniv. All possible scripts in any context

dataUniv ∷ Era era ⇒ Term era (Map (DataHash (EraCrypto era)) (Data era))

The universe of Data (and their hashes)

poolHashUniv ∷ Era era ⇒ Term era (Set (KeyHash 'StakePool (EraCrypto era)))

The universe of StakePool key hashes. These hashes hash the cold key of the Pool operators.

stakeHashUniv ∷ Era era ⇒ Term era (Set (KeyHash 'Staking (EraCrypto era)))

The universe of StakePool key hashes. These hashes hash are hashes of the Owners of a PoolParam

drepHashUniv ∷ Era era ⇒ Term era (Set (KeyHash 'DRepRole (EraCrypto era)))

The universe of DRep key hashes. These hashes hash are hashes of the DReps

genesisHashUniv ∷ Era era ⇒ Term era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

The universe of the Genesis key hashes and their signing and validating GenDelegPairs

voteCredUniv ∷ Era era ⇒ Term era (Set (Credential 'ColdCommitteeRole (EraCrypto era)))

txinUniv ∷ Era era ⇒ Term era (Set (TxIn (EraCrypto era)))

The universe of TxIns. Pairs of TxId: hashes of previously run transaction bodies, and TxIx: indexes of one of the bodies outputs.

govActionIdUniv ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

The universe of GovActionId. Pairs of TxId: hashes of previously run transaction bodies, and GovActionIx: indexes of one of the bodies Proposals .

txoutUniv ∷ Era era ⇒ Proof era → Term era (Set (TxOutF era))

The universe of TxOuts. It contains colTxoutUniv as a sublist and feeOutput as an element See also feeOutput which is defined by the universes, and is related.

colTxoutUniv ∷ Era era ⇒ Proof era → Term era (Set (TxOutF era))

The universe of TxOuts useable for collateral The collateral TxOuts consists only of VKey addresses and The collateral TxOuts do not contain any non-ADA part

feeTxOut ∷ Reflect era ⇒ Term era (TxOutF era)

A TxOut, guaranteed to have 1) no scripts in its Addr, and 2) It's Addr is in the addrUniv 3) bigCoin is stored in the Addr Value, and 4) the Addr Value has empty MutiAssets 5) be a member of the txoutUniv

feeTxIn ∷ Era era ⇒ Term era (TxIn (EraCrypto era))

A TxIn, guaranteed to have 1) be a member of the txinUniv

bigCoin ∷ Era era ⇒ Term era Coin

A Coin large enough to pay almost any fee. See also feeOutput which is related.

datumsUniv ∷ Era era ⇒ Term era [Datum era]

multiAssetUniv ∷ Era era ⇒ Term era [MultiAsset (EraCrypto era)]

keymapUniv ∷ Era era ⇒ Term era (Map (KeyHash 'Witness (EraCrypto era)) (KeyPair 'Witness (EraCrypto era)))

The universe of key hashes, and the signing and validating key pairs they represent.

currentSlot ∷ Era era ⇒ Term era SlotNo

endSlotDelta ∷ Era era ⇒ Term era SlotNo

beginSlotDelta ∷ Era era ⇒ Term era SlotNo

network ∷ Era era ⇒ Term era Network

From Globals

quorumConstant ∷ Word64

This not really a variable, But a constant that is set by the testGlobals we reflect this into a Term, so we can refer to it in the Preds.

quorum ∷ Era era ⇒ Term era Int

From Globals. Reflected here at type Int, This is set to quorumConstant in CertState. because is is used to compare the Size of things, which are computed as Int

addrUniv ∷ Era era ⇒ Term era (Set (Addr (EraCrypto era)))

ptrUniv ∷ Era era ⇒ Term era (Set Ptr)

plutusUniv ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (IsValid, ScriptF era))

spendPlutusUniv ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (IsValid, ScriptF era))

byronAddrUniv ∷ Era era ⇒ Term era (Map (KeyHash 'Payment (EraCrypto era)) (Addr (EraCrypto era), SigningKey))

The universe of all Byron addresses. In Eras, Babbage, Conway we avoid these Adresses, as they do not play well with Plutus Scripts.

newEpochStateConstr ∷ Proof era → EpochNo → Map (KeyHash 'StakePool (EraCrypto era)) Natural → Map (KeyHash 'StakePool (EraCrypto era)) Natural → EpochState era → Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)) → NewEpochState era

Abstract constuctor function for NewEpochState

newEpochStateT ∷ ∀ era. EraGov era ⇒ Proof era → RootTarget era (NewEpochState era) (NewEpochState era)

Target for NewEpochState

epochStateT ∷ ∀ era. EraGov era ⇒ Proof era → RootTarget era (EpochState era) (EpochState era)

Target for EpochState

accountStateT ∷ Era era ⇒ RootTarget era AccountState AccountState

Target for AccountState

ledgerStateT ∷ ∀ era. EraGov era ⇒ Proof era → RootTarget era (LedgerState era) (LedgerState era)

Target for LedgerState

ledgerState ∷ Reflect era ⇒ Term era (LedgerState era)

utxoStateT ∷ ∀ era. EraGov era ⇒ Proof era → RootTarget era (UTxOState era) (UTxOState era)

Target for UTxOState

unGovL ∷ Proof era → Lens' (GovState era) (GovState era)

justProtocolVersion ∷ ∀ era. Reflect era ⇒ Proof era → PParams era

certstateT ∷ ∀ era. Era era ⇒ RootTarget era (CertState era) (CertState era)

Target for CertState

vstateT ∷ ∀ era. Era era ⇒ RootTarget era (VState era) (VState era)

Target for VState

committeeL ∷ Lens' (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era))) (CommitteeState era)

pstateT ∷ ∀ era. Era era ⇒ RootTarget era (PState era) (PState era)

Target for PState

dstateT ∷ ∀ era. Era era ⇒ RootTarget era (DState era) (DState era)

Target for DState

dstate ∷ Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)) → Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)) → Map Ptr (Credential 'Staking (EraCrypto era)) → Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)) → Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → InstantaneousRewards (EraCrypto era) → DState era

Abstract construcor function for DState

instantaneousRewardsT ∷ ∀ era. Era era ⇒ RootTarget era (InstantaneousRewards (EraCrypto era)) (InstantaneousRewards (EraCrypto era))

allvars ∷ String

A String that pretty prints the complete set of variables of the NewEpochState

printTarget ∷ RootTarget era root t → IO ()

protVer ∷ Era era ⇒ Proof era → Term era ProtVer

ProtVer in pparams

prevProtVer ∷ Era era ⇒ Proof era → Term era ProtVer

ProtVer in prevPParams

minFeeA ∷ Era era ⇒ Proof era → Term era Coin

minFeeB ∷ Era era ⇒ Proof era → Term era Coin

maxBBSize ∷ Era era ⇒ Proof era → Term era Natural

Max Block Body Size

maxTxSize ∷ Era era ⇒ Proof era → Term era Natural

Max Tx Size

fromIntegralBounded ∷ ∀ a b. (HasCallStack, Integral a, Show a, Integral b, Bounded b, Show b) ⇒ String → a → b

word32NaturalL ∷ Lens' Word32 Natural

word16NaturalL ∷ Lens' Word16 Natural

maxBHSize ∷ Era era ⇒ Proof era → Term era Natural

Max Block Header Size

poolDepAmt ∷ Era era ⇒ Proof era → Term era Coin

keyDepAmt ∷ Era era ⇒ Proof era → Term era Coin

proposalDeposit ∷ ConwayEraPParams era ⇒ Proof era → Term era Coin

maxTxExUnits ∷ AlonzoEraPParams era ⇒ Proof era → Term era ExUnits

collateralPercentage ∷ AlonzoEraPParams era ⇒ Proof era → Term era Natural

drepDeposit ∷ ConwayEraPParams era ⇒ Proof era → Term era Coin

drepActivity ∷ ConwayEraPParams era ⇒ Proof era → Term era EpochInterval

maxEpoch ∷ Era era ⇒ Proof era → Term era EpochInterval

txbodyterm ∷ Reflect era ⇒ Term era (TxBodyF era)

inputs ∷ Era era ⇒ Term era (Set (TxIn (EraCrypto era)))

collateral ∷ Era era ⇒ Term era (Set (TxIn (EraCrypto era)))

refInputs ∷ Era era ⇒ Term era (Set (TxIn (EraCrypto era)))

outputs ∷ Era era ⇒ Proof era → Term era [TxOutF era]

collateralReturn ∷ Era era ⇒ Proof era → Term era (TxOutF era)

totalCol ∷ Era era ⇒ Term era Coin

The sum of all the collateral inputs. The Tx is constucted by SNothing or wrapping SJust around this value.

certs ∷ Reflect era ⇒ Term era [TxCertF era]

withdrawals ∷ ∀ era. Era era ⇒ Term era (Map (RewardAccount (EraCrypto era)) Coin)

txfee ∷ Era era ⇒ Term era Coin

ttl ∷ Era era ⇒ Term era SlotNo

validityInterval ∷ Era era ⇒ Term era ValidityInterval

mint ∷ Era era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (Map AssetName Integer))

reqSignerHashes ∷ Era era ⇒ Term era (Set (KeyHash 'Witness (EraCrypto era)))

networkID ∷ Era era ⇒ Term era (Maybe Network)

adHash ∷ Era era ⇒ Term era (Maybe (AuxiliaryDataHash (EraCrypto era)))

wppHash ∷ Era era ⇒ Term era (Maybe (SafeHash (EraCrypto era) EraIndependentScriptIntegrity))

txDonation ∷ Era era ⇒ Term era Coin

liftMultiAsset ∷ Map (ScriptHash c) (Map AssetName Integer) → MultiAsset c

lift the model type of mint into a MultiAsset

scriptsNeeded ∷ Reflect era ⇒ Term era (ScriptsNeededF era)

smNeededL ∷ ScriptsNeeded era ~ ShelleyScriptsNeeded era ⇒ Lens' (ScriptsNeededF era) (Set (ScriptHash (EraCrypto era)))

acNeededL ∷ ScriptsNeeded era ~ AlonzoScriptsNeeded era ⇒ Lens' (ScriptsNeededF era) [(PlutusPurposeF era, ScriptHash (EraCrypto era))]

extraCol ∷ Era era ⇒ Term era Coin

A Coin that needs to be added to the range of the colInputs in the UtxO that will make sure the collateral is large enough to pay the fees if needed

sumCol ∷ Era era ⇒ Term era Coin

The sum of all the collateral inputs, total colateral of the Tx is computed by adding (SJust _) to this value.

colRetAddr ∷ Era era ⇒ Term era (Addr (EraCrypto era))

colRetCoin ∷ Era era ⇒ Term era Coin

The Coin in the collateralReturn TxOut

owed ∷ Era era ⇒ Term era Coin

The amount that the collateral must cover if there is a two phase error. This is roughly the collateralPercentage * txfee . The calculation deals with rounding, but you don't need those details to understand what is going on.

txbody ∷ Reflect era ⇒ Term era (TxBodyF era)

txwits ∷ Reflect era ⇒ Term era (TxWitsF era)

txauxdata ∷ Reflect era ⇒ Term era (Maybe (TxAuxDataF era))

txisvalid ∷ Era era ⇒ Term era IsValid

valids ∷ Era era ⇒ Term era [IsValid]

txterm ∷ Reflect era ⇒ Term era (TxF era)

getRwdCredL ∷ Lens' (RewardAccount c) (Credential 'Staking c)

txOutFL ∷ Lens' (TxOutF era) (TxOut era)

valueFL ∷ Reflect era ⇒ Lens' (Value era) (ValueF era)

lensVC ∷ Val t ⇒ Lens' t Coin

valueFCoinL ∷ (HasCallStack, Reflect era) ⇒ Lens' (ValueF era) Coin

outputCoinL ∷ (HasCallStack, Reflect era) ⇒ Lens' (TxOutF era) Coin

valCoinF ∷ (HasCallStack, Reflect era) ⇒ Field era (ValueF era) Coin

a Field from (ValueF era) to Coin

valCoin ∷ (HasCallStack, Reflect era) ⇒ Term era Coin

maryValueMultiAssetL ∷ Lens' (MaryValue c) (MultiAsset c)

valueFMultiAssetL ∷ Lens' (ValueF era) (MultiAsset (EraCrypto era))

valueFMultiAssetF ∷ Reflect era ⇒ Field era (ValueF era) (MultiAsset (EraCrypto era))

a Field from (ValueF era) to MultiAsset

valueFMultiAsset ∷ Reflect era ⇒ Term era (MultiAsset (EraCrypto era))

txoutAddressF ∷ Reflect era ⇒ Field era (TxOutF era) (Addr (EraCrypto era))

a Field from (TxOut era) to (Addr era)

txoutAddress ∷ Reflect era ⇒ Term era (Addr (EraCrypto era))

txoutCoinF ∷ (HasCallStack, Reflect era) ⇒ Field era (TxOutF era) Coin

a Field from (TxOutF era) to Coin

txoutCoin ∷ (HasCallStack, Reflect era) ⇒ Term era Coin

txoutAmountF ∷ Reflect era ⇒ Field era (TxOutF era) (ValueF era)

a Field from (TxOutF era) to (ValueF era)

txoutAmount ∷ Reflect era ⇒ Term era (ValueF era)

scriptWits ∷ Reflect era ⇒ Term era (Map (ScriptHash (EraCrypto era)) (ScriptF era))

redeemers ∷ Reflect era ⇒ Term era (Map (PlutusPointerF era) (Data era, ExUnits))

bootWits ∷ ∀ era. Reflect era ⇒ Term era (Set (BootstrapWitness (EraCrypto era)))

dataWits ∷ Reflect era ⇒ Term era (Map (DataHash (EraCrypto era)) (Data era))

keyWits ∷ Reflect era ⇒ Term era (Set (WitVKey 'Witness (EraCrypto era)))

witsTarget ∷ Reflect era ⇒ Term era (Set (BootstrapWitness (EraCrypto era))) → Term era (Set (WitVKey 'Witness (EraCrypto era))) → Target era (TxWits era)

txTarget ∷ Reflect era ⇒ Term era (TxBodyF era) → Term era (Set (BootstrapWitness (EraCrypto era))) → Term era (Set (WitVKey 'Witness (EraCrypto era))) → Target era (TxF era)

txbodyTarget ∷ Reflect era ⇒ Term era Coin → Term era (Maybe (ScriptIntegrityHash (EraCrypto era))) → Term era Coin → Target era (TxBodyF era)

Need to build the TxBody with different terms that control the fee and wppHash so we parameterise this target over those two terms

allowMIRTransfer ∷ Proof era → Term era Bool

constitution ∷ Era era ⇒ Term era (Constitution era)

enactTreasury ∷ Era era ⇒ Term era Coin

enactWithdrawals ∷ ∀ era. Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

currentGovActionStates ∷ Era era ⇒ Term era (Map (GovActionId (EraCrypto era)) (GovActionState era))

currentProposalOrder ∷ Era era ⇒ Term era [GovActionId (EraCrypto era)]

prevGovActionStates ∷ Era era ⇒ Term era (Map (GovActionId (EraCrypto era)) (GovActionState era))

prevProposalOrder ∷ Era era ⇒ Term era [GovActionId (EraCrypto era)]

previousCommitteeState ∷ Era era ⇒ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (Maybe (Credential 'HotCommitteeRole (EraCrypto era))))

commMembers ∷ Era era ⇒ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

commQuorum ∷ Era era ⇒ Term era UnitInterval

committeeVar ∷ Era era ⇒ Term era (Maybe (Committee era))

type UtxoPulse era = (Map (TxIn (EraCrypto era)) (TxOutF era), DRepPulser era Identity (RatifyState era))

There are 2 forms of DRepPulsingState. This is part of the first one where the pulsing is not complete, and the snapshots are stored as fields in the datatype DRepPulser. Note that the function part of Invert : initPulser makes many transformations from the types used in the Model, and the types stored in the implementation types. In order to construct a valid DRepPulser we need the UTxO (to compute the IncrementalStake) But we cannot find a Lens that can recover the UTxO from a DRepPulser. So we introduce this type ' UtxoPulse' that pairs the two (which makes the recovery possible). W

utxoPulse ∷ (RunConwayRatify era, Reflect era) ⇒ Proof era → Term era (UtxoPulse era)

We also introduce an intermediate variable utxoPulse which can constrain this value by using the predicate [ utxoPulse p :<-: pulsingPair p ]

pulsingPairT ∷ ∀ era. (RunConwayRatify era, Reflect era) ⇒ Proof era → RootTarget era (UtxoPulse era) (UtxoPulse era)

an invertable RootTarget to compute a (UtxoPulse era)

justPulser ∷ ∀ era. (Reflect era, RunConwayRatify era) ⇒ Proof era → RootTarget era (DRepPulser era Identity (RatifyState era)) (DRepPulser era Identity (RatifyState era))

drepPulser ∷ (RunConwayRatify era, Reflect era) ⇒ Term era (DRepPulser era Identity (RatifyState era))

Variable used to constrain the DRepPulser

prevPulsingPreds ∷ (RunConwayRatify era, Reflect era) ⇒ Proof era → [Pred era]

Predicates that constrain the DRepPuser and all its prevXXX snapshots These ensure we generate state just passing the epoch boundary

pulsingPulsingStateT ∷ ∀ era. (RunConwayRatify era, Reflect era) ⇒ RootTarget era (DRepPulsingState era) (DRepPulsingState era)

Target for assembling DRPulsing form of (DRepPulsingState era) from drepPulser :: forall era. Term era (DRepPulser era Identity (RatifyState era))

pulsingStatePulserL ∷ Lens' (DRepPulsingState era) (DRepPulser era Identity (RatifyState era))

The Lens' used in pulsingPulsingStateT

initPulser ∷ ∀ era. (Reflect era, RunConwayRatify era) ⇒ Proof era → Map (TxIn (EraCrypto era)) (TxOutF era) → Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)) → Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)) → Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)) → EpochNo → Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)) → EnactState era → [GovActionState era] → DRepPulser era Identity (RatifyState era)

The abstract form of DRepPulser that transforms from the Model types used in the inputs, and the concrete types actually stored in DRepPulser

proposalsT ∷ ∀ era. Era era ⇒ Proof era → RootTarget era (Proposals era) (Proposals era)

pulsingSnapshotT ∷ ∀ era. Era era ⇒ RootTarget era (PulsingSnapshot era) (PulsingSnapshot era)

The snapshot dedicated datatype (PulsingSnapshot era) stored inside DRComplete Note this is used in dRepPulsingStateT, the second DRepPulsingState form.

pulsingSnapshotL ∷ Lens' (DRepPulsingState era) (PulsingSnapshot era)

completePulsingStateT ∷ ∀ era. Reflect era ⇒ Proof era → RootTarget era (DRepPulsingState era) (DRepPulsingState era)

There are 2 forms of DRepPulsingState. This is the second one where the pulsing is complete

ratifyState ∷ Reflect era ⇒ Term era (RatifyState era)

ratifyStateL ∷ Lens' (DRepPulsingState era) (RatifyState era)

prevProposals ∷ Era era ⇒ Proof era → Term era (Proposals era)

ratifyGovActionStatesL ∷ Lens' (DRepPulser era Identity (RatifyState era)) [GovActionState era]

partialDRepDistr ∷ Era era ⇒ Term era (Map (DRep (EraCrypto era)) Coin)

Partially computed DRepDistr inside the pulser

partialDRepDistrL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (Map (DRep (EraCrypto era)) Coin)

prevDRepState ∷ Era era ⇒ Term era (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

Snapshot of dreps from the start of the current epoch

prevDRepStateL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

prevPoolDistr ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

snapshot of poolDistr from the start of the current epoch

prevPoolDistrL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

prevDRepDelegations ∷ Era era ⇒ Term era (Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)))

Snapshot of the drepDelegation from he start of the current epoch.

prevDRepDelegationsL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (Map (Credential 'Staking (EraCrypto era)) (DRep (EraCrypto era)))

Snapshot of drepDelegation from the start of the current epoch

prevCommitteeState ∷ Era era ⇒ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)))

Snapshot of committeeState from the start of the current epoch

prevCommitteeStateL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (Map (Credential 'ColdCommitteeRole (EraCrypto era)) (CommitteeAuthorization (EraCrypto era)))

prevEnactState ∷ Reflect era ⇒ Term era (EnactState era)

Snapshot of the enactState built by enactStateT assembled from data at the start the current epoch

prevEnactStateL ∷ Lens' (DRepPulser era Identity (RatifyState era)) (EnactState era)

prevEpoch ∷ Era era ⇒ Term era EpochNo

Snapshot of currentEpoch just before the start of the current epoch. (currenEpoch - 1)

prevEpochL ∷ Lens' (DRepPulser era Identity (RatifyState era)) EpochNo

prevTreasury ∷ Era era ⇒ Term era Coin

partialIndividualPoolStake ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

conwayGovStateT ∷ ∀ era. (RunConwayRatify era, Reflect era) ⇒ Proof era → RootTarget era (ConwayGovState era) (ConwayGovState era)

proposalDeposits ∷ Era era ⇒ Term era Coin

The sum of all the gasDeposit fields of currProposals

drepDepositsView ∷ Era era ⇒ Term era (Map (Credential 'DRepRole (EraCrypto era)) Coin)

A view of currentDRepState (sum of the drepDeposit field of in the range of currentDRepState)

currProposals ∷ Era era ⇒ Proof era → Term era (Proposals era)

The current set of proposals. Proposals has a serious set of invariants. We do not attempt to state these proposals (Yes I know that is cheating) We get random Proposals (that meets its invariants) by using (genSizedRep n (ProposalsR p))

prevGovActionIds ∷ ∀ era. Reflect era ⇒ Term era (GovRelation StrictMaybe era)

Part of the EnactState, it is computed by selecting from currProposals

currGovStates ∷ Era era ⇒ Term era [GovActionState era]

This is a view of currProposals, so will compute it once once currProposals is defined

enactStateT ∷ ∀ era. Reflect era ⇒ RootTarget era (EnactState era) (EnactState era)

committeeT ∷ ∀ era. Era era ⇒ RootTarget era (Committee era) (Committee era)

One can use this Target, to make a constraint for committeeVar from the vars commMembers and commQuorum

ppUpdateChildren ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

hardForkChildren ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

committeeChildren ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

constitutionChildren ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

pparamsFL ∷ Proof era → Lens' (PParams era) (PParamsF era)

pparamsMaybeFL ∷ Proof era → Lens' (Maybe (PParams era)) (Maybe (PParamsF era))

smCommL ∷ Lens' (StrictMaybe (Committee era)) (Committee era)

proposedMapL ∷ Proof era → Lens' (ProposedPPUpdates era) (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

pair1 ∷ Era era ⇒ Rep era a → Term era a

pair2 ∷ Era era ⇒ Rep era b → Term era b

pairT ∷ ∀ era a b. (Typeable a, Typeable b, Era era) ⇒ Rep era a → Rep era b → RootTarget era (a, b) (a, b)

idV ∷ Era era ⇒ Term era (GovActionId (EraCrypto era))

committeeVotesV ∷ Era era ⇒ Term era (Map (Credential 'HotCommitteeRole (EraCrypto era)) Vote)

drepVotesV ∷ Era era ⇒ Term era (Map (Credential 'DRepRole (EraCrypto era)) Vote)

stakePoolVotesV ∷ Era era ⇒ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Vote)

depositV ∷ Era era ⇒ Term era Coin

returnAddrV ∷ Era era ⇒ Term era (RewardAccount (EraCrypto era))

actionV ∷ Era era ⇒ Term era (GovAction era)

proposedInV ∷ Era era ⇒ Term era EpochNo

expiresAfterV ∷ Era era ⇒ Term era EpochNo

childrenV ∷ Era era ⇒ Term era (Set (GovActionId (EraCrypto era)))

anchorV ∷ Era era ⇒ Term era (Anchor (EraCrypto era))

govActionStateTarget ∷ ∀ era. Era era ⇒ RootTarget era (GovActionState era) (GovActionState era)

liftId ∷ Maybe (GovActionId (EraCrypto era)) → StrictMaybe (GovPurposeId p era)

Lift the Model to the real type

dropId ∷ StrictMaybe (GovPurposeId p era) → Maybe (GovActionId (EraCrypto era))

Drop the real type back to the Model

gaPrevId ∷ Era era ⇒ Term era (Maybe (GovActionId (EraCrypto era)))

gaPParamsUpdate ∷ Reflect era ⇒ Term era (PParamsUpdateF era)

gaProtVer ∷ Reflect era ⇒ Term era ProtVer

gaRewardAccount ∷ Era era ⇒ Term era (Map (RewardAccount (EraCrypto era)) Coin)

gaRemMember ∷ Era era ⇒ Term era (Set (Credential 'ColdCommitteeRole (EraCrypto era)))

gaAddMember ∷ Era era ⇒ Term era (Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

gaThreshold ∷ Era era ⇒ Term era UnitInterval

gaPolicy ∷ Era era ⇒ Term era (Maybe (ScriptHash (EraCrypto era)))

gaConstitutionAnchor ∷ Era era ⇒ Term era (Anchor (EraCrypto era))

gaNewConstitution ∷ Era era ⇒ Term era (Constitution era)

constitutionT ∷ ∀ era. Reflect era ⇒ RootTarget era (Constitution era) (Constitution era)

parameterChangeT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

hardForkInitiationT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

treasuryWithdrawalsT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

noConfidenceT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

updateCommitteeT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

newConstitutionT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)

infoActionT ∷ ∀ era. Reflect era ⇒ RootTarget era (GovAction era) (GovAction era)