Test.Cardano.Ledger.Constrained.Vars

Documentation

field ∷ (Eq t, Show t) ⇒ Rep era s → Term era t → AnyF era s

Used in Component constraints to turn a Var Term into a component (AnyF era s) E.g. (Component foo [ field fooRep fooPart1, field fooRep fooPart2]) Where fooPart1 :: Term era a, and fooPart2 :: Term era b And fooPart1 has an (Access foo a) And fooPart2 has an (Access foo b)

getName ∷ Term era t → Name era

type NELens era t = Lens' (NewEpochState era) t

epochNo ∷ Term era EpochNo

prevBlocksMade ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

currBlocksMade ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Natural)

poolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

mockPoolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Rational)

For tests only, Like PoolDistr but has a Rational (rather than a IndividualPoolStake).

poolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

unPoolDistrL ∷ Lens' (PoolDistr c) (Map (KeyHash 'StakePool c) (IndividualPoolStake c))

rewards ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

rewardsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

delegations ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

delegationsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

stakeDeposits ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

stakeDepositsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

ptrs ∷ Term era (Map Ptr (Credential 'Staking (EraCrypto era)))

ptrsL ∷ NELens era (Map Ptr (Credential 'Staking (EraCrypto era)))

futureGenDelegs ∷ Term era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

futureGenDelegsL ∷ NELens era (Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegs ∷ Term era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

genDelegsL ∷ NELens era (Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)))

instanReserves ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanReservesL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasury ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

instanTreasuryL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

deltaReserves ∷ Term era DeltaCoin

deltaReservesNEL ∷ NELens era DeltaCoin

deltaTreasury ∷ Term era DeltaCoin

deltaTreasuryNEL ∷ NELens era DeltaCoin

regPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

regPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

futureRegPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

retiring ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

retiringL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) EpochNo)

poolDeposits ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

poolDepositsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) Coin)

dreps ∷ Term era (Set (Credential 'Voting (EraCrypto era)))

drepsL ∷ NELens era (Set (Credential 'Voting (EraCrypto era)))

ccHotKeys ∷ Term era (Map (KeyHash 'Voting (EraCrypto era)) (KeyHash 'Voting (EraCrypto era)))

ccHotKeysL ∷ NELens era (Map (KeyHash 'Voting (EraCrypto era)) (KeyHash 'Voting (EraCrypto era)))

utxo ∷ Proof era → Term era (Map (TxIn (EraCrypto era)) (TxOutF era))

utxoL ∷ Proof era → NELens era (Map (TxIn (EraCrypto era)) (TxOutF era))

unUtxoL ∷ Proof era → Lens' (UTxO era) (Map (TxIn (EraCrypto era)) (TxOutF era))

deposits ∷ Term era Coin

depositsL ∷ NELens era Coin

fees ∷ Term era Coin

feesL ∷ NELens era Coin

ppup ∷ Proof era → Term era (ShelleyPPUPState era)

ppupsL ∷ Proof era → NELens era (ShelleyPPUPState era)

proposalsT ∷ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

futureProposalsT ∷ Proof era → Term era (Map (KeyHash 'Genesis (EraCrypto era)) (PParamsUpdateF era))

ppupStateT ∷ Proof era → Target era (ShelleyPPUPState era)

governanceStateT ∷ ∀ era. Proof era → Target era (GovernanceState era)

individualPoolStakeL ∷ Lens' (IndividualPoolStake c) Rational

isPtrMapT ∷ Term era (Map Ptr Coin)

ptrMapL ∷ Lens' (NewEpochState era) (Map Ptr Coin)

isCredMapT ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

credMapL ∷ Lens' (NewEpochState era) (Map (Credential 'Staking (EraCrypto era)) Coin)

treasury ∷ Term era Coin

treasuryL ∷ NELens era Coin

reserves ∷ Term era Coin

reservesL ∷ NELens era Coin

snapshots ∷ Term era (SnapShots (EraCrypto era))

snapshotsL ∷ NELens era (SnapShots (EraCrypto era))

ppFL ∷ Proof era → Lens' (PParams era) (PParamsF era)

Lens' from the Core PParams to the Model PParamsF which embeds a (Proof era)

prevpparams ∷ Proof era → Term era (PParamsF era)

pparams ∷ Proof era → Term era (PParamsF era)

nmLikelihoodsT ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) [Float])

nmRewardPotT ∷ Term era Coin

stakeL ∷ Lens' (Stake c) (Map (Credential 'Staking c) Coin)

Helper lens that deals with the Stake newtype, and the shift from Map to VMap

vmapL ∷ Lens' (VMap VB VB k v) (Map k v)

Helper lens that deals with the shift from Map to VMap

markStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

markStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

markDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

markPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

markSnapShotT ∷ Target era (SnapShot (EraCrypto era))

setStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

setStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

setDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

setPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

setSnapShotT ∷ Target era (SnapShot (EraCrypto era))

goStake ∷ Term era (Map (Credential 'Staking (EraCrypto era)) Coin)

goStakeL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) Coin)

goDelegs ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goDelegsL ∷ NELens era (Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)))

goPools ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goPoolsL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era)))

goSnapShotT ∷ Target era (SnapShot (EraCrypto era))

markPoolDistr ∷ Term era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

markPoolDistrL ∷ NELens era (Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)))

pooldistrHelpL ∷ Lens' (PoolDistr c) (Map (KeyHash 'StakePool c) (IndividualPoolStake c))

snapShotFee ∷ Term era Coin

snapShotsT ∷ Target era (SnapShots (EraCrypto era))

deltaT ∷ Term era (Maybe DeltaCoin)

deltaTL ∷ NELens era (Maybe DeltaCoin)

deltaR ∷ Term era (Maybe DeltaCoin)

deltaRL ∷ NELens era (Maybe DeltaCoin)

deltaF ∷ Term era (Maybe DeltaCoin)

deltaFL ∷ NELens era (Maybe DeltaCoin)

rewardSet ∷ Term era (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era))))

rewardSetL ∷ NELens era (Maybe (Map (Credential 'Staking (EraCrypto era)) (Set (Reward (EraCrypto era)))))

totalAda ∷ Term era Coin

utxoCoin ∷ Term era Coin

credsUniv ∷ Term era (Set (Credential 'Staking (EraCrypto era)))

poolsUniv ∷ Term era (Set (KeyHash 'StakePool (EraCrypto era)))

genesisUniv ∷ Term era (Set (KeyHash 'Genesis (EraCrypto era)))

txinUniv ∷ Term era (Set (TxIn (EraCrypto era)))

newEpochStateConstr ∷ Proof era → EpochNo → Map (KeyHash 'StakePool (EraCrypto era)) Natural → Map (KeyHash 'StakePool (EraCrypto era)) Natural → EpochState era → Map (KeyHash 'StakePool (EraCrypto era)) (IndividualPoolStake (EraCrypto era)) → NewEpochState era

Abstract constuctor function for NewEpochState

newEpochStateT ∷ Proof era → Target era (NewEpochState era)

Target for NewEpochState

epochStateT ∷ Proof era → Target era (EpochState era)

Target for EpochState

accountStateT ∷ Target era AccountState

Target for AccountState

ledgerStateT ∷ Proof era → Target era (LedgerState era)

Target for LedgerState

utxoStateT ∷ Proof era → Target era (UTxOState era)

Target for UTxOState

dpstateT ∷ Target era (CertState era)

Target for CertState

vstateT ∷ Target era (VState era)

Target for VState

pstateT ∷ Target era (PState era)

Target for PState

dstateT ∷ Target era (DState era)

Target for DState

dstate ∷ Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) Coin → Map (Credential 'Staking (EraCrypto era)) (KeyHash 'StakePool (EraCrypto era)) → Map Ptr (Credential 'Staking (EraCrypto era)) → Map (FutureGenDeleg (EraCrypto era)) (GenDelegPair (EraCrypto era)) → Map (KeyHash 'Genesis (EraCrypto era)) (GenDelegPair (EraCrypto era)) → InstantaneousRewards (EraCrypto era) → DState era

Abstract construcor function for DState

instantaneousRewardsT ∷ Target era (InstantaneousRewards (EraCrypto era))

allvars ∷ String

A String that pretty prints the complete set of variables

withEraPParams ∷ ∀ era a. Proof era → (EraPParams era ⇒ a) → a

protVer ∷ Proof era → Term era ProtVer

ProtVer in pparams

prevProtVer ∷ Proof era → Term era ProtVer

ProtVer in prevpparams

minFeeA ∷ Proof era → Term era Coin

minFeeB ∷ Proof era → Term era Coin

maxBBSize ∷ Proof era → Term era Natural

Max Block Body Size

maxTxSize ∷ Proof era → Term era Natural

Max Tx Size

maxBHSize ∷ Proof era → Term era Natural

Max Block Header Size