Test.Cardano.Ledger.EraClass

Description

Everything you need to build Era polymorphic transactions, all in one place

Documentation

class (EraTxBody era, EraTxWits era, EraTxAuxData era, EraPParams era, NoThunks (Tx era), DecCBOR (Tx era), EncCBOR (Tx era), ToCBOR (Tx era), Show (Tx era), Eq (Tx era), EqRaw (Tx era)) ⇒ EraTx era where

A transaction.

Minimal complete definition

mkBasicTx, bodyTxL, witsTxL, auxDataTxL, sizeTxF, wireSizeTxF, validateNativeScript, getMinFeeTx, upgradeTx

Associated Types

type Tx era = (r ∷ Type) | r → era

Methods

mkBasicTx ∷ TxBody era → Tx era

bodyTxL ∷ Lens' (Tx era) (TxBody era)

witsTxL ∷ Lens' (Tx era) (TxWits era)

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

class (EraTxOut era, EraTxCert era, EraPParams era, HashAnnotated (TxBody era) EraIndependentTxBody, DecCBOR (TxBody era), EncCBOR (TxBody era), ToCBOR (TxBody era), NoThunks (TxBody era), NFData (TxBody era), Show (TxBody era), Eq (TxBody era), EqRaw (TxBody era)) ⇒ EraTxBody era where

Minimal complete definition

mkBasicTxBody, inputsTxBodyL, outputsTxBodyL, feeTxBodyL, withdrawalsTxBodyL, auxDataHashTxBodyL, spendableInputsTxBodyF, allInputsTxBodyF, certsTxBodyL, upgradeTxBody

Associated Types

type TxBody era = (r ∷ Type) | r → era

The body of a transaction.

Methods

mkBasicTxBody ∷ TxBody era

inputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxOut era))

feeTxBodyL ∷ Lens' (TxBody era) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody era) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe TxAuxDataHash)

certsTxBodyL ∷ Lens' (TxBody era) (StrictSeq (TxCert era))

class EraTxBody era ⇒ AllegraEraTxBody era where

Methods

vldtTxBodyL ∷ Lens' (TxBody era) ValidityInterval

class AllegraEraTxBody era ⇒ MaryEraTxBody era where

Minimal complete definition

mintTxBodyL, mintedTxBodyF

Methods

mintTxBodyL ∷ Lens' (TxBody era) MultiAsset

mintedTxBodyF ∷ SimpleGetter (TxBody era) (Set PolicyID)

mintValueTxBodyF ∷ SimpleGetter (TxBody era) (Value era)

class (MaryEraTxBody era, AlonzoEraTxOut era) ⇒ AlonzoEraTxBody era where

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody era) (Set TxIn)

reqSignerHashesTxBodyL ∷ Lens' (TxBody era) (Set (KeyHash 'Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody era) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Network)

redeemerPointer ∷ TxBody era → PlutusPurpose AsItem era → StrictMaybe (PlutusPurpose AsIx era)

This function is called rdptr in the spec. Given a TxBody and a plutus purpose with an item, we should be able to find the plutus purpose as in index

redeemerPointerInverse ∷ TxBody era → PlutusPurpose AsIx era → StrictMaybe (PlutusPurpose AsIxItem era)

This is an inverse of redeemerPointer. Given purpose as an index return it as an item.

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era where

Methods

currentTreasuryValueTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

Lens for getting and setting number of Coin that is expected to be in the Treasury at the current Epoch

votingProceduresTxBodyL ∷ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.

treasuryDonationTxBodyL ∷ Lens' (TxBody era) Coin

class (Val (Value era), ToJSON (TxOut era), DecCBOR (Value era), DecCBOR (CompactForm (Value era)), MemPack (CompactForm (Value era)), EncCBOR (Value era), ToCBOR (TxOut era), EncCBOR (TxOut era), DecCBOR (TxOut era), DecShareCBOR (TxOut era), Share (TxOut era) ~ Interns (Credential 'Staking), NoThunks (TxOut era), NFData (TxOut era), Show (TxOut era), Eq (TxOut era), MemPack (TxOut era), EraPParams era) ⇒ EraTxOut era where

Abstract interface into specific fields of a TxOut

Minimal complete definition

mkBasicTxOut, upgradeTxOut, valueEitherTxOutL, addrEitherTxOutL, (getMinCoinSizedTxOut | getMinCoinTxOut)

Associated Types

type TxOut era = (r ∷ Type) | r → era

The output of a UTxO for a particular era

Methods

mkBasicTxOut ∷ Addr → Value era → TxOut era

valueTxOutL ∷ Lens' (TxOut era) (Value era)

compactValueTxOutL ∷ Lens' (TxOut era) (CompactForm (Value era))

valueEitherTxOutL ∷ Lens' (TxOut era) (Either (Value era) (CompactForm (Value era)))

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

addrTxOutL ∷ Lens' (TxOut era) Addr

compactAddrTxOutL ∷ Lens' (TxOut era) CompactAddr

addrEitherTxOutL ∷ Lens' (TxOut era) (Either Addr CompactAddr)

Lens for getting and setting in TxOut either an address or its compact version by doing the least amount of work.

The utility of this function comes from the fact that TxOut usually stores the address in either one of two forms: compacted or unpacked. In order to avoid extroneous conversions in getTxOutAddr and getTxOutCompactAddr we can define just this functionality. Also sometimes it is crucial to know at the callsite which form of address we have readily available without any conversions (eg. searching millions of TxOuts for a particular address)

class (AlonzoEraPParams era, EraTxOut era) ⇒ AlonzoEraTxOut era where

Methods

dataHashTxOutL ∷ Lens' (TxOut era) (StrictMaybe DataHash)

datumTxOutF ∷ SimpleGetter (TxOut era) (Datum era)

class (AlonzoEraTxOut era, AlonzoEraScript era) ⇒ BabbageEraTxOut era where

Methods

referenceScriptTxOutL ∷ Lens' (TxOut era) (StrictMaybe (Script era))

dataTxOutL ∷ Lens' (TxOut era) (StrictMaybe (Data era))

datumTxOutL ∷ Lens' (TxOut era) (Datum era)

coinTxOutL ∷ (HasCallStack, EraTxOut era) ⇒ Lens' (TxOut era) Coin

class (EraScript era, Eq (TxWits era), EqRaw (TxWits era), Show (TxWits era), Monoid (TxWits era), NoThunks (TxWits era), ToCBOR (TxWits era), EncCBOR (TxWits era), DecCBOR (TxWits era)) ⇒ EraTxWits era where

A collection of witnesses in a Tx

Minimal complete definition

addrTxWitsL, bootAddrTxWitsL, scriptTxWitsL, upgradeTxWits

Associated Types

type TxWits era = (r ∷ Type) | r → era

Methods

mkBasicTxWits ∷ TxWits era

addrTxWitsL ∷ Lens' (TxWits era) (Set (WitVKey 'Witness))

bootAddrTxWitsL ∷ Lens' (TxWits era) (Set BootstrapWitness)

scriptTxWitsL ∷ Lens' (TxWits era) (Map ScriptHash (Script era))

class (EraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTxWits era where

Methods

datsTxWitsL ∷ Lens' (TxWits era) (TxDats era)

rdmrsTxWitsL ∷ Lens' (TxWits era) (Redeemers era)

class (Era era, Eq (TxAuxData era), EqRaw (TxAuxData era), Show (TxAuxData era), NoThunks (TxAuxData era), ToCBOR (TxAuxData era), EncCBOR (TxAuxData era), DecCBOR (TxAuxData era), HashAnnotated (TxAuxData era) EraIndependentTxAuxData) ⇒ EraTxAuxData era where

TxAuxData which may be attached to a transaction

Minimal complete definition

mkBasicTxAuxData, metadataTxAuxDataL, upgradeTxAuxData, validateTxAuxData

Associated Types

type TxAuxData era = (r ∷ Type) | r → era

Methods

validateTxAuxData ∷ ProtVer → TxAuxData era → Bool

class (Era era, ToJSON (TxCert era), DecCBOR (TxCert era), EncCBOR (TxCert era), ToCBOR (TxCert era), FromCBOR (TxCert era), NoThunks (TxCert era), NFData (TxCert era), Show (TxCert era), Ord (TxCert era), Eq (TxCert era)) ⇒ EraTxCert era where

Minimal complete definition

upgradeTxCert, getVKeyWitnessTxCert, getScriptWitnessTxCert, mkRegPoolTxCert, getRegPoolTxCert, mkRetirePoolTxCert, getRetirePoolTxCert, lookupRegStakeTxCert, lookupUnRegStakeTxCert, getTotalDepositsTxCerts, getTotalRefundsTxCerts

Associated Types

type TxCert era = (r ∷ Type) | r → era

Methods

getVKeyWitnessTxCert ∷ TxCert era → Maybe (KeyHash 'Witness)

Return a witness key whenever a certificate requires one

getScriptWitnessTxCert ∷ TxCert era → Maybe ScriptHash

Return a ScriptHash for certificate types that require a witness

mkRegPoolTxCert ∷ PoolParams → TxCert era

getRegPoolTxCert ∷ TxCert era → Maybe PoolParams

mkRetirePoolTxCert ∷ KeyHash 'StakePool → EpochNo → TxCert era

getRetirePoolTxCert ∷ TxCert era → Maybe (KeyHash 'StakePool, EpochNo)

lookupRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can register such credential

lookupUnRegStakeTxCert ∷ TxCert era → Maybe (Credential 'Staking)

Extract staking credential from any certificate that can unregister such credential

getTotalDepositsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (KeyHash 'StakePool → Bool)
Check whether stake pool is registered or not

→ f (TxCert era)

→ Coin

Compute the total deposits from a list of certificates.

getTotalRefundsTxCerts

Arguments

∷ Foldable f

⇒ PParams era

→ (Credential 'Staking → Maybe Coin)
Lookup current deposit for Staking credential if one is registered

→ (Credential 'DRepRole → Maybe Coin)
Lookup current deposit for DRep credential if one is registered

→ f (TxCert era)

→ Coin

Compute the total refunds from a list of certificates.

class ShelleyEraTxCert era ⇒ ConwayEraTxCert era where

Methods

mkRegDepositTxCert ∷ StakeCredential → Coin → TxCert era

getRegDepositTxCert ∷ TxCert era → Maybe (StakeCredential, Coin)

mkUnRegDepositTxCert ∷ StakeCredential → Coin → TxCert era

getUnRegDepositTxCert ∷ TxCert era → Maybe (StakeCredential, Coin)

mkDelegTxCert ∷ StakeCredential → Delegatee → TxCert era

getDelegTxCert ∷ TxCert era → Maybe (StakeCredential, Delegatee)

mkRegDepositDelegTxCert ∷ StakeCredential → Delegatee → Coin → TxCert era

getRegDepositDelegTxCert ∷ TxCert era → Maybe (StakeCredential, Delegatee, Coin)

mkAuthCommitteeHotKeyTxCert ∷ Credential 'ColdCommitteeRole → Credential 'HotCommitteeRole → TxCert era

getAuthCommitteeHotKeyTxCert ∷ TxCert era → Maybe (Credential 'ColdCommitteeRole, Credential 'HotCommitteeRole)

mkResignCommitteeColdTxCert ∷ Credential 'ColdCommitteeRole → StrictMaybe Anchor → TxCert era

getResignCommitteeColdTxCert ∷ TxCert era → Maybe (Credential 'ColdCommitteeRole, StrictMaybe Anchor)

mkRegDRepTxCert ∷ Credential 'DRepRole → Coin → StrictMaybe Anchor → TxCert era

getRegDRepTxCert ∷ TxCert era → Maybe (Credential 'DRepRole, Coin, StrictMaybe Anchor)

mkUnRegDRepTxCert ∷ Credential 'DRepRole → Coin → TxCert era

getUnRegDRepTxCert ∷ TxCert era → Maybe (Credential 'DRepRole, Coin)

mkUpdateDRepTxCert ∷ Credential 'DRepRole → StrictMaybe Anchor → TxCert era

getUpdateDRepTxCert ∷ TxCert era → Maybe (Credential 'DRepRole, StrictMaybe Anchor)

class (Era era, Eq (PParamsHKD Identity era), Ord (PParamsHKD Identity era), Show (PParamsHKD Identity era), NFData (PParamsHKD Identity era), EncCBOR (PParamsHKD Identity era), DecCBOR (PParamsHKD Identity era), ToCBOR (PParamsHKD Identity era), FromCBOR (PParamsHKD Identity era), NoThunks (PParamsHKD Identity era), ToJSON (PParamsHKD Identity era), FromJSON (PParamsHKD Identity era), Eq (PParamsHKD StrictMaybe era), Ord (PParamsHKD StrictMaybe era), Show (PParamsHKD StrictMaybe era), NFData (PParamsHKD StrictMaybe era), EncCBOR (PParamsHKD StrictMaybe era), DecCBOR (PParamsHKD StrictMaybe era), ToCBOR (PParamsHKD StrictMaybe era), FromCBOR (PParamsHKD StrictMaybe era), NoThunks (PParamsHKD StrictMaybe era), ToJSON (PParamsHKD StrictMaybe era)) ⇒ EraPParams era where

Minimal complete definition

emptyPParamsIdentity, emptyPParamsStrictMaybe, upgradePParamsHKD, downgradePParamsHKD, hkdMinFeeAL, hkdMinFeeBL, hkdMaxBBSizeL, hkdMaxTxSizeL, hkdMaxBHSizeL, hkdKeyDepositL, hkdPoolDepositL, hkdEMaxL, hkdNOptL, hkdA0L, hkdRhoL, hkdTauL, hkdDL, hkdExtraEntropyL, hkdProtocolVersionL, hkdMinUTxOValueL, hkdMinPoolCostL

Methods

ppProtocolVersionL ∷ Lens' (PParams era) ProtVer

data PParams era

Protocol parameters

emptyPParams ∷ EraPParams era ⇒ PParams era

ppMinFeeAL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The linear factor for the minimum fee calculation

ppMinFeeBL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The constant factor for the minimum fee calculation

ppMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal block body size

ppMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word32

Maximal transaction size

ppMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParams era) Word16

Maximal block header size

ppKeyDepositL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a key registration deposit

ppPoolDepositL ∷ EraPParams era ⇒ Lens' (PParams era) Coin

The amount of a pool registration deposit

ppEMaxL ∷ EraPParams era ⇒ Lens' (PParams era) EpochInterval

epoch bound on pool retirement

ppNOptL ∷ EraPParams era ⇒ Lens' (PParams era) Word16

Desired number of pools

ppA0L ∷ EraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

Pool influence

ppRhoL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Monetary expansion

ppTauL ∷ EraPParams era ⇒ Lens' (PParams era) UnitInterval

Treasury expansion

ppDL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParams era) UnitInterval

Decentralization parameter

data PParamsUpdate era

The type of updates to Protocol parameters

emptyPParamsUpdate ∷ EraPParams era ⇒ PParamsUpdate era

ppuMinFeeAL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The linear factor for the minimum fee calculation

ppuMinFeeBL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The constant factor for the minimum fee calculation

ppuMaxBBSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal block body size

ppuMaxTxSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word32)

Maximal transaction size

ppuMaxBHSizeL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Maximal block header size

ppuKeyDepositL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a key registration deposit

ppuPoolDepositL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

The amount of a pool registration deposit

ppuEMaxL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

epoch bound on pool retirement

ppuNOptL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Word16)

Desired number of pools

ppuA0L ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

Pool influence

ppuRhoL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Monetary expansion

ppuTauL ∷ EraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Treasury expansion

ppuDL ∷ (EraPParams era, ProtVerAtMost era 6) ⇒ Lens' (PParamsUpdate era) (StrictMaybe UnitInterval)

Decentralization parameter

class EraPParams era ⇒ AlonzoEraPParams era

Minimal complete definition

hkdCoinsPerUTxOWordL, hkdCostModelsL, hkdPricesL, hkdMaxTxExUnitsL, hkdMaxBlockExUnitsL, hkdMaxValSizeL, hkdCollateralPercentageL, hkdMaxCollateralInputsL

ppCoinsPerUTxOWordL ∷ (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParams era) CoinPerWord

ppCostModelsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) CostModels

ppPricesL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Prices

ppMaxTxExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxBlockExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) ExUnits

ppMaxValSizeL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppCollateralPercentageL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppMaxCollateralInputsL ∷ AlonzoEraPParams era ⇒ Lens' (PParams era) Natural

ppuCoinsPerUTxOWordL ∷ (AlonzoEraPParams era, ExactEra AlonzoEra era) ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerWord)

ppuCostModelsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CostModels)

ppuPricesL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Prices)

ppuMaxTxExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxBlockExUnitsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe ExUnits)

ppuMaxValSizeL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCollateralPercentageL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuMaxCollateralInputsL ∷ AlonzoEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

class AlonzoEraPParams era ⇒ BabbageEraPParams era where

Methods

hkdCoinsPerUTxOByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f CoinPerByte)

ppCoinsPerUTxOByteL ∷ BabbageEraPParams era ⇒ Lens' (PParams era) CoinPerByte

ppuCoinsPerUTxOByteL ∷ BabbageEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe CoinPerByte)

class BabbageEraPParams era ⇒ ConwayEraPParams era

Minimal complete definition

modifiedPPGroups, ppuWellFormed, hkdPoolVotingThresholdsL, hkdDRepVotingThresholdsL, hkdCommitteeMinSizeL, hkdCommitteeMaxTermLengthL, hkdGovActionLifetimeL, hkdGovActionDepositL, hkdDRepDepositL, hkdDRepActivityL, hkdMinFeeRefScriptCostPerByteL

ppPoolVotingThresholdsL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) PoolVotingThresholds

ppDRepVotingThresholdsL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) DRepVotingThresholds

ppCommitteeMinSizeL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) Natural

ppCommitteeMaxTermLengthL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionLifetimeL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionDepositL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepDepositL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepActivityL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppMinFeeRefScriptCostPerByteL ∷ ConwayEraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

ppuPoolVotingThresholdsL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe PoolVotingThresholds)

ppuDRepVotingThresholdsL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe DRepVotingThresholds)

ppuCommitteeMinSizeL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCommitteeMaxTermLengthL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionLifetimeL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionDepositL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepDepositL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepActivityL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuMinFeeRefScriptCostPerByteL ∷ ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

class (Era era, Show (Script era), Eq (Script era), EqRaw (Script era), ToCBOR (Script era), EncCBOR (Script era), DecCBOR (Script era), NoThunks (Script era), SafeToHash (Script era), Eq (NativeScript era), Show (NativeScript era), NFData (NativeScript era), NoThunks (NativeScript era), EncCBOR (NativeScript era), DecCBOR (NativeScript era)) ⇒ EraScript era where

Typeclass for script data types. Allows for script validation and hashing. You must understand the role of SafeToHash and scriptPrefixTag to make new instances. scriptPrefixTag is a magic number representing the tag of the script language. For each new script language defined, a new tag is chosen and the tag is included in the script hash for a script. The safeToHash constraint ensures that Scripts are never reserialised.

Minimal complete definition

upgradeScript, scriptPrefixTag, getNativeScript, fromNativeScript

Associated Types

type Script era = (r ∷ Type) | r → era

Scripts which may lock transaction outputs in this era

type NativeScript era = (r ∷ Type) | r → era

Methods

getNativeScript ∷ Script era → Maybe (NativeScript era)

fromNativeScript ∷ NativeScript era → Script era

hashScript ∷ EraScript era ⇒ Script era → ScriptHash

Compute ScriptHash of a Script for a particular era.