Test.Cardano.Ledger.Examples.STSTestUtils

Documentation

initUTxO ∷ ∀ era. (EraTxOut era, PostShelley era) ⇒ Proof era → UTxO era

mkGenesisTxIn ∷ (HashAlgorithm (HASH c), HasCallStack) ⇒ Integer → TxIn c

mkTxDats ∷ Era era ⇒ Data era → TxDats era

someAddr ∷ ∀ era. Era era ⇒ Proof era → Addr (EraCrypto era)

someKeys ∷ ∀ era. Era era ⇒ Proof era → KeyPair 'Payment (EraCrypto era)

someScriptAddr ∷ ∀ era. Scriptic era ⇒ Script era → Addr (EraCrypto era)

testBBODY ∷ (Reflect era, HasCallStack) ⇒ WitRule "BBODY" era → ShelleyBbodyState era → Block (BHeaderView (EraCrypto era)) era → Either (NonEmpty (PredicateFailure (EraRule "BBODY" era))) (ShelleyBbodyState era) → PParams era → Assertion

runLEDGER ∷ ∀ era. (GoodCrypto (EraCrypto era), EraTx era, EraGov era) ⇒ WitRule "LEDGER" era → LedgerState era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era))

testUTXOW ∷ ∀ era. (Reflect era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use an equality test on the expected and computed [PredicateFailure]

testUTXOWsubset ∷ ∀ era. (Reflect era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use a subset test on the expected and computed [PredicateFailure]

testUTXOspecialCase ∷ ∀ era. (Reflect era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either (NonEmpty (PredicateFailure (EraRule "UTXOW" era))) (State (EraRule "UTXOW" era)) → Assertion

Use a test where any two (ValidationTagMismatch x y) failures match regardless of x and y

trustMeP ∷ Proof era → Bool → Tx era → Tx era

alwaysFailsHash ∷ ∀ era. Scriptic era ⇒ Natural → Proof era → ScriptHash (EraCrypto era)

alwaysSucceedsHash ∷ ∀ era. Scriptic era ⇒ Natural → Proof era → ScriptHash (EraCrypto era)

timelockScript ∷ PostShelley era ⇒ Int → Proof era → Script era

timelockHash ∷ ∀ era. PostShelley era ⇒ Int → Proof era → ScriptHash (EraCrypto era)

timelockStakeCred ∷ PostShelley era ⇒ Proof era → StakeCredential (EraCrypto era)