Test.Cardano.Ledger.Examples.STSTestUtils

Documentation

class AlonzoBased era failure where

We use this to write set of tests that raise AlonzoBased PredicateFailures. The type we use is for '(fail ~ PredicateFailure (EraRule UTXOW era))' . There are 4 types of AlonzoBased PredicateFailures: UtxowPredFail, UtxosPredicateFailure, UtxoPredicateFailure, and UtxowPredicateFailure . The idea is to make tests that only raise these failures, in Alonzo and future Eras.

Methods

fromUtxos ∷ AlonzoUtxosPredFailure era → failure

fromUtxo ∷ AlonzoUtxoPredFailure era → failure

fromUtxow ∷ ShelleyUtxowPredFailure era → failure

fromPredFail ∷ AlonzoUtxowPredFailure era → failure

freeCostModelV1 ∷ CostModel

A cost model that sets everything as being free

freeCostModelV2 ∷ CostModel

A cost model that sets everything as being free

initUTxO ∷ ∀ era. (EraTxOut era, PostShelley era) ⇒ Proof era → UTxO era

mkGenesisTxIn ∷ (HashAlgorithm (HASH c), HasCallStack) ⇒ Integer → TxIn c

mkTxDats ∷ Era era ⇒ Data era → TxDats era

someAddr ∷ ∀ era. Era era ⇒ Proof era → Addr (EraCrypto era)

someKeys ∷ ∀ era. Era era ⇒ Proof era → KeyPair 'Payment (EraCrypto era)

someScriptAddr ∷ ∀ era. Scriptic era ⇒ Script era → Addr (EraCrypto era)

testBBODY ∷ (GoodCrypto (EraCrypto era), HasCallStack) ⇒ WitRule "BBODY" era → ShelleyBbodyState era → Block (BHeaderView (EraCrypto era)) era → Either [PredicateFailure (AlonzoBBODY era)] (ShelleyBbodyState era) → PParams era → Assertion

testUTXOW ∷ ∀ era. (GoodCrypto (EraCrypto era), PostShelley era, EraTx era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either [PredicateFailure (EraRule "UTXOW" era)] (State (EraRule "UTXOW" era)) → Assertion

Use an equality test on the expected and computed [PredicateFailure]

testUTXOWsubset ∷ ∀ era. (GoodCrypto (EraCrypto era), PostShelley era, EraTx era, EraGovernance era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either [PredicateFailure (EraRule "UTXOW" era)] (State (EraRule "UTXOW" era)) → Assertion

Use a subset test on the expected and computed [PredicateFailure]

testUTXOspecialCase ∷ ∀ era. (GoodCrypto (EraCrypto era), PostShelley era, EraTx era, EraGovernance era, HasCallStack) ⇒ WitRule "UTXOW" era → UTxO era → PParams era → Tx era → Either [PredicateFailure (EraRule "UTXOW" era)] (State (EraRule "UTXOW" era)) → Assertion

Use a test where any two (ValidationTagMismatch x y) failures match regardless of x and y

trustMeP ∷ Proof era → Bool → Tx era → Tx era

keyBy ∷ Ord k ⇒ (a → k) → [a] → Map k a

alwaysFailsHash ∷ ∀ era. Scriptic era ⇒ Natural → Proof era → ScriptHash (EraCrypto era)

alwaysSucceedsHash ∷ ∀ era. Scriptic era ⇒ Natural → Proof era → ScriptHash (EraCrypto era)

timelockScript ∷ PostShelley era ⇒ Int → Proof era → Script era

timelockHash ∷ ∀ era. PostShelley era ⇒ Int → Proof era → ScriptHash (EraCrypto era)

timelockStakeCred ∷ PostShelley era ⇒ Proof era → StakeCredential (EraCrypto era)