Test.Cardano.Ledger.Generic.Fields

Documentation

data TxField era

Constructors

Body (TxBody era)

BodyI [TxBodyField era]

TxWits (TxWits era)

WitnessesI [WitnessesField era]

AuxData (StrictMaybe (TxAuxData era))

Valid IsValid

Bundled Patterns

pattern AuxData' ∷ [TxAuxData era] → TxField era

pattern Valid' ∷ Bool → TxField era

data TxBodyField era

Constructors

Inputs (Set (TxIn (EraCrypto era)))

Collateral (Set (TxIn (EraCrypto era)))

RefInputs (Set (TxIn (EraCrypto era)))

Outputs (StrictSeq (TxOut era))

CollateralReturn (StrictMaybe (TxOut era))

TotalCol (StrictMaybe Coin)

Certs (StrictSeq (TxCert era))

Withdrawals' (Withdrawals (EraCrypto era))

Txfee Coin

Vldt ValidityInterval

TTL SlotNo

Update (StrictMaybe (Update era))

ReqSignerHashes (Set (KeyHash 'Witness (EraCrypto era)))

Mint (MultiAsset (EraCrypto era))

WppHash (StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

AdHash (StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

Txnetworkid (StrictMaybe Network)

ProposalProc (OSet (ProposalProcedure era))

VotingProc (VotingProcedures era)

CurrentTreasuryValue (StrictMaybe Coin)

TreasuryDonation Coin

Bundled Patterns

pattern Inputs' ∷ [TxIn (EraCrypto era)] → TxBodyField era

pattern Collateral' ∷ [TxIn (EraCrypto era)] → TxBodyField era

pattern RefInputs' ∷ [TxIn (EraCrypto era)] → TxBodyField era

pattern Outputs' ∷ [TxOut era] → TxBodyField era

pattern Certs' ∷ [TxCert era] → TxBodyField era

pattern CollateralReturn' ∷ [TxOut era] → TxBodyField era

pattern Update' ∷ [Update era] → TxBodyField era

pattern ReqSignerHashes' ∷ [KeyHash 'Witness (EraCrypto era)] → TxBodyField era

pattern WppHash' ∷ [ScriptIntegrityHash (EraCrypto era)] → TxBodyField era

pattern AdHash' ∷ [AuxiliaryDataHash (EraCrypto era)] → TxBodyField era

pattern Txnetworkid' ∷ [Network] → TxBodyField era

data WitnessesField era

Constructors

AddrWits (Set (WitVKey 'Witness (EraCrypto era)))

BootWits (Set (BootstrapWitness (EraCrypto era)))

ScriptWits (Map (ScriptHash (EraCrypto era)) (Script era))

DataWits (TxDats era)

RdmrWits (Redeemers era)

Bundled Patterns

pattern AddrWits' ∷ Era era ⇒ [WitVKey 'Witness (EraCrypto era)] → WitnessesField era

pattern BootWits' ∷ Era era ⇒ [BootstrapWitness (EraCrypto era)] → WitnessesField era

pattern ScriptWits' ∷ ∀ era. EraScript era ⇒ [Script era] → WitnessesField era

pattern DataWits' ∷ Era era ⇒ [Data era] → WitnessesField era

data PParamsField era

Constructors

MinfeeA Coin

MinfeeB Coin
The constant factor for the minimum fee calculation

MaxBBSize Word32
Maximal block body size

MaxTxSize Word32
Maximal transaction size

MaxBHSize Word16
Maximal block header size

KeyDeposit Coin
The amount of a key registration deposit

PoolDeposit Coin
The amount of a pool registration deposit

EMax EpochInterval
epoch bound on pool retirement

NOpt Natural
Desired number of pools

A0 NonNegativeInterval
Pool influence

Rho UnitInterval
Monetary expansion

Tau UnitInterval
Treasury expansion

D UnitInterval
Decentralization parameter

ExtraEntropy Nonce
Extra entropy

ProtocolVersion ProtVer
Protocol version

MinPoolCost Coin
Minimum Stake Pool Cost

MinUTxOValue Coin
Minimum Lovelace in a UTxO deprecated by AdaPerUTxOWord

CoinPerUTxOWord CoinPerWord
Cost in ada per 8 bytes of UTxO storage instead of _minUTxOValue

CoinPerUTxOByte CoinPerByte
Cost in ada per 1 byte of UTxO storage instead of _coinsPerUTxOWord

Costmdls CostModels
Cost models for non-native script languages

Prices Prices
Prices of execution units for non-native script languages

MaxTxExUnits ExUnits
Max total script execution resources units allowed per tx

MaxBlockExUnits ExUnits
Max total script execution resources units allowed per block

MaxValSize Natural
Max size of a Value in an output

CollateralPercentage Natural
The scaling percentage of the collateral relative to the fee

MaxCollateralInputs Natural
Maximum number of collateral inputs allowed in a transaction

PoolVotingThreshold PoolVotingThresholds
These are new to Conway

DRepVotingThreshold DRepVotingThresholds

MinCommitteeSize Natural

CommitteeTermLimit EpochInterval

GovActionExpiration EpochInterval

GovActionDeposit Coin

DRepDeposit Coin

DRepActivity EpochInterval

data TxOutField era

Constructors

Address (Addr (EraCrypto era))

Amount (Value era)

DHash (StrictMaybe (DataHash (EraCrypto era)))

FDatum (Datum era)

RefScript (StrictMaybe (Script era))

Bundled Patterns

pattern DHash' ∷ [DataHash (EraCrypto era)] → TxOutField era

pattern RefScript' ∷ [Script era] → TxOutField era

initVI ∷ ValidityInterval

initWithdrawals ∷ Withdrawals c

initialTxBody ∷ Era era ⇒ Proof era → TxBody era

initialWitnesses ∷ Era era ⇒ Proof era → TxWits era

initialTx ∷ ∀ era. Proof era → Tx era

initialTxOut ∷ Era era ⇒ Proof era → TxOut era

valid ∷ IsValid → Bool

abstractTx ∷ Proof era → Tx era → [TxField era]

abstractTxBody ∷ Proof era → TxBody era → [TxBodyField era]

abstractTxOut ∷ Era era ⇒ Proof era → TxOut era → [TxOutField era]

abstractWitnesses ∷ Proof era → TxWits era → [WitnessesField era]

abstractPParams ∷ Proof era → PParams era → [PParamsField era]