Test.Cardano.Ledger.Generic.Functions

Description

Functions in this module take a (Proof era) as their first parameter and do something potentially different in each Era.

Documentation

maxCollateralInputs' ∷ Proof era → PParams era → Natural

maxTxExUnits' ∷ Proof era → PParams era → ExUnits

collateralPercentage' ∷ Proof era → PParams era → Natural

protocolVersion ∷ Proof era → ProtVer

depositsAndRefunds ∷ (EraPParams era, ShelleyEraTxCert era) ⇒ PParams era → [TxCert era] → Map (Credential 'Staking (EraCrypto era)) Coin → Coin

Positive numbers are "deposits owed", negative amounts are "refunds gained"

scriptWitsNeeded' ∷ Proof era → MUtxo era → TxBody era → Set (ScriptHash (EraCrypto era))

Compute the set of ScriptHashes for which there should be ScriptWitnesses. In Babbage Era and later, where inline Scripts are allowed, they should not appear in this set.

scriptsNeeded' ∷ Proof era → MUtxo era → TxBody era → Set (ScriptHash (EraCrypto era))

txInBalance ∷ ∀ era. EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → MUtxo era → Coin

txoutFields ∷ Proof era → TxOut era → (Addr (EraCrypto era), Value era, [TxOutField era])

Break a TxOut into its mandatory and optional parts

injectFee ∷ EraTxOut era ⇒ Proof era → Coin → TxOut era → TxOut era

getTxOutRefScript ∷ Proof era → TxOut era → StrictMaybe (Script era)

emptyPPUPstate ∷ ∀ era. Proof era → ShelleyGovState era

maxRefInputs ∷ Proof era → Int

isValid' ∷ Proof era → Tx era → IsValid

txoutEvidence ∷ ∀ era. Proof era → TxOut era → ([Credential 'Payment (EraCrypto era)], Maybe (DataHash (EraCrypto era)))

Does the TxOut have evidence of credentials and data. Evidence of data is either ScriptHash or (in Babbage) an inline Datum Evidence of credentials can come from the Addr

addrCredentials ∷ Addr c → [Credential 'Payment c]

paymentCredAddr ∷ Addr c → Maybe (Credential 'Payment c)

stakeCredAddr ∷ Addr c → Maybe (Credential 'Staking c)

getBody ∷ EraTx era ⇒ Proof era → Tx era → TxBody era

getCollateralInputs ∷ Proof era → TxBody era → Set (TxIn (EraCrypto era))

getCollateralOutputs ∷ Proof era → TxBody era → [TxOut era]

getInputs ∷ EraTxBody era ⇒ Proof era → TxBody era → Set (TxIn (EraCrypto era))

getOutputs ∷ EraTxBody era ⇒ Proof era → TxBody era → StrictSeq (TxOut era)

getScriptWits ∷ EraTxWits era ⇒ Proof era → TxWits era → Map (ScriptHash (EraCrypto era)) (Script era)

allInputs ∷ EraTxBody era ⇒ Proof era → TxBody era → Set (TxIn (EraCrypto era))

getWitnesses ∷ EraTx era ⇒ Proof era → Tx era → TxWits era

primaryLanguage ∷ Proof era → Maybe Language

alwaysTrue ∷ ∀ era. Proof era → Maybe Language → Natural → Script era

alwaysFalse ∷ ∀ era. Proof era → Maybe Language → Natural → Script era

certs ∷ (ShelleyEraTxBody era, EraTx era) ⇒ Proof era → Tx era → [TxCert era]

createRUpdNonPulsing' ∷ ∀ era. Proof era → Model era → RewardUpdateOld (EraCrypto era)

Create an old style RewardUpdate to be used in tests, in any Era.

languagesUsed ∷ ∀ era. Proof era → Tx era → UTxO era → Set (ScriptHash (EraCrypto era)) → Set Language

languages ∷ ∀ era. (EraUTxO era, AlonzoEraScript era) ⇒ Tx era → UTxO era → Set (ScriptHash (EraCrypto era)) → Set Language

Compute the Set of Languages in an era, where AlonzoScripts are used

class TotalAda t where

Compute the total Ada from Ada pots within t

Methods

totalAda ∷ t → Coin

govStateTotalAda ∷ ∀ era. Reflect era ⇒ GovState era → Coin

adaPots ∷ Proof era → EpochState era → AdaPots