Test.Cardano.Ledger.Generic.Proof

Documentation

data Proof era where

Proof of a valid (predefined) era

Constructors

Shelley ∷ Proof ShelleyEra

Mary ∷ Proof MaryEra

Allegra ∷ Proof AllegraEra

Alonzo ∷ Proof AlonzoEra

Babbage ∷ Proof BabbageEra

Conway ∷ Proof ConwayEra

class (EraGov era, EraTx era, EraUTxO era, EraTxAuxData era, EraScript era, EraStake era, ShelleyEraTxCert era, EraCertState era, EraTest era) ⇒ Reflect era where

Minimal complete definition

reify

Methods

reify ∷ Proof era

lift ∷ ∀ a. (Proof era → a) → a

data Some (t ∷ k → Type) where

Hide the index for a singleton type t

Constructors

Some ∷ ∀ {k} (t ∷ k → Type) (i ∷ k). Singleton t ⇒ t i → Some t

runSTS ∷ ∀ s e ans. (BaseM (EraRule s e) ~ ShelleyBase, STS (EraRule s e)) ⇒ TRC (EraRule s e) → (Either (NonEmpty (PredicateFailure (EraRule s e))) (State (EraRule s e)) → ans) → ans

goSTS ∷ ∀ s e ans env state sig. (BaseM (EraRule s e) ~ ShelleyBase, STS (EraRule s e), env ~ Environment (EraRule s e), state ~ State (EraRule s e), sig ~ Signal (EraRule s e)) ⇒ env → state → sig → (Either (NonEmpty (PredicateFailure (EraRule s e))) (State (EraRule s e)) → ans) → ans

Like runSTS, but makes the components of the TRC triple explicit. in case you can't remember, in ghci type @@ :t goSTS (UTXOW Babbage) goSTS (LEDGER Babbage) :: LedgerEnv (BabbageEra ) -> (UTxOState (BabbageEra ), CertState ) -> Cardano.Ledger.Alonzo.Tx.AlonzoTx (BabbageEra ) -> (Either [LedgerPredicateFailure (BabbageEra )] (UTxOState (BabbageEra ), CertState ) -> ans) -> ans @@ it will tell you what type env state and sig are for Babbage

preShelley ∷ [Some Proof]

preAllegra ∷ [Some Proof]

preMary ∷ [Some Proof]

preAlonzo ∷ [Some Proof]

preBabbage ∷ [Some Proof]

preConway ∷ [Some Proof]

postShelley ∷ [Some Proof]

postAllegra ∷ [Some Proof]

postMary ∷ [Some Proof]

postAlonzo ∷ [Some Proof]

postBabbage ∷ [Some Proof]

postConway ∷ [Some Proof]

data ShelleyEra

data AllegraEra

data MaryEra

data AlonzoEra

data BabbageEra

data ConwayEra

specialize ∷ ∀ constraint era t. (constraint ShelleyEra, constraint AllegraEra, constraint MaryEra, constraint AlonzoEra, constraint BabbageEra, constraint ConwayEra) ⇒ Proof era → (constraint era ⇒ t) → t

Specialize (action :: (constraint era => t)) to all known era, because we know (constraint era) holds for all known era. In order for this to work it is best to type apply specialize to a concrete constraint. So a call site looks like '(specialize @EraBlockBody proof action). This way the constraint does not percolate upwards, past the call site of action

unReflect ∷ (Reflect era ⇒ Proof era → a) → Proof era → a

lift a function (Proof era -> a) that has a (Reflect era) constraint to one that does not. This is possible because every inhabited term of type (Proof era) packs a (Reflect era) instance. so instead of writing: f proof arg1 .. argn one writes: unReflect f proof arg1 .. argn which will not require a (Reflect era) instance

runSTS' ∷ ∀ s e. (BaseM (EraRule s e) ~ ShelleyBase, STS (EraRule s e)) ⇒ TRC (EraRule s e) → Either (NonEmpty (PredicateFailure (EraRule s e))) (State (EraRule s e))

data ValueWit era where

Constructors

ValueShelleyToAllegra ∷ Value era ~ Coin ⇒ ValueWit era

ValueMaryToConway ∷ Value era ~ MaryValue ⇒ ValueWit era

data TxOutWit era where

Constructors

TxOutShelleyToMary ∷ (TxOut era ~ ShelleyTxOut era, EraTxOut era, AtMostEra "Babbage" era) ⇒ TxOutWit era

TxOutAlonzoToAlonzo ∷ (TxOut era ~ AlonzoTxOut era, AlonzoEraTxOut era, AtMostEra "Babbage" era) ⇒ TxOutWit era

TxOutBabbageToConway ∷ (TxOut era ~ BabbageTxOut era, BabbageEraTxOut era) ⇒ TxOutWit era

data TxCertWit era where

Constructors

TxCertShelleyToBabbage ∷ (TxCert era ~ ShelleyTxCert era, ShelleyEraTxCert era, AtMostEra "Babbage" era) ⇒ TxCertWit era

TxCertConwayToConway ∷ (TxCert era ~ ConwayTxCert era, ConwayEraTxCert era, ConwayEraPParams era) ⇒ TxCertWit era

data PParamsWit era where

Constructors

PParamsShelleyToMary ∷ (PParamsHKD Identity era ~ ShelleyPParams Identity era, EraPParams era) ⇒ PParamsWit era

PParamsAlonzoToAlonzo ∷ (PParamsHKD Identity era ~ AlonzoPParams Identity era, AlonzoEraPParams era) ⇒ PParamsWit era

PParamsBabbageToBabbage ∷ (PParamsHKD Identity era ~ BabbagePParams Identity era, BabbageEraPParams era) ⇒ PParamsWit era

PParamsConwayToConway ∷ (PParamsHKD Identity era ~ ConwayPParams Identity era, ConwayEraPParams era) ⇒ PParamsWit era

data UTxOWit era where

Constructors

UTxOShelleyToMary ∷ (EraUTxO era, ScriptsNeeded era ~ ShelleyScriptsNeeded era, TxWits era ~ ShelleyTxWits era) ⇒ UTxOWit era

UTxOAlonzoToConway ∷ (EraUTxO era, AlonzoEraScript era, AlonzoEraTxBody era, AlonzoEraPParams era, AlonzoEraTxOut era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Script era ~ AlonzoScript era, TxWits era ~ AlonzoTxWits era) ⇒ UTxOWit era

data ScriptWit era where

Constructors

ScriptShelleyToShelley ∷ (Script era ~ MultiSig era, EraScript era) ⇒ ScriptWit era

ScriptAllegraToMary ∷ (Script era ~ Timelock era, EraScript era) ⇒ ScriptWit era

ScriptAlonzoToConway ∷ (Script era ~ AlonzoScript era, EraScript era) ⇒ ScriptWit era

data GovStateWit era where

Constructors

GovStateShelleyToBabbage ∷ (EraGov era, GovState era ~ ShelleyGovState era) ⇒ GovStateWit era

GovStateConwayToConway ∷ (ConwayEraPParams era, RunConwayRatify era, EraGov era, GovState era ~ ConwayGovState era) ⇒ GovStateWit era

data CertStateWit era where

Constructors

CertStateShelleyToBabbage ∷ (EraCertState era, CertState era ~ ShelleyCertState era) ⇒ CertStateWit era

CertStateConwayToConway ∷ (ConwayEraCertState era, CertState era ~ ConwayCertState era) ⇒ CertStateWit era

whichValue ∷ Proof era → ValueWit era

whichTxOut ∷ Proof era → TxOutWit era

whichTxCert ∷ Proof era → TxCertWit era

whichPParams ∷ Proof era → PParamsWit era

whichUTxO ∷ Proof era → UTxOWit era

whichScript ∷ Proof era → ScriptWit era

whichGovState ∷ Proof era → GovStateWit era

whichCertState ∷ Proof era → CertStateWit era