Test.Cardano.Ledger.Generic.Proof

Documentation

type Mock = C_Crypto

type Standard = StandardCrypto

data Evidence c where

Constructors

Standard ∷ Evidence Standard

Mock ∷ Evidence Mock

data Proof era where

Proof of a valid (predefined) era

Constructors

Shelley ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (ShelleyEra c)

Mary ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (MaryEra c)

Allegra ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (AllegraEra c)

Alonzo ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (AlonzoEra c)

Babbage ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (BabbageEra c)

Conway ∷ ∀ c. Crypto c ⇒ Evidence c → Proof (ConwayEra c)

getCrypto ∷ Proof era → Evidence (EraCrypto era)

type GoodCrypto c = (Crypto c, DSignable c (Hash (HASH c) EraIndependentTxBody), DSIGNAlgorithm (DSIGN c), Signable (DSIGN c) (OCertSignable c), Signable (VRF c) Seed, Signable (KES c) (BHBody c), ContextKES (KES c) ~ (), ContextVRF (VRF c) ~ (), HashAlgorithm (HASH c), PraosCrypto c)

class GoodCrypto c ⇒ ReflectC c where

Minimal complete definition

evidence

Methods

evidence ∷ Evidence c

liftC ∷ ∀ a. (Evidence c → a) → a

class (EraGovernance era, EraTx era, ReflectC (EraCrypto era)) ⇒ Reflect era where

Minimal complete definition

reify

Methods

reify ∷ Proof era

lift ∷ ∀ a. (Proof era → a) → a

data Some (t ∷ k → Type) where

Hide the index for a singleton type t

Constructors

Some ∷ ∀ {k} (t ∷ k → Type) (i ∷ k). Singleton t ⇒ t i → Some t

data WitRule (s ∷ Symbol) (e ∷ Type) where

Constructors

UTXOW ∷ Proof era → WitRule "UTXOW" era

LEDGER ∷ Proof era → WitRule "LEDGER" era

BBODY ∷ Proof era → WitRule "BBODY" era

LEDGERS ∷ Proof era → WitRule "LEDGERS" era

MOCKCHAIN ∷ Proof era → WitRule "MOCKCHAIN" era

ruleProof ∷ WitRule s e → Proof e

runSTS ∷ ∀ s e ans. (BaseM (EraRule s e) ~ ShelleyBase, STS (EraRule s e)) ⇒ WitRule s e → TRC (EraRule s e) → (Either [PredicateFailure (EraRule s e)] (State (EraRule s e)) → ans) → ans

goSTS ∷ ∀ s e ans env state sig. (BaseM (EraRule s e) ~ ShelleyBase, STS (EraRule s e), env ~ Environment (EraRule s e), state ~ State (EraRule s e), sig ~ Signal (EraRule s e)) ⇒ WitRule s e → env → state → sig → (Either [PredicateFailure (EraRule s e)] (State (EraRule s e)) → ans) → ans

Like runSTS, but makes the components of the TRC triple explicit. in case you can't remember, in ghci type @@ :t goSTS (UTXOW (Babbage Mock)) goSTS (LEDGER (Babbage Mock)) :: LedgerEnv (BabbageEra C_Crypto) -> (UTxOState (BabbageEra C_Crypto), CertState C_Crypto) -> Cardano.Ledger.Alonzo.Tx.AlonzoTx (BabbageEra C_Crypto) -> (Either [LedgerPredicateFailure (BabbageEra C_Crypto)] (UTxOState (BabbageEra C_Crypto), CertState C_Crypto) -> ans) -> ans @@ it will tell you what type env state and sig are for Babbage

preShelley ∷ [Some Proof]

preAllegra ∷ [Some Proof]

preMary ∷ [Some Proof]

preAlonzo ∷ [Some Proof]

preBabbage ∷ [Some Proof]

preConway ∷ [Some Proof]

postShelley ∷ [Some Proof]

postAllegra ∷ [Some Proof]

postMary ∷ [Some Proof]

postAlonzo ∷ [Some Proof]

postBabbage ∷ [Some Proof]

postConway ∷ [Some Proof]

data ShelleyEra c

data AllegraEra c

The Allegra era

data MaryEra era

data AlonzoEra c

The Alonzo era

data BabbageEra c

The Babbage era

data ConwayEra c

The Conway era

data C_Crypto

specialize ∷ ∀ constraint era t. (constraint (ShelleyEra (EraCrypto era)), constraint (AllegraEra (EraCrypto era)), constraint (MaryEra (EraCrypto era)), constraint (AlonzoEra (EraCrypto era)), constraint (BabbageEra (EraCrypto era)), constraint (ConwayEra (EraCrypto era))) ⇒ Proof era → (constraint era ⇒ t) → t

Specialize (action :: (constraint era => t)) to all known era, because we know (constraint era) holds for all known era. In order for this to work it is best to type apply specialize to a concrete constraint. So a call site looks like '(specialize @EraSegWits proof action). This way the constraint does not percolate upwards, past the call site of action

unReflect ∷ (Reflect era ⇒ Proof era → a) → Proof era → a

lift a function (Proof era -> a) that has a (Reflect era) constraint to one that does not. This is possible because every inhabited term of type (Proof era) packs a (Reflect era) instance. so instead of writing: f proof arg1 .. argn one writes: unReflect f proof arg1 .. argn which will not require a (Reflect era) instance