Test.Cardano.Ledger.Generic.Properties

Documentation

genTxAndUTXOState ∷ Reflect era ⇒ Proof era → GenSize → Gen (TRC (EraRule "UTXOW" era), GenState era)

genTxAndLEDGERStateShelley ∷ GenSize → Gen (TRC (EraRule "LEDGER" Shelley), GenState Shelley)

testTxValidForLEDGERShelley ∷ (TRC (EraRule "LEDGER" Shelley), GenState Shelley) → Property

genTxAndLEDGERState ∷ ∀ era. (Reflect era, Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, Environment (EraRule "LEDGER" era) ~ LedgerEnv era) ⇒ Proof era → GenSize → Gen (Box era)

testTxValidForLEDGER ∷ (Reflect era, Signal (EraRule "LEDGER" era) ~ Tx era, State (EraRule "LEDGER" era) ~ LedgerState era, PrettyA (PredicateFailure (EraRule "LEDGER" era))) ⇒ Proof era → Box era → Property

txOutRoundTrip ∷ EraTxOut era ⇒ Proof era → TxOut era → Property

txRoundTrip ∷ EraTx era ⇒ Proof era → Tx era → Property

txBodyRoundTrip ∷ EraTxBody era ⇒ Proof era → TxBody era → Property

txWitRoundTrip ∷ EraTxWits era ⇒ Proof era → TxWits era → Property

coreTypesRoundTrip ∷ TestTree

txPreserveAda ∷ GenSize → TestTree

A single Tx preserves Ada

adaIsPreserved ∷ (Reflect era, HasTrace (MOCKCHAIN era) (Gen1 era)) ⇒ Proof era → Int → GenSize → TestTree

Ada is preserved over a trace of length 100

tracePreserveAda ∷ Int → GenSize → TestTree

adaIsPreservedBabbage ∷ Int → GenSize → TestTree

stakeInvariant ∷ EraTxOut era ⇒ MockChainState era → MockChainState era → Property

The incremental Stake invaraint is preserved over a trace of length 100=

incrementStakeInvariant ∷ (Reflect era, HasTrace (MOCKCHAIN era) (Gen1 era)) ⇒ Proof era → GenSize → TestTree

incrementalStake ∷ GenSize → TestTree

genericProperties ∷ GenSize → TestTree

epochPreserveAda ∷ GenSize → TestTree

adaIsPreservedInEachEpoch ∷ ∀ era. Reflect era ⇒ Proof era → GenSize → TestTree

twiddleInvariantHolds ∷ ∀ a. (Arbitrary a, Show a, Twiddle a) ⇒ String → TestTree

twiddleInvariantHoldsEras ∷ TestTree

main ∷ IO ()

main8 ∷ IO ()

test ∷ Int → Proof era → IO ()

makeGen ∷ Reflect era ⇒ Proof era → (Proof era → GenRS era b) → Gen b

Construct a random (Gen b)

runTest ∷ (Reflect era, PrettyA a) ⇒ (Proof era → GenRS era a) → (a → IO ()) → Proof era → IO ()

main2 ∷ IO ()

main3 ∷ IO ()