Test.Cardano.Ledger.Generic.Trace

Documentation

genRsTxAndModel ∷ ∀ era. EraGenericGen era ⇒ Int → SlotNo → GenRS era (Tx era)

Generate a Tx and an internal Model of the state after the tx has been applied. That model can be used to generate the next Tx

genRsTxSeq ∷ ∀ era. EraGenericGen era ⇒ Int → Int → [(StrictSeq (Tx era), SlotNo)] → SlotNo → GenRS era (Vector (StrictSeq (Tx era), SlotNo))

Generate a Vector of (StrictSeq (Tx era)) representing a (Vector Block)

genTxSeq ∷ ∀ era. EraGenericGen era ⇒ GenSize → Int → GenRS era () → Gen (Vector (StrictSeq (Tx era), SlotNo), GenState era)

Generate a Vector of Blocks, and an initial LedgerState

runTest ∷ IO ()

initialMockChainState ∷ ∀ era. (Reflect era, ShelleyEraAccounts era) ⇒ GenState era → MockChainState era

makeEpochState ∷ (Reflect era, ShelleyEraAccounts era) ⇒ GenState era → LedgerState era → EpochState era

snaps ∷ (EraTxOut era, EraCertState era, ShelleyEraAccounts era) ⇒ LedgerState era → SnapShots

raiseMockError ∷ ∀ era. Reflect era ⇒ Word64 → SlotNo → EpochState era → NonEmpty (MockChainFailure era) → [Tx era] → GenState era → String

badScripts ∷ Proof era → [MockChainFailure era] → Set ScriptHash

shortTxOut ∷ EraTxOut era ⇒ TxOut era → Expr

smartTxBody ∷ EraTest era ⇒ MUtxo era → TxBody era → Expr

data Gen1 era

Constructors

Gen1 (Vector (StrictSeq (Tx era), SlotNo)) (GenState era)

mapProportion ∷ EpochNo → Word64 → Int → Map k Int → Gen k

chooseIssuer ∷ EpochNo → Word64 → Int → PoolDistr → Gen (KeyHash 'StakePool)

genTrace ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → GenRS era () → Gen (Trace (MOCKCHAIN era))

traceProp ∷ ∀ era prop. (HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → (MockChainState era → MockChainState era → prop) → Gen prop

forEachEpochTrace ∷ ∀ era prop. (Testable prop, HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → (Trace (MOCKCHAIN era) → prop) → Gen Property

forAllTraceFromInitState

Arguments

∷ ∀ sts traceGenEnv prop. (HasTrace sts traceGenEnv, Testable prop, Show (Environment sts))

⇒ BaseEnv sts

→ Word64
Maximum trace length.

→ traceGenEnv

→ Maybe (IRC sts → Gen (Either (NonEmpty (PredicateFailure sts)) (State sts)))
Optional generator of STS initial state

→ (Trace sts → prop)

→ Property

Check a property on the sts traces.

Takes an optional generator for initial state of the STS.

testPropMax ∷ Testable prop ⇒ Int → String → prop → TestTree

We are making smoke tests, testing for smoke rather than fire. Under the assumption that shorter tests have advantages like not getting turned off because the tests take too long. A glaring failure is likely to be caught in n tests, rather than the standard '100'

chainTest ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), Eq (StashedAVVMAddresses era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → TestTree

testTraces ∷ Int → TestTree

multiEpochTest ∷ ∀ era. (HasTrace (MOCKCHAIN era) (Gen1 era), EraGenericGen era, ShelleyEraAccounts era) ⇒ Int → GenSize → TestTree

Show that Ada is preserved across multiple Epochs