Test.Cardano.Ledger.Generic.TxGen

Documentation

alonzoMkRedeemers ∷ ∀ era. AlonzoEraScript era ⇒ [(PlutusPurpose AsIx era, (Data era, ExUnits))] → Redeemers era

alonzoMkRedeemersFromTags ∷ (AlonzoEraScript era, EraModel era) ⇒ [((PlutusPurposeTag, Word32), (Data era, ExUnits))] → Redeemers era

mkAlonzoPlutusPurposePointer ∷ ∀ era. Era era ⇒ PlutusPurposeTag → Word32 → AlonzoPlutusPurpose AsIx era

mkConwayPlutusPurposePointer ∷ PlutusPurposeTag → Word32 → ConwayPlutusPurpose AsIx era

genAlonzoTx ∷ ∀ era. EraGenericGen era ⇒ SlotNo → GenRS era (UTxO era, Tx era)

runSTSWithContext ∷ ∀ era. (BaseM (EraRule "LEDGER" era) ~ ShelleyBase, STS (EraRule "LEDGER" era)) ⇒ RuleContext 'Transition (EraRule "LEDGER" era) → Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (State (EraRule "LEDGER" era))

genUTxO ∷ EraGenericGen era ⇒ GenRS era (MUtxo era, Maybe (UtxoEntry era))

Generate a somewhat arbitrary MUtxo. Occasionally add a bit of the MUtxo in the Model to the one generated. This way the Tx we generate may spend some of the old UTxo. The result has at most 1 entry from the old MUtxo, and If it has only one entry, that entry is not from the old MUtxo